% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyopenms.R
\name{addXIC}
\alias{addXIC}
\title{Add XIC to pyopenms experiment}
\usage{
addXIC(ropenms, expriment, xic, nativeId)
}
\arguments{
\item{ropenms}{(pyopenms module) get this python module through get_ropenms().}

\item{expriment}{(python object) an MSExperiment() created using ropenms.}

\item{xic}{(data-frame) must have two numeric columns.}

\item{nativeId}{(integer) transition ID of the xic.}
}
\value{
(None)
}
\description{
A chromatogram and its repective native ID (transition ID) is added to MSExperiment object.
}
\examples{
ropenms = get_ropenms(condaEnv = "TricEnvr", useConda=TRUE)
expriment = ropenms$MSExperiment()
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR)
xic <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run0"]][["14299_QFNNTDIVLLEDFQK/3"]][[1]]
\dontrun{
DIAlignR:::addXIC(ropenms, expriment, xic, 34L)
chroms = expriment$getChromatograms()
reticulate::py_to_r(chroms[[0]]$getNativeID())
reticulate::py_to_r(chroms[[0]]$get_peaks())
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-06
}
\keyword{internal}
