% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_aligned_obj.R
\docType{class}
\name{AffineAlignObjMedium-class}
\alias{AffineAlignObjMedium-class}
\alias{AffineAlignObjMedium}
\title{An S4 object for class AffineAlignObj. It only contains similarity matrix and aligned indices.}
\description{
An S4 object for class AffineAlignObj. It only contains similarity matrix and aligned indices.
}
\seealso{
\code{\link{doAffineAlignmentCpp}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-14
}
