% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{alignChromatograms_cpp}
\alias{alignChromatograms_cpp}
\title{Outputs aligned chromatograms.}
\usage{
alignChromatograms_cpp(l1, l2, tA, tB, Normalization, SimType, B1p, B2p,
  noBeef, goFactor = 0.125, geFactor = 40, cosAngleThresh = 0.3,
  OverlapAlignment = TRUE, dotProdThresh = 0.96, gapQuantile = 0.5,
  hardConstrain = FALSE, samples4gradient = 100)
}
\arguments{
\item{ROW_SIZE}{(int) Number of rows of a matrix}

\item{COL_SIZE}{(int) Number of columns of a matrix}
}
\value{
affineAlignObj (S4class) A S4class dummy object from C++ AffineAlignObj struct
}
\description{
Outputs aligned chromatograms.
}
\examples{
simMeasure <- "dotProductMasked"
run_pair <- c("run1", "run2")
peptide <- peptides[1]
r1 <- lapply(StrepChroms[[run_pair[1]]][[peptide]], `[[`, 2)
r2 <- lapply(StrepChroms[[run_pair[2]]][[peptide]], `[[`, 2)
tRunAVec <- StrepChroms[[run_pair[1]]][[peptide]][[1]][["time"]]
tRunBVec <- StrepChroms[[run_pair[2]]][[peptide]][[1]][["time"]]
noBeef <- 6
B1p <- predict(Loess.fit, tRunAVec[1]); B2p <- predict(Loess.fit, tRunAVec[length(tRunAVec)])
Alignobj <- alignChromatograms_cpp(r1, r2, tRunAVec, tRunBVec, "mean", simMeasure, B1p, B2p, noBeef)

}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
