% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dia_runs.R
\name{alignIthAnalyte}
\alias{alignIthAnalyte}
\title{Aligns an analyte across runs}
\usage{
alignIthAnalyte(
  rownum,
  peptideIDs,
  multipeptide,
  refRuns,
  precursors,
  prec2chromIndex,
  fileInfo,
  mzPntrs,
  params,
  globalFits,
  RSE
)
}
\arguments{
\item{rownum}{(integer) represnts the index of the multipepetide to be aligned.}

\item{multipeptide}{(list) contains multiple data-frames that are collection of features
associated with analytes. This is an output of \code{\link{getMultipeptide}}.}

\item{refRuns}{(data-frame) output of \code{\link{getRefRun}}. Must have two columsn : transition_group_id and run.}

\item{precursors}{(data-frame) atleast two columns transition_group_id and transition_ids are required.}

\item{prec2chromIndex}{(list) a list of dataframes having following columns: \cr
transition_group_id: it is PRECURSOR.ID from osw file. \cr
chromatogramIndex: index of chromatogram in mzML file.}

\item{fileInfo}{(data-frame) output of \code{\link{getRunNames}}.}

\item{mzPntrs}{(list) a list of mzRpwiz.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{globalFits}{(list) each element is either of class lm or loess. This is an output of \code{\link{getGlobalFits}}.}

\item{RSE}{(list) Each element represents Residual Standard Error of corresponding fit in globalFits.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\value{
(dataframe) a collection of aligned features related to analyte_chr.
}
\description{
For the ith analyte in multipeptide, this function aligns all runs to the reference run. The result is
a dataframe that contains aligned features corresponding to the analyte across all runs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
}
\seealso{
\code{\link{alignTargetedRuns}, \link{alignToRef}, \link{getAlignedTimes}, \link{getMultipeptide}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-26
}
\keyword{internal}
