% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alignmentStats}
\alias{alignmentStats}
\title{Prints alignment summary}
\usage{
alignmentStats(finalTbl, params)
}
\arguments{
\item{finalTbl}{(dataframe) an output of  \code{\link{writeTables}} function.}

\item{params}{(list) must have following elements: alignedFDR and unalignedFDR.}
}
\value{
Invisible NULL
}
\description{
Prints alignment summary
}
\seealso{
\code{\link{paramsDIAlignR}, \link{writeTables}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-15
}
\keyword{internal}
