% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_metabo_runs.R
\name{alignMetaboruns}
\alias{alignMetaboruns}
\title{Align XICs of precursors across multiple Targeted-MS runs and outputs quantitative data matrix.}
\usage{
alignMetaboruns(dataPath, alignType = "hybrid", oswMerged = TRUE,
  nameCutPattern = "(.*)(/)(.*)", maxFdrQuery = 0.05,
  maxFdrLoess = 0.01, spanvalue = 0.1, runType = "DIA_Metabolomics",
  normalization = "mean", simMeasure = "dotProductMasked",
  SgolayFiltOrd = 4, SgolayFiltLen = 5, goFactor = 0.125,
  geFactor = 40, cosAngleThresh = 0.3, OverlapAlignment = TRUE,
  dotProdThresh = 0.96, gapQuantile = 0.5, hardConstrain = FALSE,
  samples4gradient = 100, samplingTime = NULL, RSEdistFactor = 3.5)
}
\value{
Saves intensity table in the current directory.
}
\description{
Align XICs of precursors across multiple Targeted-MS runs and outputs quantitative data matrix.
}
