% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_mrm_runs.R
\name{alignMRMruns}
\alias{alignMRMruns}
\title{Align XICs of precursors across multiple MRM runs and outputs quantitative data matrix.}
\usage{
alignMRMruns(dataPath, alignType = "local", oswMerged = FALSE,
  nameCutPattern = "(.*)(/)(.*)", maxFdrQuery = 0.05,
  maxFdrLoess = 0.01, spanvalue = 0.1, runType = "MRM_Proteomics",
  normalization = "mean", simMeasure = "dotProductMasked",
  SgolayFiltOrd = 2, SgolayFiltLen = 3, goFactor = 0.125,
  geFactor = 400, cosAngleThresh = 0.3, OverlapAlignment = TRUE,
  dotProdThresh = 0.96, gapQuantile = 0.98, hardConstrain = FALSE,
  samples4gradient = 100, expRSE = 8, samplingTime = 3.4,
  RSEdistFactor = 3.5)
}
\value{
Saves intensity table in the current directory.
}
\description{
In MRM runs, since XICs have less interference, higher quantile value from similarity
matrix is needed to get base-gap-penalty.
}
