% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alignObj_DIAlignR}
\alias{alignObj_DIAlignR}
\title{Alignment object of a peptide.}
\format{
A S4 object of 16 slots:
\describe{
  \item{s}{similarity score matrix.}
  \item{M}{Match or Mismatch matrix, residues of A and B are aligned without a gap. M(i,j) = Best score upto (i,j) given Ai is aligned to Bj.}
  \item{A}{Insert in sequence A, residue in A is aligned to gap in B. A(i,j) is the best score given that Ai is aligned to a gap in B.}
  \item{B}{Insert in sequence B, residue in B is aligned to gap in A. B(i,j) is the best score given that Bj is aligned to a gap in A.}
  \item{Traceback}{Traceback matrices store source matrix name and direction as matrices are filled with dynamic programming.}
  \item{path}{Path matrix would represent alignment path through similarity matrix as binary-hot encoding.}
  \item{signalA_len}{Number of data-points in signal A.}
  \item{signalB_len}{Number of data-points in signal B.}
  \item{GapOpen}{Penalty for Gap opening. For n consecutive gaps: Penalty = GapOpen + (n-1)*GapExten.}
  \item{GapExten}{Penalty for Gap extension. For n consecutive gaps: Penalty = GapOpen + (n-1)*GapExten.}
  \item{FreeEndGaps}{True for Overlap alignment.}
  \item{indexA_aligned}{Aligned signalA indices after affine alignment.}
  \item{indexB_aligned}{Aligned signalB indices after affine alignment.}
  \item{score}{Cumulative score along the aligned path.}
  \item{simScore_forw}{Not needed, will be removed.}
  \item{nGaps}{Total number of gaps in the alignment path.}
}
}
\source{
C++ code is exaplained at \href{https://shubham1637.github.io/DIAlignR/src/doc/html/structDIAlign_1_1AffineAlignObj.html}{DIAlign namespace}.
File test_GenerateData.R has \href{https://github.com/shubham1637/DIAlignR/tree/master/data-raw}{source code}
to generate the example data.
}
\usage{
alignObj_DIAlignR
}
\description{
Aligned XICs of peptide 14299_QFNNTDIVLLEDFQK/3 across two SWATH runs:

run1 : hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML\cr
run2 : hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML
}
\keyword{datasets}
