% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nature_paper.R
\name{alignTargetedRuns2}
\alias{alignTargetedRuns2}
\title{Star alignment for few ids}
\usage{
alignTargetedRuns2(
  dataPath,
  outFile = "DIAlignR.tsv",
  params,
  ids = NULL,
  oswMerged = TRUE,
  runs = NULL,
  applyFun = lapply
)
}
\arguments{
\item{dataPath}{(string) path to mzml and osw directory.}

\item{outFile}{(string) name of the output file.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{runs}{(string) names of mzml file without extension.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\value{
(None)
}
\description{
Star alignment for few ids
}
\examples{
\dontrun{
ids <- as.integer(scan(file = "data/ids.txt"))
params <- paramsDIAlignR()
params[["maxFdrQuery"]] <- 1.0
params[["maxPeptideFdr"]] <- 1.0
params[["kernelLen"]] <- 11L
params$globalAlignmentFdr <- 0.001
params[["batchSize"]] <- 1000L
BiocParallel::register(BiocParallel::MulticoreParam(workers = 6, log = FALSE, threshold = "INFO", stop.on.error = TRUE))
alignTargetedRuns2(dataPath = ".", params = params, outFile = "te1.tsv", ids = ids, applyFun = BiocParallel::bplapply)
}
}
\seealso{
\code{\link{alignTargetedRuns}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-01-20
}
\keyword{internal}
