% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dia_runs.R
\name{alignToRef2}
\alias{alignToRef2}
\title{Aligns an analyte from an experiment to the reference run}
\usage{
alignToRef2(
  eXp,
  ref,
  idx,
  refIdx,
  fileInfo,
  XICs,
  XICs.ref,
  params,
  df,
  globalFits,
  RSE
)
}
\arguments{
\item{eXp}{(string) name of the run to be aligned to reference run. Must be in the rownames of fileInfo.}

\item{ref}{(string) name of the reference run. Must be in the rownames of fileInfo.}

\item{df}{(dataframe) a collection of features related to analyte_chr.}

\item{analyte_chr}{(string) Precursor ID of the requested analyte.}

\item{XICs.ref.s}{(list of dataframes) Smoothed fragment-ion chromatograms of the analyte_chr from the reference run.}
}
\value{
(dataframe) aligned features of analyte_chr in eXp run.
}
\description{
df contains unaligned features for an analyte across multiple runs. This function aligns eXp run to
ref run and updates corresponding features.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
}
\seealso{
\code{\link{alignTargetedRuns}, \link{alignIthAnalyte}, \link{setAlignmentRank}, \link{getMultipeptide}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-26
}
\keyword{internal}
