% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstAlign.R
\name{alignToRefMST}
\alias{alignToRefMST}
\title{Aligns an analyte for an edge of MST}
\usage{
alignToRefMST(iNet, net, fileInfo, XICs, params, analytes, df, globalFits, RSE)
}
\arguments{
\item{iNet}{(integer) the index of edge to be aligned in the net.}

\item{net}{(matrix) each row represents an edge of MST.}

\item{fileInfo}{(data-frame) output of \code{\link{getRunNames}}.}

\item{XICs}{(list of dataframes) fragment-ion chromatograms of the analytes for all runs.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{analytes}{(string) precursor IDs of the requested peptide.}

\item{df}{(dataframe) a collection of features related to analytes.}

\item{globalFits}{(list) each element is either of class lm or loess. This is an output of \code{\link{getGlobalFits}}.}

\item{RSE}{(list) Each element represents Residual Standard Error of corresponding fit in globalFits.}
}
\value{
invisible NULL
}
\description{
df contains unaligned features for an analyte across multiple runs. This function aligns eXp run to
ref run and updates corresponding features.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
}
\seealso{
\code{\link{mstAlignRuns}, \link{MSTperBatch}, \link{setAlignmentRank}, \link{getMultipeptide}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-05-15
}
\keyword{internal}
