% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_runs.R
\name{alignTragetedRuns}
\alias{alignTragetedRuns}
\title{Align XICs of precursors across multiple Targeted-MS runs and outputs quantitative data matrix.}
\usage{
alignTragetedRuns(dataPath, alignType = "hybrid", oswMerged = TRUE,
  nameCutPattern = "(.*)(/)(.*)", maxFdrQuery = 0.05,
  maxFdrLoess = 0.01, spanvalue = 0.1, Metabolome = FALSE,
  normalization = "mean", simMeasure = "dotProductMasked",
  SgolayFiltOrd = 4, SgolayFiltLen = 9, goFactor = 0.125,
  geFactor = 40, cosAngleThresh = 0.3, OverlapAlignment = TRUE,
  dotProdThresh = 0.96, gapQuantile = 0.5, hardConstrain = FALSE,
  samples4gradient = 100, expRSE = 8, samplingTime = 3.4,
  RSEdistFactor = 3.5)
}
\value{
Saves intensity table in the current directory.
}
\description{
Align XICs of precursors across multiple Targeted-MS runs and outputs quantitative data matrix.
}
