% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytes.R
\name{analytesFromFeatures}
\alias{analytesFromFeatures}
\title{Outputs analytes below FDR}
\usage{
analytesFromFeatures(features, analyteFDR = 1, commonAnalytes = TRUE)
}
\arguments{
\item{features}{(list of data-frames) contains features and their properties identified in each run.}

\item{analyteFDR}{(numeric) only analytes that have m-score less than this, will be included in the output.}

\item{commonAnalytes}{(logical) TRUE: intersect across all runs, FASLE: union across all runs.}
}
\value{
a vector of integer.
}
\description{
Provides all found analytes or only common analytes that have m-score less than analyteFDR.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- getRunNames(dataPath = dataPath)
features <- getFeatures(fileInfo, maxFdrQuery = 1.00, runType = "DIA_proteomics")
\dontrun{
commonAnalytes <- analytesFromFeatures(features, analyteFDR = 0.01, commonAnalytes = TRUE)
}
}
\seealso{
\code{\link{getFeatures}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-04-14
}
\keyword{internal}
