% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_aligned_obj.R
\docType{methods}
\name{as.list,AlignObj-method}
\alias{as.list,AlignObj-method}
\title{Converts instances of class AlignObj into list}
\usage{
\S4method{as.list}{AlignObj}(x)
}
\arguments{
\item{x}{An object of class AlignObj}
}
\value{
list
}
\description{
Converts instances of class AlignObj into list
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2020-03-31
}
