% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqMass.R
\name{blobXICs}
\alias{blobXICs}
\title{Format XICs to blob}
\usage{
blobXICs(XICs, nativeIds, compress = TRUE)
}
\arguments{
\item{XICs}{(list) a list of data-frames. Each data frame has elution time and intensity of fragment-ion XIC.}

\item{compress}{(logical) if TRUE, time and intensity are compressed.}

\item{nativeId}{(integer) transition ID of the xic.}
}
\value{
(data.frame)
}
\description{
Format XICs to blob
}
\details{
DATA_TYPE is one of 0 = mz, 1 = intensity, 2 = rt \cr
COMPRESSION is one of 0 = no, 1 = zlib, 2 = np-linear,
 3 = np-slof, 4 = np-pic, 5 = np-linear + zlib,
  6 = np-slof + zlib, 7 = np-pic + zlib
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR)
XICs <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt"]][["4618"]]
nativeIds <- 27706:27711
\dontrun{
blobXICs(XICs, nativeIds)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-01-16
}
\keyword{internal}
