% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_sim_matrix.R
\name{calcNoBeefMaskGlobal}
\alias{calcNoBeefMaskGlobal}
\title{Hard constraining of similarity matrix outside of global fit window}
\usage{
calcNoBeefMaskGlobal(tRunAVec, tRunBVec, Fit, noBeef = 50)
}
\arguments{
\item{tRunAVec}{A numeric vector.}

\item{tRunBVec}{A numeric vector.}

\item{Fit}{An object of class "loess".}

\item{noBeef}{A parameter which defines the window size of no constraining.}
}
\description{
This function takes in timeVectors from both runs, a global-fit object and
sample-length of window of no constraining. Outside of window, all elements
of matrix are equally weighted.
}
