% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_sim_matrix.R
\name{calcNoBeefMaskGlobalWSlope}
\alias{calcNoBeefMaskGlobalWSlope}
\title{Soft constraining of similarity matrix outside of global fit window}
\usage{
calcNoBeefMaskGlobalWSlope(tRunAVec, tRunBVec, Fit, noBeef = 50)
}
\arguments{
\item{tRunAVec}{A numeric vector.}

\item{tRunBVec}{A numeric vector.}

\item{Fit}{An object of class "loess".}

\item{noBeef}{A parameter which defines the window size of no constraining.}
}
\description{
This function takes in datafile, precursor id, names of run pair. Based on
similarity measure, it calculates the similarity matrix. Outside of window,
elements of matrix are weighted proportional to distance from window-boundry.
}
