% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_area.R
\name{calculateIntensity}
\alias{calculateIntensity}
\title{Calculates area of a peak in XIC group}
\usage{
calculateIntensity(
  XICs,
  left,
  right,
  integrationType,
  baselineType,
  fitEMG = FALSE,
  baseSubtraction = TRUE,
  transitionIntensity = FALSE
)
}
\arguments{
\item{XICs}{(list) list of extracted ion chromatograms of a precursor.}

\item{left}{(numeric) left boundary of the peak.}

\item{right}{(numeric) right boundary of the peak.}

\item{integrationType}{(string) method to ompute the area of a peak contained in XICs. Must be
from "intensity_sum", "trapezoid", "simpson".}

\item{baselineType}{(string) method to estimate the background of a peak contained in XICs. Must be
from "base_to_base", "vertical_division_min", "vertical_division_max".}

\item{fitEMG}{(logical) enable/disable exponentially modified gaussian peak model fitting.}

\item{baseSubtraction}{(logical) TRUE: remove background from peak signal using estimated noise levels.}

\item{transitionIntensity}{(logical) TRUE: return intensity of each transition, FALSE: return sum of all transitions.}
}
\value{
(numeric)
}
\description{
Retention time from reference run is mapped to experiment run using AlignObj.
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
XICs <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
\dontrun{
calculateIntensity(XICs, 5220, 5261, integrationType = "intensity_sum",
 baselineType = "base_to_base", fitEMG = FALSE)
}
}
\seealso{
\code{\link{getMultipeptide}, \link{setAlignmentRank}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-04-13
}
\keyword{internal}
