% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkOverlap}
\alias{checkOverlap}
\title{Overlap of two time ranges}
\usage{
checkOverlap(x, y)
}
\arguments{
\item{x}{(numeric) must have only two values and x[2] > x[1].}

\item{y}{(numeric) must have only two values and y[2] > y[1].}
}
\value{
(logical) TRUE: both time ranges overlap. FALSE: both time ranges do not overlap.
}
\description{
Overlap of two time ranges
}
\examples{
\dontrun{
checkOverlap(c(9.1, 13.1), c(2.1, 3.1))
checkOverlap(c(1.1, 3.1), c(3.2, 7.1))
}
}
\seealso{
getChildFeature
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-17
}
\keyword{internal}
