% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkParams}
\alias{checkParams}
\title{Checks all the alignment parameters}
\usage{
checkParams(params)
}
\arguments{
\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}
}
\value{
Invisible NULL
}
\description{
Parameters are defined in the \code{\link{paramsDIAlignR}} function. If parameters are found
incongruous, the program will terminate.
}
\examples{
params <- paramsDIAlignR()
\dontrun{
checkParams(params)
}
}
\seealso{
\code{\link{paramsDIAlignR}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-01
}
\keyword{internal}
