% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_chromatogram.R
\name{childXIC}
\alias{childXIC}
\title{Get child chromatogram from parents}
\usage{
childXIC(
  XIC.ref,
  XIC.eXp,
  alignedIndices,
  method = "spline",
  polyOrd = 4,
  kernelLen = 9,
  splineMethod = "fmm"
)
}
\arguments{
\item{alignedIndices}{(data-frame) must have two columns "indexAligned.ref" and "indexAligned.eXp".}

\item{method}{(string) must be either "spline", "sgolay" or "linear".}

\item{polyOrd}{(integer) must be less than kernelLen.}

\item{kernelLen}{(integer) must be an odd integer.}

\item{splineMethod}{(string) must be either "fmm" or "natural".}

\item{XICs.ref}{(data-frame) extracted ion chromatograms from reference run.}

\item{XICs.eXp}{(data-frame) extracted ion chromatograms from experiment run.}
}
\value{
(data-frame) a chromatogram.
}
\description{
Get child chromatogram from parents
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
data(alignObj_DIAlignR, package="DIAlignR")
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run1"]][["14299_QFNNTDIVLLEDFQK/3"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run2"]][["14299_QFNNTDIVLLEDFQK/3"]]
alignedIndices <- cbind(alignObj_DIAlignR@indexA_aligned, alignObj_DIAlignR@indexB_aligned)
colnames(alignedIndices) <- c("indexAligned.ref", "indexAligned.eXp")
alignedIndices[, 1:2][alignedIndices[, 1:2] == 0] <- NA_integer_
\dontrun{
childXIC(XICs.ref[[1]], XICs.eXp[[1]], alignedIndices)
}
}
\seealso{
\code{\link{mergeXIC}, \link{alignedXIC}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-23
}
\keyword{internal}
