% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_chromatogram.R
\name{childXICs}
\alias{childXICs}
\title{Get child chromatograms from parents}
\usage{
childXICs(
  XICs.ref,
  XICs.eXp,
  alignedIndices,
  method = "spline",
  polyOrd = 4,
  kernelLen = 9,
  splineMethod = "fmm",
  w.ref = 0.5,
  mergeStrategy = "avg",
  keepFlanks = TRUE
)
}
\arguments{
\item{XICs.ref}{(list of data-frames) extracted ion chromatograms from reference run.}

\item{XICs.eXp}{(list of data-frames) extracted ion chromatograms from experiment run.}

\item{alignedIndices}{(data-frame) must have two columns "indexAligned.ref" and "indexAligned.eXp".}

\item{method}{(string) must be either "spline", "sgolay" or "linear".}

\item{polyOrd}{(integer) must be less than kernelLen.}

\item{kernelLen}{(integer) must be an odd integer.}

\item{splineMethod}{(string) must be either "fmm" or "natural".}

\item{mergeStrategy}{(string) must be either ref, avg, refStart or refEnd.}

\item{keepFlanks}{(logical) TRUE: Flanking chromatogram is not removed.}
}
\value{
(list) the first element is a list of chromatograms. The second element is aligned parent time-vectors.
}
\description{
Get child chromatograms from parents
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
data(alignObj_DIAlignR, package="DIAlignR")
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
alignedIndices <- cbind(alignObj_DIAlignR@indexA_aligned, alignObj_DIAlignR@indexB_aligned)
colnames(alignedIndices) <- c("indexAligned.ref", "indexAligned.eXp")
alignedIndices[, 1:2][alignedIndices[, 1:2] == 0] <- NA_integer_
newXICs <- childXICs(XICs.ref, XICs.eXp, alignedIndices)[[1]]
plotXICgroup(XICs.ref)
plotXICgroup(newXICs)
}
\seealso{
\code{\link{childXIC}, \link{mergeStrategy}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-23
}
