% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osw_mzml.R
\name{chromatogramIdAsInteger}
\alias{chromatogramIdAsInteger}
\title{Selects only chromatogramId, chromatogramIndex columns and convert them into integer.}
\usage{
chromatogramIdAsInteger(chromHead)
}
\value{
Invisible NULL
}
\description{
Selects only chromatogramId, chromatogramIndex columns and convert them into integer.
}
