% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrain_sim_matrix.R
\name{constrainSimilarity}
\alias{constrainSimilarity}
\title{Hard constraining of similarity matrix outside of global fit window}
\usage{
constrainSimilarity(s, mask, unitConstrain)
}
\arguments{
\item{s}{A numeric similarity matrix.}

\item{mask}{A numeric matrix with weights for constraining.}

\item{unitConstrain}{Constraining value for unit weight.}
}
\description{
This function takes in timeVectors from both runs, a global-fit object and
sample-length of window of no constraining. Outside of window, all elements
of matrix are equally weighted.
}
