% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{constrainSimMain}
\alias{constrainSimMain}
\title{Constrain similarity matrix using mask and constrainValue.}
\usage{
constrainSimMain(sim, MASK, samples4gradient = 100)
}
\arguments{
\item{ROW_SIZE}{(int) Number of rows of a matrix}

\item{COL_SIZE}{(int) Number of columns of a matrix}
}
\value{
affineAlignObj (S4class) A S4class dummy object from C++ AffineAlignObj struct
}
\description{
Constrain similarity matrix using mask and constrainValue.
}
\examples{

}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
