% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyopenms.R
\name{createMZML}
\alias{createMZML}
\title{Create an mzML file}
\usage{
createMZML(ropenms, filename, XICs, transitionIDs)
}
\arguments{
\item{ropenms}{(pyopenms module) get this python module through get_ropenms().}

\item{filename}{(string) Name of the mzML file to be written.}

\item{XICs}{(list of data-frames) extracted ion chromatograms.}

\item{transitionIDs}{(integer) length must be the same as of XICs.}
}
\value{
(None)
}
\description{
Writes an mzML file having chromatograms and their native IDs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
filename <- paste0(dataPath, "/mzml/hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML")
mz <- mzR::openMSfile(filename, backend = "pwiz")
chromHead <- mzR::chromatogramHeader(mz)
XICs <- mzR::chromatograms(mz, chromHead$chromatogramIndex[1:10])
nativeIds <- chromHead$chromatogramId[1:10]
ropenms <- get_ropenms(condaEnv = "TricEnvr")
createMZML(ropenms, "testfile.mzML", XICs, nativeIds)
XICs <- mzR::chromatograms(mzR::openMSfile("testfile.mzML", backend = "pwiz"))
}
\seealso{
\code{\link{get_ropenms}, \link{addXIC}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-06
}
