% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIAlignR.R
\docType{package}
\name{DIAlignR}
\alias{DIAlignR}
\title{DIAlignR}
\description{
This package implements dynamic programming with affine gap penalty to find a highest-scoring scoring path. A hybrid approach of global alignment through MS2 features and local alignment with MS2 chromatograms is implemented in this tool.
}
\author{
Shubham Gupta, Hannes Rost
}
