% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_fit.R
\name{dialignrLoess}
\alias{dialignrLoess}
\title{Modified loess for condition handling}
\usage{
dialignrLoess(RUNS_RT, spanvalue)
}
\arguments{
\item{RUNS_RT}{(data-frame) must have three calumns: transition_group_id, RT.eXp, and RT.ref.}

\item{spanvalue}{(numeric) spanvalue for LOESS fit. For targeted proteomics 0.1 could be used.}
}
\value{
An object of class "loess" or "lm".
}
\description{
Modified loess for condition handling
}
\examples{
df <- data.frame("transition_group_id" = 1:10, "RT.eXp" = 2:11, "RT.ref" = 10:19)
\dontrun{
dialignrLoess(df, 0.1)
dialignrLoess(df[1:4,], 0.1)
}
}
\seealso{
\code{\link{getLinearfit}, \link{getLOESSfit}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-11
}
\keyword{internal}
