% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doAlignmentCpp}
\alias{doAlignmentCpp}
\title{Perform non-affine global and overlap alignment on a similarity matrix}
\usage{
doAlignmentCpp(sim, gap, OverlapAlignment)
}
\arguments{
\item{sim}{(NumericMatrix) A numeric matrix with similarity values of two sequences or signals.}

\item{gap}{(double) Penalty for introducing gaps in alignment.}

\item{OverlapAlignment}{(logical) An input for alignment with free end-gaps. False: Global alignment, True: overlap alignment.}
}
\value{
AlignObj (S4class) An object from C++ class of AlignObj.
}
\description{
Perform non-affine global and overlap alignment on a similarity matrix
}
\examples{
# Get sequence similarity of two DNA strings
Match=10; MisMatch=-2
seq1 = "GCAT"; seq2 = "CAGTG"
s <- getSeqSimMatCpp(seq1, seq2, Match, MisMatch)
obj_Global <- doAlignmentCpp(s, 22, FALSE)
obj_Global@score # -2 -4 -6 4 -18
obj_Olap <- doAlignmentCpp(s, 22, TRUE)
obj_Olap@score # 0 10 20 18 18 18

Match=1; MisMatch=-1
seq1 = "TTTC"; seq2 = "TGC"
s <- getSeqSimMatCpp(seq1, seq2, Match, MisMatch)
obj_Global <- doAlignmentCpp(s, 2, FALSE)
obj_Global@optionalPaths
matrix(data = c(1,1,1,1,1,1,1,1,1,2,1,2,1,3,3,1,1,3,6,3), nrow = 5, ncol =4, byrow = TRUE)
obj_Global@M_forw
matrix(data = c(0,-2,-4,-6,-2,-7,-22,-45,-4,-20,-72,-184,-6,-41,-178,-547,-8,-72,-366,-1274), nrow = 5, ncol =4, byrow = TRUE)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
