% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{extractXIC_group}
\alias{extractXIC_group}
\title{Extract XICs of chromIndices}
\usage{
<<<<<<< HEAD
extractXIC_group(
  mz,
  chromIndices,
  XICfilter = "sgolay",
  SgolayFiltOrd = 4,
  SgolayFiltLen = 9
)
=======
extractXIC_group(mz, chromIndices)
>>>>>>> a73122ae9a62d6a347f9ff0e7d9569beb8752b66
}
\arguments{
\item{mz}{(mzRpwiz object)}

\item{chromIndices}{(vector of Integers) Indices of chromatograms to be extracted.}
}
\value{
A list of data-frames. Each data frame has elution time and intensity of fragment-ion XIC.
}
\description{
Extracts XICs using mz object. Each chromatogram represents a transition of precursor.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
mzmlName<-paste0(dataPath,"/mzml/hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML")
mz <- mzR::openMSfile(mzmlName, backend = "pwiz")
chromIndices <- c(37L, 38L, 39L, 40L, 41L, 42L)
\dontrun{
XIC_group <- extractXIC_group(mz, chromIndices)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
\keyword{internal}
