% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{extractXIC_group2}
\alias{extractXIC_group2}
\title{Extract XICs of chromIndices}
\usage{
extractXIC_group2(con, chromIndices)
}
\arguments{
\item{chromIndices}{(vector of Integers) Indices of chromatograms to be extracted.}

\item{mz}{(SQLiteConnection object)}
}
\value{
A list of data-frames. Each data frame has elution time and intensity of fragment-ion XIC.
}
\description{
DATA_TYPE is one of 0 = mz, 1 = intensity, 2 = rt
Extracts XICs using connection to sqMass file Each chromatogram represents a transition of precursor.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
sqName <- paste0(dataPath,"/xics/hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt.chrom.sqMass")
chromIndices <- c(36L, 37L, 38L, 39L, 40L, 41L)
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = sqName)
XIC_group <- extractXIC_group2(con, chromIndices)
DBI::dbDisconnect(con)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-12-25
}
\keyword{internal}
