% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osw.R
\name{fetchAnalytesInfo}
\alias{fetchAnalytesInfo}
\title{Get a table of analytes, their chromatogram indices and scores.}
\usage{
fetchAnalytesInfo(oswName, maxFdrQuery, oswMerged, analytes, filename,
  runType)
}
\value{
A data-frame.
}
\description{
Get a table of analytes, their chromatogram indices and scores.
}
