% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osw.R
\name{fetchPeptidesInfo}
\alias{fetchPeptidesInfo}
\title{Get scores of all peptides}
\usage{
fetchPeptidesInfo(oswName, runType, context)
}
\arguments{
\item{oswName}{(char) path to the osw file.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{context}{(string) Context used in pyprophet peptide. Must be either "run-specific", "experiment-wide", or "global".}
}
\value{
(dataframe) with following columns:
\item{peptide_id}{(integer) a unique id for each precursor.}
\item{run}{(character) retention time as in FEATURE.EXP_RT of osw files.}
\item{score}{(numeric) peak intensity as in FEATURE_MS2.AREA_INTENSITY of osw files.}
\item{pvalue}{(numeric) as in FEATURE.LEFT_WIDTH of osw files.}
\item{qvalue}{(numeric) as in FEATURE.RIGHT_WIDTH of osw files.}
}
\description{
Return a scores, pvalues, and qvalues for all peptides from the osw file.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- getRunNames(dataPath = dataPath)
oswName <- fileInfo[["featureFile"]][1]
\dontrun{
precursorsInfo <- fetchPeptidesInfo(fileInfo, runType = "DIA_proteomics", context = "experiment-wide")
}
}
\seealso{
\code{\link{getPeptideQuery}, \link{getPeptideScores}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-01
}
\keyword{internal}
