% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osw.R
\name{fetchTransitionsFromRun}
\alias{fetchTransitionsFromRun}
\title{Get transitions from a feature file}
\usage{
fetchTransitionsFromRun(
  filename,
  runID,
  maxFdrQuery = 1,
  runType = "DIA_proteomics"
)
}
\arguments{
\item{filename}{(string) Path to the feature file.}

\item{runID}{(string) id in RUN.ID column of the feature file.}

\item{maxFdrQuery}{(numeric) A numeric value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}
}
\value{
(data-frames) Data-frame has following columns:
\item{transition_group_id}{(integer) a unique id for each precursor.}
\item{RT}{(numeric) retention time as in FEATURE.EXP_RT of osw files.}
\item{intensity}{(list) of peak intensities as in FEATURE_TRANSITION.AREA_INTENSITY of osw files.}
\item{leftWidth}{(numeric) as in FEATURE.LEFT_WIDTH of osw files.}
\item{rightWidth}{(numeric) as in FEATURE.RIGHT_WIDTH of osw files.}
\item{peak_group_rank}{(integer) rank of each feature associated with transition_group_id.}
\item{m_score}{(numeric) q-value of each feature associated with transition_group_id.}
}
\description{
Get a data-frame of OpenSwath features that contains retention time, transition intensities, boundaries etc.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- getRunNames(dataPath = dataPath)
\dontrun{
transitionsInfo <- fetchTransitionsFromRun(fileInfo$featureFile[1], fileInfo$spectraFileID[1],
 maxFdrQuery = 0.05)
dim(transitionsInfo) # 211  8
}
}
\seealso{
\code{\link{getRunNames}, \link{getTransitions}, \link{getTransitionsQuery}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-11-15
}
\keyword{internal}
