% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filenames.R
\name{filenamesFromMZML}
\alias{filenamesFromMZML}
\title{Get mzML filenames from the directory.}
\usage{
filenamesFromMZML(dataPath)
}
\value{
A dataframe with single column.
}
\description{
Get mzML filenames from the directory.
}
