% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filenames.R
\name{filenamesFromOSW}
\alias{filenamesFromOSW}
\title{Get mzML filenames from osw RUN table.}
\usage{
filenamesFromOSW(dataPath, pattern)
}
\value{
A dataframe with single column.
}
\description{
Get mzML filenames from osw RUN table.
}
