% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{fillOswFiles}
\alias{fillOswFiles}
\title{Get list of peptides and their chromatogram indices.}
\usage{
fillOswFiles(oswFiles, dataPath, filenames, maxFdrQuery, oswMerged,
  peptides, runType)
}
\value{
A list of data-frames.
}
\description{
Get list of peptides and their chromatogram indices.
}
