% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{get1OswFiles}
\alias{get1OswFiles}
\title{Get list of peptides and their chromatogram indices.}
\usage{
get1OswFiles(filenames, dataPath = ".", peptides = NULL,
  query = NULL, oswMerged = TRUE, maxFdrQuery = 1,
  nameCutPattern = "(.*)(/)(.*)", runType = "DIA_Proteomics")
}
\value{
A list of data-frames.
}
\description{
Get list of peptides and their chromatogram indices.
}
