% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyopenms.R
\name{get_ropenms}
\alias{get_ropenms}
\title{Get ropenms}
\usage{
get_ropenms(pythonPath = NULL, condaEnv = NULL, useConda = TRUE)
}
\arguments{
\item{pythonPath}{(string) path of python. Must have pyopenms module.}

\item{condaEnv}{(string) name of the conda environment that has pyopenms module.}

\item{useConda}{(logical) TRUE: Use conda environment. FALSE: Use python through pythonPath.}
}
\value{
(pyopenms module)
}
\description{
Python path can also be set using Sys.setenv(RETICULATE_PYTHON = pythonPath).
}
\examples{
ropenms <- get_ropenms(condaEnv = "TricEnvr", useConda=TRUE)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-06
}
