% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_aligned_obj.R
\name{getAffineAlignObj}
\alias{getAffineAlignObj}
\title{Creates affine alignment object}
\usage{
getAffineAlignObj(s, go, ge, FreeEndGaps = TRUE)
}
\arguments{
\item{s}{A numeric matrix}

\item{go}{A numeric value}

\item{ge}{A numeric value}

\item{FreeEndGaps}{A logical scalar}
}
\value{
A S4 object
}
\description{
This function calculates three matrices for affine gap alignment using input
similarity matrix and affine gap opening and gap closing penalties. An
implementation of Needleman-Wunsch alignment and overlap alignment is also
provided. All three matrices are clubbed together in an output S4 object.
}
\examples{
M = 10; MM = -2; go = 22; ge = 7
seq1 = "GCAT"
seq2 = "CAGTG"
s = matrix(NA, nrow = nchar(seq1), ncol= nchar(seq2))
for(i in 1:nrow(s)){
   for(j in 1: ncol(s)){
       if(substr(seq1, start = i, stop = i) == substr(seq2, start = j, stop = j)) s[i,j] <- M
       else s[i,j] <- MM
   }
}
alignObj <- getAffineAlignObj(s, go, ge, FreeEndGaps=FALSE)
}
