% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{getAlignedFigs}
\alias{getAlignedFigs}
\title{Plot aligned XICs group for a specific peptide.
AlignObjOutput is the output from getAlignObjs fucntion.}
\usage{
getAlignedFigs(AlignObj, XICs.ref, XICs.eXp, refPeakLabel,
  annotatePeak = FALSE)
}
\description{
Plot aligned XICs group for a specific peptide.
AlignObjOutput is the output from getAlignObjs fucntion.
}
