% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getAlignedInices_S4}
\alias{getAlignedInices_S4}
\title{Initialize a S4 object AffineAlignObj1}
\usage{
getAlignedInices_S4(obj)
}
\arguments{
\item{seq1Len}{(int) Length of sequence1}

\item{seq2Len}{(int) Length of sequence2}
}
\value{
affineAlignObj (S4class) An object from C++ class of AffineAlignObj
}
\description{
Initialize a S4 object AffineAlignObj1
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
