% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pep_peak_corp.R
\name{getAlignedTimesFast}
\alias{getAlignedTimesFast}
\title{Get aligned Retention times.}
\usage{
getAlignedTimesFast(XICs.ref, XICs.eXp, globalFit, adaptiveRT, params)
}
\arguments{
\item{XICs.ref}{List of extracted ion chromatograms from reference run.}

\item{XICs.eXp}{List of extracted ion chromatograms from experiment run.}

\item{globalFit}{Linear or loess fit object between reference and experiment run.}

\item{adaptiveRT}{(numeric) Similarity matrix is not penalized within adaptive RT.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}
}
\value{
(matrix) the first column corresponds to the aligned reference time, the second column is the aligned experiment time.
}
\description{
This function aligns XICs of reference and experiment runs.
It produces aligned retention times between refernce run and experiment run.
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
data(oswFiles_DIAlignR, package="DIAlignR")
run1 <- "hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"
run2 <- "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"
XICs.ref <- lapply(XIC_QFNNTDIVLLEDFQK_3_DIAlignR[[run1]][["4618"]], as.matrix)
XICs.eXp <- lapply(XIC_QFNNTDIVLLEDFQK_3_DIAlignR[[run2]][["4618"]], as.matrix)
params <- paramsDIAlignR()
params[["globalAlignment"]] <- "linear"
globalFit <- getGlobalAlignment(oswFiles_DIAlignR, ref = "run2", eXp = "run0",
 fitType = params[["globalAlignment"]], maxFdrGlobal = 0.05, spanvalue = 0.1)
adaptiveRT <- 77.82315 #3.5*getRSE(globalFit, params[["globalAlignment"]])
globalFit <- coef(globalFit)
getAlignedTimesFast(XICs.ref, XICs.eXp, globalFit, adaptiveRT, params)
}
\seealso{
\code{\link{alignChromatogramsCpp}, \link{getAlignObj}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-01-02
}
