% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_alignment_path.R
\name{getAlignment}
\alias{getAlignment}
\title{Computes indices of row and column of maximum scored traceback path}
\usage{
getAlignment(alignObj)
}
\arguments{
\item{alignObj}{An object of class NeedleObjAffine}
}
\value{
A list of matrices having row and column indices of traceback path.
}
\description{
This function takes aligned object of the class NeedleObjAffine as an input.
With the provided traceback path, it calculates row and column index pair
associated with the highest scoring path through similarity matrix. Output is
a list of row-column index pairs of all highest scoring traceback paths.
}
