% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_alignment_path.R
\name{getAlignmentPath}
\alias{getAlignmentPath}
\title{Output matrix with alignment path represented as binary-hot encoding.}
\usage{
getAlignmentPath(alignedIndicesObj, s)
}
\arguments{
\item{alignedIndicesObj}{A numeric matrix}

\item{s}{A numeric matrix}
}
\value{
A numeric matrix
}
\description{
This function outputs matix of the same size of input similarity matrix s.
The output matrix has unit cell value for row and column indices given in
the matrix alignedIndicesObj.
}
