% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dia_runs.R
\name{getAlignObjs}
\alias{getAlignObjs}
\title{AlignObj for analytes between a pair of runs}
\usage{
getAlignObjs(
  analytes,
  runs,
  dataPath = ".",
  refRun = NULL,
  oswMerged = TRUE,
  params = paramsDIAlignR(),
  objType = "light"
)
}
\arguments{
\item{analytes}{(vector of integers) transition_group_ids for which features are to be extracted.}

\item{runs}{(a vector of string) names of mzml file without extension.}

\item{dataPath}{(string) path to mzml and osw directory.}

\item{refRun}{(string) reference for alignment. If no run is provided, m-score is used to select reference run.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{objType}{(char) Must be selected from light, medium and heavy.}
}
\value{
A list of fileInfo and AlignObjs. Each AlignObj is an S4 object. Three most-important slots are:
\item{indexA_aligned}{(integer) aligned indices of reference run.}
\item{indexB_aligned}{(integer) aligned indices of experiment run.}
\item{score}{(numeric) cumulative score of alignment.}
}
\description{
This function expects osw and mzml directories at dataPath. It first reads osw files and fetches chromatogram indices for each requested analyte.
It then align XICs of each analyte to its reference XICs. AlignObj is returned which contains aligned indices and cumulative score along the alignment path.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
params <- paramsDIAlignR()
params[["context"]] <- "experiment-wide"
runs <- c("hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt",
 "hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt",
 "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt")
analytes <- c(32L, 898L, 2474L)
AlignObjOutput <- getAlignObjs(analytes, runs, dataPath = dataPath)
plotAlignedAnalytes(AlignObjOutput)

}
\references{
Gupta S, Ahadi S, Zhou W, Röst H. "DIAlignR Provides Precise Retention Time Alignment Across Distant Runs in DIA and Targeted Proteomics." Mol Cell Proteomics. 2019 Apr;18(4):806-817. doi: https://doi.org/10.1074/mcp.TIR118.001132 Epub 2019 Jan 31.
}
\seealso{
\code{\link{plotAlignedAnalytes}, \link{getRunNames}, \link{getFeatures}, \link{getXICs4AlignObj}, \link{getAlignObj}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
