% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getAnalytesName}
\alias{getAnalytesName}
\title{Get names of analytes found in all runs.}
\usage{
getAnalytesName(runs, dataPath = ".", query = NULL, oswMerged = TRUE,
  nameCutPattern = "(.*)(/)(.*)", maxFdrQuery = 0.05)
}
\value{
A vector of strings.
}
\description{
Get names of analytes found in all runs.
}
