% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osw_query.R
\name{getAnalytesQuery}
\alias{getAnalytesQuery}
\title{Generate SQL query to fetch limited information from osw files.}
\usage{
getAnalytesQuery(
  maxFdrQuery,
  oswMerged = TRUE,
  filename = NULL,
  runType = "DIA_Proteomics",
  analyteInGroupLabel = FALSE
)
}
\arguments{
\item{maxFdrQuery}{(numeric) value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{filename}{(string) as mentioned in RUN table of osw files..}

\item{runType}{(char) This must be one of the strings "DIA_Proteomics", "DIA_Metabolomics".}

\item{analyteInGroupLabel}{(logical) TRUE for getting analytes as PRECURSOR.GROUP_LABEL from osw file.}
}
\value{
SQL query to be searched.
}
\description{
Generate SQL query to fetch limited information from osw files.
}
\seealso{
\code{\link{getOswAnalytes}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
\keyword{internal}
