% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osw_mzml.R
\name{getChromatogramIndices}
\alias{getChromatogramIndices}
\title{Get chromatogram indices of precursors.}
\usage{
getChromatogramIndices(fileInfo, precursors, mzPntrs, applyFun = lapply)
}
\arguments{
\item{fileInfo}{(data-frame) Output of getRunNames function.}

\item{precursors}{(data-frame) Atleast two columns transition_group_id and transition_ids are required.}

\item{mzPntrs}{A list of mzRpwiz.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\value{
(list) A list of dataframes having following columns:
\item{transition_group_id}{(string) it is PRECURSOR.ID from osw file.}
\item{chromatogramIndex}{(integer) index of chromatogram in mzML file.}
}
\description{
This function reads the header of chromatogram files. It then fetches chromatogram indices by matching transition_id(osw) with chromatogramID(mzml).
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- DIAlignR::getRunNames(dataPath = dataPath)
precursors <- getPrecursors(fileInfo, oswMerged = TRUE, context = "experiment-wide")
mzPntrs <- getMZMLpointers(fileInfo)
prec2chromIndex <- getChromatogramIndices(fileInfo, precursors, mzPntrs)
rm(mzPntrs)
}
\seealso{
\code{\link{chromatogramIdAsInteger}, \link{mapPrecursorToChromIndices}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-04-07
}
