% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getChromSimMat}
\alias{getChromSimMat}
\title{Calculate similarity matrix of two fragment-ion chromatogram group.}
\usage{
getChromSimMat(l1, l2, Normalization, SimType, dotProdThresh = 0.96,
  cosAngleThresh = 0.3)
}
\arguments{
\item{l1}{(list) A list of vectors. Length should be same as of l2.}

\item{l2}{(list) A list of vectors. Length should be same as of l1.}

\item{Normalization}{(char) Normalization to be used. L2, mean}

\item{SimType}{(char) Method for calculating the similarity matrix. dotProductMasked, dotProduct, cosineAngle, cosine2Angle, euclidianDist, covariance, correlation.}
}
\value{
s (matrix) Numeric similarity matrix. Rows and columns expresses seq1 and seq2, respectively
}
\description{
Calculate similarity matrix of two fragment-ion chromatogram group.
}
\examples{
# Get similarity matrix of dummy chromatograms
r1 <- list(c(1.0,3.0,2.0,4.0), c(0.0,0.0,0.0,1.0), c(4.0,4.0,4.0,5.0))
r2 <- list(c(1.4,2.0,1.5,4.0), c(0.0,0.5,0.0,0.0), c(2.0,3.0,4.0,0.9))
getChromSimMat(r1, r2, "L2", "dotProduct")
matrix(c(0.1251213, 0.1623915, 0.1437564, 0.2076481, 0.1863509,
 0.2395940, 0.2129724, 0.3128033, 0.2329386, 0.2728709, 0.2529048,
  0.3460802, 0.1011619, 0.2076481, 0.1544050, 0.2728709), 4, 4, byrow = F)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-05
}
