% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osw_query.R
\name{getFeaturesQuery}
\alias{getFeaturesQuery}
\title{Get features from a SQLite file}
\usage{
getFeaturesQuery(runType = "DIA_Proteomics")
}
\arguments{
\item{runType}{(char) This must be one of the strings "DIA_Proteomics", "DIA_Metabolomics".}
}
\value{
SQL query to be searched.
}
\description{
Query is generated to identify features below a FDR cut-off from a run.
}
\seealso{
\code{\link{fetchFeaturesFromRun}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2020-04-07
}
\keyword{internal}
