% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gap_penalty.R
\name{getGapPenalty}
\alias{getGapPenalty}
\title{Calculates gap penalty}
\usage{
getGapPenalty(s, gapQuantile, type = "dotProduct")
}
\arguments{
\item{s}{A numeric matrix}
}
\value{
The gap penalty for dynamic programming
}
\description{
Takes a matrix, vectorize it and find the given quantile. This value is given out as gap penalty.
}
