% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_fit.R
\name{getGlobalAlignment}
\alias{getGlobalAlignment}
\title{Calculates global alignment between RT of two runs}
\usage{
getGlobalAlignment(oswFiles, ref, eXp, fitType = "linear",
  maxFdrGlobal = 0.01, spanvalue = 0.1)
}
\arguments{
\item{oswFiles}{(list of data-frames) it is output from getOswFiles function.}

\item{ref}{(string) Must be a combination of "run" and an iteger e.g. "run2".}

\item{eXp}{(string) Must be a combination of "run" and an iteger e.g. "run2".}

\item{fitType}{(char) Must be from "loess" or "linear".}

\item{maxFdrGlobal}{(numeric) A numeric value between 0 and 1. Features should have m-score lower than this value for participation in global fit.}

\item{spanvalue}{(numeric) Spanvalue for LOESS fit. For targeted proteomics 0.1 could be used.}
}
\value{
An object of class "loess".
}
\description{
This function selects features from oswFiles which has m-score < maxFdrLoess. It then fit linear/loess regression on these feature.
Retention-time mapping is established from reference to experiment run.
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
Loess.fit <- getGlobalAlignment(oswFiles = oswFiles_DIAlignR, ref = "run1", eXp = "run2",
 maxFdrGlobal = 0.05, spanvalue = 0.1, fit = "loess")
}
\seealso{
\code{\link{getOswFiles}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
