% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getListObj}
\alias{getListObj}
\title{Find and Extract an obj from a structured list}
\usage{
getListObj(x, name)
}
\arguments{
\item{x}{A list object}

\item{name}{Name of sub-components in list}
}
\value{
returns the named object in list
}
\description{
Find and Extract an obj from a structured list
}
