% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_fit.R
\name{getLOESSfit}
\alias{getLOESSfit}
\title{Calculates LOESS fit between RT of two runs}
\usage{
getLOESSfit(oswFiles, ref, eXp, maxFdrGlobal, spanvalue = 0.1)
}
\arguments{
\item{oswFiles}{(list of data-frames) it is output from getOswFiles function.}

\item{ref}{(string) Must be a combination of "run" and an iteger e.g. "run2".}

\item{eXp}{(string) Must be a combination of "run" and an iteger e.g. "run2".}

\item{maxFdrGlobal}{(numeric) A numeric value between 0 and 1. Features should have m-score lower than this value for participation in LOESS fit.}

\item{spanvalue}{(numeric) Spanvalue for LOESS fit. For targeted proteomics 0.1 could be used.}
}
\value{
An object of class "loess".
}
\description{
This function selects features from oswFiles which has m-score < maxFdrLoess. It then fit LOESS on these feature.
Loess mapping is established from reference to experiment run.
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
Loess.fit <- getLOESSfit(oswFiles = oswFiles_DIAlignR, ref = "run1", eXp = "run2", maxFdrGlobal = 0.05, spanvalue = 0.1)
}
\seealso{
\code{\link{getLinearfit}, \link{getOswFiles}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-14
}
