% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pep_peak_corp.R
\name{getMappedRT}
\alias{getMappedRT}
\title{Get mapping of reference RT on experiment run.}
\usage{
getMappedRT(refRT, XICs.ref, XICs.eXp, globalFit, alignType, adaptiveRT,
  normalization, simMeasure, goFactor, geFactor, cosAngleThresh,
  OverlapAlignment, dotProdThresh, gapQuantile, hardConstrain,
  samples4gradient, objType = "light")
}
\arguments{
\item{refRT}{Peak's retention-time in reference run.}

\item{XICs.ref}{List of extracted ion chromatograms from reference run.}

\item{XICs.eXp}{List of extracted ion chromatograms from experiment run.}

\item{globalFit}{Linear or loess fit object between reference and experiment run.}

\item{alignType}{Available alignment methods are "global", "local" and "hybrid".}

\item{adaptiveRT}{(numeric) Similarity matrix is not penalized within adaptive RT.}

\item{normalization}{(character) Must be selected from "mean", "l2".}

\item{simMeasure}{(string) Must be selected from dotProduct, cosineAngle,
cosine2Angle, dotProductMasked, euclideanDist, covariance and correlation.}

\item{goFactor}{(numeric) Penalty for introducing first gap in alignment. This value is multiplied by base gap-penalty.}

\item{geFactor}{(numeric) Penalty for introducing subsequent gaps in alignment. This value is multiplied by base gap-penalty.}

\item{cosAngleThresh}{(numeric) In simType = dotProductMasked mode, angular similarity should be higher than cosAngleThresh otherwise similarity is forced to zero.}

\item{OverlapAlignment}{(logical) An input for alignment with free end-gaps. False: Global alignment, True: overlap alignment.}

\item{dotProdThresh}{(numeric) In simType = dotProductMasked mode, values in similarity matrix higher than dotProdThresh quantile are checked for angular similarity.}

\item{gapQuantile}{(numeric) Must be between 0 and 1. This is used to calculate base gap-penalty from similarity distribution.}

\item{hardConstrain}{(logical) If FALSE; indices farther from noBeef distance are filled with distance from linear fit line.}

\item{samples4gradient}{(numeric) This parameter modulates penalization of masked indices.}

\item{objType}{(char) Must be selected from light, medium and heavy.}
}
\value{
(numeric)
}
\description{
This function aligns XICs of reference and experiment runs. Using alignment, it maps retention time from refernce run on experiment run.
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
data(oswFiles_DIAlignR, package="DIAlignR")
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run1"]][["14299_QFNNTDIVLLEDFQK/3"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run2"]][["14299_QFNNTDIVLLEDFQK/3"]]
globalFit <- getGlobalAlignment(oswFiles_DIAlignR, ref = "run2", eXp = "run0",
 maxFdrGlobal = 0.05, spanvalue = 0.1)
adaptiveRT <- 77.82315 #3.5*globalFit$s
\dontrun{
getMappedRT(refRT = 5238.35, XICs.ref, XICs.eXp, globalFit, alignType = "hybrid",
 adaptiveRT = adaptiveRT, normalization = "mean",
  simMeasure = "dotProductMasked", goFactor = 0.125, geFactor = 40, cosAngleThresh = 0.3,
  OverlapAlignment = TRUE, dotProdThresh = 0.96, gapQuantile = 0.5, hardConstrain = FALSE,
  samples4gradient = 100)
}
}
\seealso{
\code{\link{alignChromatogramsCpp}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
\keyword{internal}
