% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_metabo_runs.R
\name{getMetaboAlignObjs}
\alias{getMetaboAlignObjs}
\title{AlignObj for analytes between a pair for DIA_metabolomics runs.}
\usage{
getMetaboAlignObjs(analytes, runs, dataPath = ".",
  alignType = "hybrid", query = NULL, oswMerged = TRUE,
  nameCutPattern = "(.*)(/)(.*)", maxFdrQuery = 0.05,
  maxFdrLoess = 0.01, spanvalue = 0.1, normalization = "mean",
  simMeasure = "dotProductMasked", SgolayFiltOrd = 2,
  SgolayFiltLen = 3, goFactor = 0.125, geFactor = 400,
  cosAngleThresh = 0.3, OverlapAlignment = TRUE,
  dotProdThresh = 0.96, gapQuantile = 0.98, hardConstrain = FALSE,
  samples4gradient = 100, samplingTime = NULL, RSEdistFactor = 3.5)
}
\value{
A list of AlignObj. Each AlignObj contains alignment path, similarity matrix and related parameters.
}
\description{
AlignObj for analytes between a pair for DIA_metabolomics runs.
}
