% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getMetaboswFiles}
\alias{getMetaboswFiles}
\title{Get list of analytes and their chromatogram indices.}
\usage{
getMetaboswFiles(filenames, dataPath = ".", peptides = NULL,
  query = NULL, oswMerged = FALSE, maxFdrQuery = 1)
}
\value{
A list of data-frames.
}
\description{
Get list of analytes and their chromatogram indices.
}
