% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMultipeptide}
\alias{getMultipeptide}
\title{Get multipeptides}
\usage{
getMultipeptide(precursors, features, applyFun = lapply)
}
\arguments{
\item{precursors}{(data-frames) Contains precursors and associated transition IDs.}

\item{features}{(list of data-frames) Contains features and their properties identified in each run.}
}
\value{
(list) of dataframes having following columns:
\item{transition_group_id}{(integer) a unique id for each precursor.}
\item{run}{(string) run identifier.}
\item{RT}{(numeric) retention time as in FEATURE.EXP_RT of osw files.}
\item{Intensity}{(numeric) peak intensity as in FEATURE_MS2.AREA_INTENSITY of osw files.}
\item{leftWidth}{(numeric) as in FEATURE.LEFT_WIDTH of osw files.}
\item{rightWidth}{(numeric) as in FEATURE.RIGHT_WIDTH of osw files.}
\item{peak_group_rank}{(integer) rank of each feature associated with transition_group_id.}
\item{m_score}{(numeric) q-value of each feature associated with transition_group_id.}
\item{alignment_rank}{(integer) rank of each feature post-alignment.}
}
\description{
Each element of the multipeptide is a collection of features associated with a peptide.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- getRunNames(dataPath, oswMerged = TRUE)
precursors <- getPrecursors(fileInfo, oswMerged = TRUE, context = "experiment-wide")
features <- getFeatures(fileInfo, maxFdrQuery = 0.05)
multipeptide <- getMultipeptide(precursors, features)
multipeptide[["9861"]]
}
\seealso{
\code{\link{getPrecursors}, \link{getFeatures}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-04-08
}
