% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsqMassPntrs.R
\name{getMZandChromHead}
\alias{getMZandChromHead}
\title{Get chromatogram data points and chromatogram indexes per chromatogram file}
\usage{
getMZandChromHead(
  data_table = NULL,
  chromatogram_file_i = NULL,
  run = NULL,
  sql_query = NULL,
  mzPntrs = NULL
)
}
\arguments{
\item{input}{A shiny input variable that contains Working Directory Information}

\item{global}{A list variable containing paths to chromatogram files}
}
\value{
(A list of mzRpwiz)
}
\description{
Get chromatogram data points and chromatogram indexes per chromatogram file
}
