% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzml.R
\name{getMZMLpointers}
\alias{getMZMLpointers}
\title{Get pointers to each mzML file.}
\usage{
getMZMLpointers(fileInfo)
}
\arguments{
\item{fileInfo}{(data-frame) Output of DIAlignR::getRunNames function}
}
\value{
(A list of mzRpwiz)
}
\description{
Returns instantiated mzRpwiz object associated to mzML file.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- getRunNames(dataPath = dataPath)
mzPntrs <- getMZMLpointers(fileInfo)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
