% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_files.R
\name{getNativeIDs}
\alias{getNativeIDs}
\title{Fetch NativeIDs}
\usage{
getNativeIDs(oswIn, peps, params = paramsDIAlignR(), oswMerged = TRUE)
}
\arguments{
\item{oswIn}{(string) path to the osw feature file.}

\item{peps}{(integer) ids of peptides to be aligned. If NULL, align all peptides.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{oswMerged}{(logical) TRUE if merged file from pyprophet is used.}
}
\value{
(integer) a vector of transition IDs.
}
\description{
Get transition(native) IDs for the peptides.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
oswIn <- file.path(dataPath, "osw", "merged.osw")
peps <- c(3L, 11L) # No transitions for 3L.
params <- paramsDIAlignR()
params[["context"]] <- "experiment-wide"
getNativeIDs(oswIn, peps, params) # 106468 106469 106470 106471 106472 106473
}
\seealso{
\code{\link{reduceXICs}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2022) + GPL-3
Date: 2022-04-19
}
