% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_runs.R
\name{getNodeRun}
\alias{getNodeRun}
\title{Create a child run from two parent runs}
\usage{
getNodeRun(
  runA,
  runB,
  mergeName,
  dataPath,
  fileInfo,
  features,
  mzPntrs,
  prec2chromIndex,
  precursors,
  params,
  adaptiveRTs,
  refRuns,
  multipeptide,
  peptideScores,
  ropenms
)
}
\arguments{
\item{runA}{(string) name of a run to be merged with runB. Must be in the rownames of fileInfo.}

\item{runB}{(string) name of a run to be merged with runA. Must be in the rownames of fileInfo.}

\item{mergeName}{(string) name of the node that is generated with merging of runA and runB.}

\item{dataPath}{(string) path to mzml and osw directory.}

\item{fileInfo}{(data-frame) output of \code{\link{getRunNames}}.}

\item{features}{(list of data-frames) contains features and their properties identified in each run.}

\item{mzPntrs}{(list) a list of mzRpwiz.}

\item{prec2chromIndex}{(list) a list of dataframes having following columns: \cr
transition_group_id: it is PRECURSOR.ID from osw file. \cr
chromatogramIndex: index of chromatogram in mzML file.}

\item{precursors}{(data-frame) atleast two columns transition_group_id and transition_ids are required.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{adaptiveRTs}{(environment) an empty environment used to store data for downstream analysis.}

\item{refRuns}{(environment) an empty environment used to store data for downstream analysis.}

\item{ropenms}{(pyopenms module) get this python module through \code{\link{get_ropenms}}.}
}
\value{
(None)
}
\description{
Get merged features and merged chromatograms from parent runs. Chromatograms are written on the disk
at dataPath/mzml. For each precursor aligned parent time-vectors and corresponding child time-vector
are also calculated and written as *_av.rda at dataPath.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
params <- paramsDIAlignR()
fileInfo <- getRunNames(dataPath = dataPath)
mzPntrs <- list2env(getMZMLpointers(fileInfo))
features <- list2env(getFeatures(fileInfo, maxFdrQuery = 1.00, runType = "DIA_proteomics"))
precursors <- getPrecursors(fileInfo, oswMerged = TRUE, runType = params[["runType"]],
 context = "experiment-wide", maxPeptideFdr = params[["maxPeptideFdr"]])
precursors <- dplyr::arrange(precursors, .data$peptide_id, .data$transition_group_id)
peptideIDs <- unique(precursors$peptide_id)
peptideScores <- getPeptideScores(fileInfo, peptideIDs, oswMerged = TRUE, params[["runType"]], params[["context"]])
peptideScores <- lapply(peptideIDs, function(pep) dplyr::filter(peptideScores, .data$peptide_id == pep))
names(peptideScores) <- as.character(peptideIDs)
peptideScores <- list2env(peptideScores)
multipeptide <- getMultipeptide(precursors, features)
prec2chromIndex <- list2env(getChromatogramIndices(fileInfo, precursors, mzPntrs))
mergeName <- "master"
adaptiveRTs <- new.env()
refRuns <- new.env()
\dontrun{
ropenms <- get_ropenms(condaEnv = "envName", useConda=TRUE)
getNodeRun(runA = "run2", runB = "run0", mergeName = mergeName, dataPath = ".", fileInfo, features,
 mzPntrs, prec2chromIndex, precursors, params, adaptiveRTs, refRuns, multipeptide, peptideScores, ropenms)
file.remove(file.path(".", "mzml", paste0(mergeName, ".chrom.mzML")))
}
}
\seealso{
\code{\link{childXICs}, \link{getChildXICs}, \link{traverseUp}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-06
}
