% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osw.R
\name{getOswAnalytes}
\alias{getOswAnalytes}
\title{Get a table of analytes, their chromatogram indices and scores.}
\usage{
getOswAnalytes(dataPath, filenames, oswMerged = TRUE,
  analyteInGroupLabel = FALSE, maxFdrQuery = 0.05,
  runType = "DIA_proteomics")
}
\value{
A data-frame.
}
\description{
Get a table of analytes, their chromatogram indices and scores.
}
