% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osw_mzml.R
\name{getOswFiles}
\alias{getOswFiles}
\title{Get list of peptides and their chromatogram indices.}
\usage{
getOswFiles(dataPath, filenames, maxFdrQuery, analyteFDR, oswMerged,
  peptides = NULL, runType = "DIA_proteomics")
}
\value{
A list of data-frames.
}
\description{
Get list of peptides and their chromatogram indices.
}
