% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pep_peak_corp.R
\name{getPepPeakCorp}
\alias{getPepPeakCorp}
\title{Establishes pairwise correspondence between extracted-ion chromatograms.}
\usage{
getPepPeakCorp(featureTable, pairName, XICs.A, XICs.B, alignType,
  oswFeatureList, expRSE, samplingTime = 3.4, samples4gradient = 100,
  RSEdistFactor = 3.5, hardConstrain = FALSE, gapQuantile = 0.5,
  goFactor = 0.125, geFactor = 40, normalization = "mean",
  simMeasure = "dotProductMasked")
}
\arguments{
\item{featureTable}{A numeric matrix}

\item{pairName}{A numeric value}

\item{XICs.A}{A numeric value}

\item{XICs.B}{A logical scalar}

\item{alignType}{A logical scalar}

\item{oswFeatureList}{A logical scalar}

\item{expRSE}{A logical scalar}

\item{samplingTime}{A logical scalar}

\item{samples4gradient}{A logical scalar}

\item{RSEdistFactor}{A logical scalar}

\item{hardConstrain}{A logical scalar}

\item{gapQuantile}{A logical scalar}

\item{goFactor}{A logical scalar}

\item{geFactor}{A logical scalar}

\item{normalization}{A logical scalar}

\item{simMeasure}{A logical scalar}
}
\value{
A numeric matrix
}
\description{
This function calculates three matrices for affine gap alignment using input
similarity matrix and affine gap opening and gap closing penalties. An
implementation of Needleman-Wunsch alignment and overlap alignment is also
provided. All three matrices are clubbed together in an output S4 object.
}
\examples{
pair <- "run1_run2"
MappedTime2 <- getPepPeakCorp(StrepAnnot, pair, StrepChroms[["run1"]], StrepChroms[["run2"]],
"hybrid", oswOutStrep, expRSE = 7.4)
}
