% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osw_query.R
\name{getPeptideQuery2}
\alias{getPeptideQuery2}
\title{Get peptide scores}
\usage{
getPeptideQuery2(runType = "DIA_Proteomics")
}
\arguments{
\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}
}
\value{
SQL query to be searched.
}
\description{
For each peptide, its score, pvalue and qvalues are fetched across all runs.
}
\seealso{
\code{\link{getPeptideScores}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-11-18
}
\keyword{internal}
