% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osw_mzml.R
\name{getPrecursorIndices}
\alias{getPrecursorIndices}
\title{Get MS1 chromatogram indices of precursors.}
\usage{
getPrecursorIndices(
  fileInfo,
  precursors,
  mzPntrs,
  ions = ".*",
  applyFun = lapply
)
}
\arguments{
\item{fileInfo}{(data-frame) Output of getRunNames function.}

\item{precursors}{(data-frame) Atleast two columns transition_group_id and transition_ids are required.}

\item{mzPntrs}{A list of mzRpwiz.}

\item{ions}{(character) Strings to specify ions. eg. i0, i1.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\value{
(list) A list of dataframes having following columns:
\item{transition_group_id}{(string) it is PRECURSOR.ID from osw file.}
\item{chromatogramIndex}{(integer) index of MS1 chromatogram in mzML file.}
}
\description{
This function reads the header of chromatogram files. It then fetches chromatogram indices by matching transition_group_id(osw) with chromatogramID(xics).
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- getRunNames(dataPath = dataPath)
precursors <- getPrecursors(fileInfo, oswMerged = TRUE, context = "experiment-wide")
mzPntrs <- getMZMLpointers(fileInfo)
prec2chromIndex <- getPrecursorIndices(fileInfo, precursors, mzPntrs)
for(mz in mzPntrs) DBI::dbDisconnect(mz)
}
\seealso{
\code{\link{chromatogramIdAsInteger}, \link{mapPrecursorToChromIndices}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2022) + GPL-3
Date: 2022-01-15
}
