% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osw.R
\name{getPrecursors}
\alias{getPrecursors}
\title{Get precursors from all feature files}
\usage{
getPrecursors(
  fileInfo,
  oswMerged = TRUE,
  runType = "DIA_proteomics",
  context = "global",
  maxPeptideFdr = 0.05
)
}
\arguments{
\item{fileInfo}{(data-frame) Output of \code{\link{getRunNames}} function.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{context}{(string) Context used in pyprophet peptide. Must be either "run-specific", "experiment-wide", or "global".}

\item{maxPeptideFdr}{(numeric) A numeric value between 0 and 1. It is used to filter peptides from osw file which have SCORE_PEPTIDE.QVALUE less than itself.}
}
\value{
(data-frames) A data-frame having following columns:
\item{transition_group_id}{(integer) a unique id for each precursor.}
\item{peptide_id}{(integer) a unique id for each peptide. A peptide can have multiple precursors.}
\item{sequence}{(string) amino-acid sequence of the precursor with possible modifications.}
\item{charge}{(integer) charge on the precursor.}
\item{group_label}{(string) TODO Figure it out.}
\item{transition_ids}{(list) fragment-ion ID associated with transition_group_id. This is matched with chromatogram ID in mzML file.}
}
\description{
Get a data-frame of analytes' transition_group_id, transition_ids, peptide_id and amino-acid sequences.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- getRunNames(dataPath = dataPath)
precursorsInfo <- getPrecursors(fileInfo, oswMerged = TRUE, runType = "DIA_proteomics",
context = "experiment-wide", maxPeptideFdr = 0.05)
dim(precursorsInfo) # 234  6
}
\seealso{
\code{\link{getRunNames}, \link{fetchPrecursorsInfo}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-04-06
}
