% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRefExpFeatureMap}
\alias{getRefExpFeatureMap}
\title{Alignment Feature Mapping Table}
\usage{
getRefExpFeatureMap(precursors, features, applyFun = lapply)
}
\arguments{
\item{precursors}{(data-frames) Contains precursors and associated transition IDs.}

\item{features}{(list of data-frames) Contains features and their properties identified in each run.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\value{
(list) of dataframes per precursor peptide id having following columns:
\item{reference_feature_id}{(integer) id of reference feature.}
\item{experiment_feature_id}{(integer) id of experiment feature aligned to reference feature.}
}
\description{
Construct an alignment feature mapping table for star align method, to map aligned features in experiment runs to the reference run
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- getRunNames(dataPath, oswMerged = TRUE)
precursors <- getPrecursors(fileInfo, oswMerged = TRUE, context = "experiment-wide")
features <- getFeatures(fileInfo, maxFdrQuery = 0.05)
multiFeatureAlignmentMap <- getRefExpFeatureMap(precursors, features)
multiFeatureAlignmentMap[["9861"]]
}
\seealso{
\code{\link{getPrecursors}, \link{getFeatures}}
}
\author{
Justin Sing, \email{justinc.sing@mail.utoronto.ca}

ORCID: 0000-0003-0386-0092

License: (c) Author (2020) + GPL-3
Date: 2022-11-07
}
