% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRefRun}
\alias{getRefRun}
\title{Fetch the reference run-index.}
\usage{
getRefRun(multipeptide)
}
\arguments{
\item{multipeptide}{(list of data-frames) Each element of the is collection of features associated with a precursor.}
}
\value{
(dataframe) has two columns:
\item{transition_group_id}{(integer) a unique id for each precursor.}
\item{run}{(string) run identifier.}
}
\description{
Provides the reference run-index based on lowest m-score.
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
\dontrun{
getRefRun(multipeptide)
}
}
\seealso{
\code{\link{getMultipeptide}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2020-04-08
}
\keyword{internal}
