% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_fit.R
\name{getRSE}
\alias{getRSE}
\title{Calculates Residual Standard Error of the fit}
\usage{
getRSE(fit, globalAlignment)
}
\arguments{
\item{fit}{(lm or loess) Linear or loess fit object between reference and experiment run.}
}
\value{
(numeric)
}
\description{
Calculates Residual Standard Error of the fit
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
\dontrun{
Loess.fit <- getGlobalAlignment(oswFiles = oswFiles_DIAlignR, ref = "run1", eXp = "run2",
maxFdrGlobal = 0.05, spanvalue = 0.1, fit = "loess")
getRSE(Loess.fit)
}
}
\seealso{
\code{\link{getLOESSfit}, \link{getLinearfit}, \link{getGlobalAlignment}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-04-08
}
\keyword{internal}
