% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_fit.R
\name{getRTdf}
\alias{getRTdf}
\title{Calculates global alignment between RT of two runs}
\usage{
getRTdf(oswFiles, ref, eXp, maxFdrGlobal)
}
\arguments{
\item{oswFiles}{(list of data-frames) it is output from getFeatures function.}

\item{ref}{(string) Must be a combination of "run" and an iteger e.g. "run2".}

\item{eXp}{(string) Must be a combination of "run" and an iteger e.g. "run2".}

\item{maxFdrGlobal}{(numeric) A numeric value between 0 and 1. Features should have m-score lower than this value for participation in global fit.}
}
\value{
A data-frame
}
\description{
This function selects features from oswFiles which has m-score < maxFdrLoess. It fits linear/loess regression on these feature.
Retention-time mapping is established from reference to experiment run.
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
df <- getRTdf(oswFiles = oswFiles_DIAlignR, ref = "run1", eXp = "run2", maxFdrGlobal = 0.05)
}
\seealso{
\code{\link{getGlobalAlignment}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2019-03-01
}
