% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getRunNames}
\alias{getRunNames}
\title{Get names of all runs}
\usage{
getRunNames(dataPath, oswMerged = TRUE, nameCutPattern = "(.*)(/)(.*)")
}
\value{
A vector with names of all runs.
}
\description{
Get names of all runs
}
