% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getseqSimMat}
\alias{getseqSimMat}
\title{Calculate similarity matrix for two sequences}
\usage{
getseqSimMat(seq1, seq2, Match, MisMatch, s)
}
\arguments{
\item{seq1}{(char) A single string}

\item{seq2}{(char) A single string}

\item{Match}{(float) Score for character match}

\item{MisMatch}{(float) score for character mismatch}

\item{s}{(matrix) A similarity matrix}
}
\value{
s (matrix) Updated similarity matrix
}
\description{
Calculate similarity matrix for two sequences
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-05
}
