% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSeqSimMatCpp}
\alias{getSeqSimMatCpp}
\title{Calculates similarity matrix for two sequences}
\usage{
getSeqSimMatCpp(seq1, seq2, match, misMatch)
}
\arguments{
\item{seq1}{(char) A single string.}

\item{seq2}{(char) A single string.}

\item{match}{(double) Score for character match.}

\item{misMatch}{(double) score for character mismatch.}
}
\value{
s (matrix) Numeric similarity matrix. Rows and columns expresses seq1 and seq2, respectively.
}
\description{
Calculates similarity matrix for two sequences
}
\examples{
# Get sequence similarity of two DNA strings
Match=10; MisMatch=-2
seq1 = "GCAT"; seq2 = "CAGTG"
getSeqSimMatCpp(seq1, seq2, Match, MisMatch)
matrix(c(-2, 10, -2, -2, -2, -2, 10, -2, 10, -2, -2, -2, -2, -2, -2, 10, 10, -2, -2, -2), 4, 5, byrow = F)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-05
}
