% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similarity_matrix.R
\name{getSimilarityMatrix}
\alias{getSimilarityMatrix}
\title{calculates similarity matrix between two chromatogram groups}
\usage{
getSimilarityMatrix(data, pep, runA, runB, removeOutlierChrom = FALSE,
  type = "dotProduct")
}
\arguments{
\item{data}{Chromatographic data}

\item{pep}{precursor ID}

\item{runA}{First run of the run-pair}

\item{runB}{Second run of the run-pair}

\item{type}{A character string}
}
\description{
This function takes in datafile, precursor id, names of run pair. Based on
similarity measure, it calculates the similarity matrix.
}
