% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_order.R
\name{getTree}
\alias{getTree}
\title{Create a phylogenetic tree}
\usage{
getTree(distMat)
}
\arguments{
\item{distMat}{(dist) pairwise distance matrix.}
}
\value{
(phylo) a tree.
}
\description{
Create a phylogenetic tree
}
\examples{
m <- matrix(c(0,1,2,3, 1,0,1.5,1.5, 2,1.5,0,1, 3,1.5,1,0), byrow = TRUE,
            ncol = 4, dimnames = list(c("run1", "run2", "run3", "run4"),
                                      c("run1", "run2", "run3", "run4")))
distMat <- as.dist(m, diag = FALSE, upper = FALSE)
labels(distMat); length(distMat)
\dontrun{
getTree(distMat)
}
}
\seealso{
\code{\link[phangorn]{upgma}, \link{getNodeIDs}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-31
}
\keyword{internal}
