% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getXICs}
\alias{getXICs}
\title{Get XICs for a list of peptides}
\usage{
getXICs(analytes, runs, dataPath = ".", maxFdrQuery = 1,
  SgolayFiltOrd = 4, SgolayFiltLen = 9, runType = "DIA_proteomics",
  oswMerged = TRUE, nameCutPattern = "(.*)(/)(.*)",
  analyteInGroupLabel = FALSE)
}
\value{
A list of list. Each list contains XICs for that run.
}
\description{
Get XICs for a list of peptides
}
