% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getXICs4AlignObj}
\alias{getXICs4AlignObj}
\title{Extract XICs of all analytes from oswFiles}
\usage{
getXICs4AlignObj(dataPath, runs, oswFiles, analytes,
  XICfilter = "sgolay", SgolayFiltOrd = 4, SgolayFiltLen = 9)
}
\arguments{
\item{dataPath}{(char) path to mzml and osw directory.}

\item{runs}{(vector of string) names of mzML files without extension. Names of the vector must be a combination of "run" and an iteger e.g. "run2".}

\item{oswFiles}{(list of data-frames) it is output from getOswFiles function.}

\item{analytes}{(string) analyte is as PRECURSOR.GROUP_LABEL or as PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}

\item{XICfilter}{(string) this must be one of the strings "sgolay", "none".}

\item{SgolayFiltOrd}{(integer) it defines the polynomial order of filer.}

\item{SgolayFiltLen}{(integer) must be an odd number. It defines the length of filter.}
}
\value{
A list of list of data-frames. Each data frame has elution time and intensity of fragment-ion XIC.
}
\description{
For all the analytes requested, it fetches chromatogram indices from oswFiles and
extract chromatograms from mzML files.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
filenames <- DIAlignR::getRunNames(dataPath = dataPath)
runs <- c("run1" = "hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt",
          "run0" =  "hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt")
oswFiles <- DIAlignR::getOswFiles(dataPath, filenames)
analytes <- "QFNNTDIVLLEDFQK_3"
XICs <- getXICs4AlignObj(dataPath, runs, oswFiles, analytes)
}
\seealso{
\code{\link{getOswFiles}, \link{getRunNames}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
