% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeMatrix}
\alias{initializeMatrix}
\title{Outputs a NumericMatrix of given row and column size.}
\usage{
initializeMatrix(initVal, ROW_SIZE, COL_SIZE)
}
\arguments{
\item{initVal}{(char) Matrix intialization value}

\item{ROW_SIZE}{(int) Number of rows}

\item{COL_SIZE}{(int) Number of columns}
}
\value{
s (NumericMatrix) A matrix
}
\description{
Outputs a NumericMatrix of given row and column size.
}
\examples{
# Get a matrix of type NumericMatrix
initializeMatrix(0, ROW_SIZE = 4, COL_SIZE = 5)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-05
}
