% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ipfReassignFDR}
\alias{ipfReassignFDR}
\title{Re-Assign FDR Aligned Peaks}
\usage{
ipfReassignFDR(dt, refRuns, fileInfo, params)
}
\arguments{
\item{dt}{(data.table) Aligned results tables from writeTables}

\item{refRuns}{(dataframe) A peptide reference run table from getRefRun}

\item{fileInfo}{(dataframe) A table with run filename information from getRunNames}

\item{params}{(list) A list of parameters used by DIAlignR generated from paramsDIAlignR}
}
\value{
(data.table) Aligned results table with reassigned m-scores
}
\description{
This method will re-assign peaks with either the MS2 FDR if there was not enough confidence (identifying transitions) in the IPF FDR, or it will assign the user defined alignedFDR. This assumes the reference run has a really good peak that meets the IPF FDR threshold.
}
\examples{
\dontrun{
finalTbl <- ipfReassignFDR(finalTbl, refRuns, fileInfo, params)
}
}
\seealso{
writeTables, getRefRun, getRunNames, paramsDIAlignR
}
\author{
Justin Sing, \email{justinc.sing@mail.utoronto.ca}

ORCID: 0000-0003-0386-0092

License: (c) Author (2020) + GPL-3
Date: 2020-07-17
}
\keyword{internal}
