% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplit_alignment.R
\name{iRTAlignment}
\alias{iRTAlignment}
\title{Builds assay library and pseudo iRT library}
\usage{
iRTAlignment(
  masterLibPath,
  iRTLibPath,
  msplitAssaySuffix,
  lib_lib_align = "linear",
  spanvalue = 0.1
)
}
\arguments{
\item{masterLibPath}{(numeric) path of the master library to be converted to sample-specific assay library.}

\item{iRTLibPath}{(string) pattern of pseudo iRT spectra files. Pattern must start with '*'.}

\item{msplitAssaySuffix}{(numeric) suffix of MSPLIT assay library.}

\item{lib_lib_align}{(integer) start time of the run in second. Usually it is 0.}

\item{spanvalue}{(integer) stop time of the run in seconds. Usually it is 7200 sec (2 hour run).}
}
\value{
(None) writes _norm_iRT_assaylib.tsv and _norm_masterLib.tsv files.
}
\description{
Builds assay library and pseudo iRT library
}
\examples{
masterLibPath <- "masterLib.tsv"
iRTLibPath <- "iRTassays.tsv"
msplitAssaySuffix <- "_Openswath_assaylib.tsv"
\dontrun{
iRTAlignment(masterLibPath, iRTLibPath, msplitAssaySuffix, lib_lib_align = "loess")
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-25
}
\keyword{internal}
