% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isListObj}
\alias{isListObj}
\title{Find if an object exists in a list
Searches name structure for name search parameter}
\usage{
isListObj(x, name)
}
\arguments{
\item{x}{A list object}

\item{name}{Name of sub-components in list}
}
\value{
logical value if name is present in list
}
\description{
Find if an object exists in a list
Searches name structure for name search parameter
}
