% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_alignment.R
\name{mapIdxToTime}
\alias{mapIdxToTime}
\title{Establishes mapping back from index to time}
\usage{
mapIdxToTime(timeVec, idx)
}
\arguments{
\item{timeVec}{A numeric vector}

\item{idx}{An integer vector}
}
\value{
A mutated time vector
}
\description{
Takes a time vector and index vector of same length. This function create a
new time vector given indices specified in \code{idx}. For \code{NA} indices
it uses last index to fill time value. For \code{NA} appearing at the start
of \code{idx}, it uses next index to get time value.
}
