% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_alignment.R
\name{mappedRTfromAlignObj}
\alias{mappedRTfromAlignObj}
\title{Use AlignObj to map reference retention time on experiment run.}
\usage{
mappedRTfromAlignObj(refRT, tVec.ref, tVec.eXp, AlignObj)
}
\value{
A numeric
}
\description{
Use AlignObj to map reference retention time on experiment run.
}
