% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osw_mzml.R
\name{mapPrecursorToChromIndices}
\alias{mapPrecursorToChromIndices}
\title{Merge precursor and transitions mapping with chromatogram header}
\usage{
mapPrecursorToChromIndices(prec2transition, chromHead)
}
\arguments{
\item{prec2transition}{(dataframe) This has two columns: transition_group_id and transition_ids with integer values.}

\item{chromHead}{(dataframe) This has two columns: chromatogramId and chromatogramIndex with integer values.}
}
\value{
(dataframe) having following columns:
\item{transition_group_id}{(string) it is either fetched from PRECURSOR.GROUP_LABEL or a combination of PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}
\item{chromatogramIndex}{(integer) Index of chromatogram in mzML file.}
}
\description{
Merges dataframes on transition_ids(OSW) = chromatogramId(mzML).
}
\seealso{
\code{\link{getChromatogramIndices}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2020-04-07
}
\keyword{internal}
