% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{masterXICs_DIAlignR}
\alias{masterXICs_DIAlignR}
\title{Master fragment-ion chromatograms from two parents}
\format{
The format is similar to the output of \code{\link{childXICs}}. A list of two elements: First element contains six fragmentation chromatograms.
The second element has aligned parent time-vectors and corresponding child time-vector. It has
five columns:
\describe{
\item{indexAligned.ref}{(integer) aligned indices of reference run.}
\item{indexAligned.eXp}{(integer) aligned indices of experiment run.}
\item{tAligned.ref}{(numeric) aligned time-vector of reference run.}
\item{tAligned.eXp}{(numeric) aligned time-vector of experiment run.}
\item{alignedChildTime}{(numeric) aligned time-vector of master run.}
}
}
\source{
File test_GenerateData.R has \href{https://github.com/shubham1637/DIAlignR/tree/master/data-raw}{source code}
to generate the example data.
}
\usage{
masterXICs_DIAlignR
}
\description{
Created merged XICs of peptide (ID = 4618) 14299_QFNNTDIVLLEDFQK/3 from two SWATH runs: \cr
run1 : hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML\cr
run2 : hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML
}
\keyword{datasets}
