% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_paper.R
\name{mcp_2019}
\alias{mcp_2019}
\title{DIAlignR paper results}
\usage{
mcp_2019(
  annotationFile,
  peptidesFile,
  dataPath = ".",
  oswMerged = TRUE,
  globalAlignment = "loess",
  globalAlignmentFdr = 0.01,
  globalAlignmentSpan = 0.1,
  XICfilter = "sgolay",
  kernelLen = 9,
  polyOrd = 3,
  RSEdistFactor = 3.5,
  alignType = "hybrid",
  normalization = "mean",
  simMeasure = "dotProductMasked",
  goFactor = 0.125,
  geFactor = 40,
  cosAngleThresh = 0.3,
  OverlapAlignment = TRUE,
  dotProdThresh = 0.96,
  gapQuantile = 0.65,
  kerLen = 9L,
  hardConstrain = FALSE,
  samples4gradient = 100
)
}
\description{
DIAlignR paper results
}
\examples{
\dontrun{
peptidesFile <- "data/peptides.txt"
annotationFile <- "data/SkylineResult500Peptides.csv"
DIAlignR:::mcp_2019(annotationFile, peptidesFile)
}
}
\keyword{internal}
