% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{mergeOsw_ChromHeader}
\alias{mergeOsw_ChromHeader}
\title{Merge dataframes from OSW and mzML files.}
\usage{
mergeOsw_ChromHeader(chromHead, analytesInfo, peptides, runType)
}
\value{
Update list of data-frame
}
\description{
Merge dataframes from OSW and mzML files.
}
