% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osw_mzml.R
\name{mergeOswAnalytes_ChromHeader}
\alias{mergeOswAnalytes_ChromHeader}
\title{Merge dataframes from OSW and mzML files.}
\usage{
mergeOswAnalytes_ChromHeader(oswAnalytes, chromHead, analyteFDR,
  runType = "DIA_proteomics")
}
\value{
Update list of data-frame
}
\description{
Merge dataframes from OSW and mzML files.
}
