% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_chromatogram.R
\name{mergeXIC}
\alias{mergeXIC}
\title{Merge two XICs into one}
\usage{
mergeXIC(XIC.ref, XIC.eXp)
}
\arguments{
\item{XICs.ref}{(data-frame) extracted ion chromatograms from reference run.}

\item{XICs.eXp}{(data-frame) extracted ion chromatograms from experiment run.}
}
\value{
(data-frame) a chromatogram.
}
\description{
Both extracted-ion chromatograms must have same number of rows.
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run1"]][["14299_QFNNTDIVLLEDFQK/3"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run2"]][["14299_QFNNTDIVLLEDFQK/3"]]
\dontrun{
mergeXIC(XICs.ref[[1]], XICs.eXp[[1]])
}
}
\seealso{
\code{\link{alignChromatogramsCpp}, \link{getAlignObj}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-23
}
\keyword{internal}
