% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplit_alignment.R
\name{msplit_masterLib}
\alias{msplit_masterLib}
\title{Get master library from MSPLIT}
\usage{
msplit_masterLib(msplitFilteredPattern, msplitAssayPattern, spanvalue = 0.1)
}
\arguments{
\item{msplitFilteredPattern}{(string) pattern of MSPLITfiltered.txt files. Pattern must start with '*'.}

\item{msplitAssayPattern}{(string) pattern of MSPLITassay_library.tsv files. Pattern must start with '*'.}

\item{spanvalue}{(numeric) spanvalue for LOESS fit. For targeted proteomics 0.1 could be used.}
}
\value{
(None) writes a masterLib.tsv file.
}
\description{
MSPLIT returns an OpenSWATH assay libraries which can be combined to obtain a comprehensive library
This library allows to perform false-discovery rate extimation on all samples together and further use
alignment algorithm of DIAlignR.
}
\examples{
msplitFilteredPattern <- "*_MSPLITfiltered.txt"
msplitAssayPattern <- "*_Openswath_assaylib.tsv"
\dontrun{
msplit_masterLib(msplitFilteredPattern, msplitAssayPattern)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-14
}
