% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplit_alignment.R
\name{msplit_to_openms}
\alias{msplit_to_openms}
\title{Converts to OpenMS format}
\usage{
msplit_to_openms(filename)
}
\arguments{
\item{filename}{(string) Path to the MSPLIT output library. The file extension must be .txt.}
}
\value{
(None) writes a .tsv file.
}
\description{
MSPLIT returns an OpenSWATH assay library which is not compatible for OpenMS version 3.0
This function reformat the library for TargetedFileConverter (OpenMS).
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-14
}
