% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstAlign.R
\name{mstAlignRuns}
\alias{mstAlignRuns}
\title{Peptide quantification through MST alignment}
\usage{
mstAlignRuns(
  dataPath,
  params,
  outFile = "DIAlignR",
  oswMerged = TRUE,
  scoreFile = NULL,
  runs = NULL,
  peps = NULL,
  mstNet = NULL,
  applyFun = lapply
)
}
\arguments{
\item{dataPath}{(string) path to xics and osw directory.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{outFile}{(string) name of the output file.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{scoreFile}{(string) path to the peptide score file, needed when oswMerged is FALSE.}

\item{runs}{(string) names of xics file without extension.}

\item{peps}{(integer) ids of peptides to be aligned. If NULL, align all peptides.}

\item{mstNet}{(matrix) array of tree-edges. Look up \code{\link{getMST}}.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\value{
(None)
}
\description{
This function expects osw and xics directories at dataPath. It first reads osw files and fetches
 chromatogram indices for each analyte. To perform alignment, first a guide-tree is built using \code{\link[ape]{mst}} which
 can also be provided with mstNet parameter. As we traverse from the start node to the other nodes,
 runs are aligned pairwise.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
params <- paramsDIAlignR()
params[["context"]] <- "experiment-wide"
mstAlignRuns(dataPath, params = params, outFile = "DIAlignR")
\dontrun{
y <- strsplit("run0 run1\nrun2 run2", split = '\n')[[1]]
y <- cbind(A = strsplit(y[1], " ")[[1]], B = strsplit(y[2], " ")[[1]])
plot(igraph::graph_from_edgelist(y, directed = FALSE))
}
}
\seealso{
\code{\link{alignTargetedRuns}, \link{progAlignRuns}, \link{getMST}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-05-15
}
