% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstAlign.R
\name{MSTperBatch}
\alias{MSTperBatch}
\title{Aligns an analyte for a batch of peptides}
\usage{
MSTperBatch(
  iBatch,
  nets,
  peptides,
  multipeptide,
  refRuns,
  precursors,
  prec2chromIndex,
  fileInfo,
  mzPntrs,
  params,
  globalFits,
  RSE,
  applyFun = lapply
)
}
\arguments{
\item{nets}{(list) set of trees ordered for traversal obained from \code{\link{traverseMST}}.}

\item{peptides}{(integer) vector of peptide IDs.}

\item{multipeptide}{(list) contains multiple data-frames that are collection of features
associated with analytes. This is an output of \code{\link{getMultipeptide}}.}

\item{refRuns}{(data-frame) output of \code{\link{getRefRun}}. Must have two columsn : transition_group_id and run.}

\item{precursors}{(data-frame) atleast two columns transition_group_id and transition_ids are required.}

\item{prec2chromIndex}{(list) a list of dataframes having following columns: \cr
transition_group_id: it is PRECURSOR.ID from osw file. \cr
chromatogramIndex: index of chromatogram in mzML file.}

\item{fileInfo}{(data-frame) output of \code{\link{getRunNames}}.}

\item{mzPntrs}{(list) a list of mzRpwiz.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{globalFits}{(list) each element is either of class lm or loess. This is an output of \code{\link{getGlobalFits}}.}

\item{RSE}{(list) Each element represents Residual Standard Error of corresponding fit in globalFits.}
}
\value{
invisible NULL
}
\description{
For the ith analyte in the batch, this function traverse the MST from start node and align runs
pairwise. In the process it updates multipeptide with aligned features.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
}
\seealso{
\code{\link{mstAlignRuns}, \link{alignToRefMST}, \link{getAlignedTimesFast}, \link{getMultipeptide}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-05-15
}
\keyword{internal}
