% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{mstScript1}
\alias{mstScript1}
\title{Extract features and generate minimum spanning tree.}
\usage{
mstScript1(
  dataPath,
  outFile = "DIAlignR",
  params = paramsDIAlignR(),
  oswMerged = TRUE,
  runs = NULL,
  mstNet = NULL,
  applyFun = lapply
)
}
\arguments{
\item{dataPath}{(string) path to xics and osw directory.}

\item{outFile}{(string) name of the output file.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{oswMerged}{(logical) TRUE if merged file from pyprophet is used.}

\item{runs}{(string) names of xics file without extension.}

\item{mstNet}{(string) minimum spanning tree in string format. See example of \link{mstScript2}.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\description{
Extract features and generate minimum spanning tree.
}
\examples{
params <- paramsDIAlignR()
params[["context"]] <- "experiment-wide"
dataPath <- system.file("extdata", package = "DIAlignR")
BiocParallel::register(BiocParallel::MulticoreParam(workers = 4, progressbar = TRUE))
mstScript1(dataPath, outFile = "testDIAlignR", params = params, applyFun = BiocParallel::bplapply)
file.remove(file.path(dataPath, "testDIAlignR_mst1.RData"))
}
\seealso{
\code{\link{mstAlignRuns}, \link{mstScript2}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2022) + GPL-3
Date: 2022-04-19
}
