% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{mstScript2}
\alias{mstScript2}
\title{Performs alignment using mstScript1 output}
\usage{
mstScript2(
  dataPath,
  outFile = "DIAlignR",
  params = paramsDIAlignR(),
  oswMerged = TRUE,
  scoreFile = NULL,
  peps = NULL,
  mstNet = NULL,
  applyFun = lapply
)
}
\arguments{
\item{dataPath}{(string) path to xics and osw directory.}

\item{outFile}{(string) name of the output file.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{oswMerged}{(logical) TRUE if merged file from pyprophet is used.}

\item{scoreFile}{(string) path to the peptide score file, needed when oswMerged is FALSE.}

\item{peps}{(integer) ids of peptides to be aligned. If NULL, align all peptides.}

\item{mstNet}{(string) minimum spanning tree in string format. See example of \link{mstScript2}.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\description{
Performs alignment using mstScript1 output
}
\examples{
params <- paramsDIAlignR()
params[["context"]] <- "experiment-wide"
dataPath <- system.file("extdata", package = "DIAlignR")
BiocParallel::register(BiocParallel::MulticoreParam(workers = 4, progressbar = TRUE))
mstScript1(dataPath, outFile = "testDIAlignR", params = params, applyFun = BiocParallel::bplapply)
mstNet <- "run0 run0\nrun1 run2"
mstScript2(dataPath, outFile = "testDIAlignR", params = params, mstNet = mstNet, applyFun = lapply)
file.remove(file.path(dataPath, "testDIAlignR_mst1.RData"))
file.remove("testDIAlignR.tsv")
}
\seealso{
\code{\link{mstAlignRuns}, \link{mstScript1}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2022) + GPL-3
Date: 2022-04-19
}
