% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_aligned_obj.R
\docType{class}
\name{NeedleObjAffine-class}
\alias{NeedleObjAffine-class}
\alias{NeedleObjAffine}
\title{An S4 object to group matrices produced by the alignmnet}
\arguments{
\item{Traceback}{A list of}

\item{M}{A numeric matrix}

\item{A}{A numeric matrix}

\item{B}{A numeric matrix}

\item{GapOpen}{A numeric scalar value}

\item{GapExten}{A numeric scalar value}

\item{FreeEndGaps}{A logical value}
}
\description{
This class combines M, A, B matrices and Traceback list into one object.
}
