% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_order.R
\name{newFunction}
\alias{newFunction}
\title{Create a child run from two parent runs}
\usage{
newFunction(
  dataPath,
  params,
  outFile = "DIAlignR.tsv",
  ropenms,
  oswMerged = TRUE,
  runs = NULL,
  newickTree = NULL
)
}
\arguments{
\item{ropenms}{(pyopenms module) get this python module through get_ropenms().}

\item{filename}{(string) Name of the mzML file to be written.}

\item{XICs}{(list of data-frames) extracted ion chromatograms.}

\item{transitionIDs}{(integer) length must be the same as of XICs.}
}
\value{
(None)
}
\description{
Get merged features and merged chromatograms from parent runs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
params <- paramsDIAlignR()
ropenms <- get_ropenms(condaEnv = "TricEnvr")
\dontrun{
newFunction(dataPath, params = params, outFile = "x1.tsv", ropenms = ropenms)
# Removing aligned vectors
file.remove(list.files(dataPath, pattern = "*_av.rds", full.names = TRUE))
# Removing temporarily created master chromatograms
file.remove(list.files(file.path(dataPath, "mzml"), pattern = "^master[0-9]+\\\\.chrom\\\\.mzML$", full.names = TRUE))
}
}
\seealso{
\code{\link{childXICs}, \link{addXIC}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-10
}
