% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_order.R
\name{nrDesc}
\alias{nrDesc}
\title{Number of descendants}
\usage{
nrDesc(tree)
}
\arguments{
\item{tree}{(phylo) a phylogenetic tree.}
}
\value{
(numeric)
}
\description{
Get the number of descendants for each node in the tree.
}
\examples{
m <- matrix(c(0,1,2,3, 1,0,1.5,1.5, 2,1.5,0,1, 3,1.5,1,0), byrow = TRUE,
            ncol = 4, dimnames = list(c("run1", "run2", "run3", "run4"),
                                      c("run1", "run2", "run3", "run4")))
distMat <- as.dist(m, diag = FALSE, upper = FALSE)
\dontrun{
tree <- getTree(distMat)
getNodeIDs(tree)
nrDesc(tree)
}
}
\seealso{
\code{\link[ape]{Ntip}, \link{getNodeIDs}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
}
\keyword{internal}
