% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{otherChildXICpp}
\alias{otherChildXICpp}
\title{Get child chromatogram for other precursors using main precursor alignment}
\usage{
otherChildXICpp(
  l1,
  l2,
  kernelLen,
  polyOrd,
  mat,
  childTime,
  wRef = 0.5,
  splineMethod = "natural"
)
}
\arguments{
\item{l1}{(list) A list of numeric matrix of two columns. l1 and l2 should have same length.}

\item{l2}{(list) A list of numeric matrix of two columns. l1 and l2 should have same length.}

\item{kernelLen}{(integer) length of filter. Must be an odd number.}

\item{polyOrd}{(integer) TRUE: remove background from peak signal using estimated noise levels.}

\item{mat}{(matrix) aligned time and child time from the main precursor.}

\item{childTime}{(numeric) iime vector from the child chromatogram.}

\item{wRef}{(numeric) Weight of the reference XIC. Must be between 0 and 1.}

\item{splineMethod}{(string) must be either "fmm" or "natural".}
}
\value{
(List) of chromatograms.
}
\description{
Get child chromatogram for other precursors using main precursor alignment
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
XICs <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR
XICs.ref <- lapply(XICs[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]], as.matrix)
XICs.eXp <- lapply(XICs[["hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]], as.matrix)
B1p <- 4964.752
B2p <- 5565.462
chrom <- getChildXICpp(XICs.ref, XICs.eXp, 11L, 4L, alignType = "hybrid", adaptiveRT = 77.82315,
 normalization = "mean", simType = "dotProductMasked", B1p = B1p, B2p = B2p,
 goFactor = 0.125, geFactor = 40, cosAngleThresh = 0.3, OverlapAlignment = TRUE,
 dotProdThresh = 0.96, gapQuantile = 0.5, hardConstrain = FALSE, samples4gradient = 100,
 wRef = 0.5, keepFlanks= TRUE)
chrom2 <- otherChildXICpp(XICs.ref, XICs.eXp, 11L, 4L, chrom[[2]], chrom[[1]][[1]][,1],
0.5, "natural")
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2021) + MIT
Date: 2021-01-08
}
