% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_alignment.R
\name{pickNearestFeature}
\alias{pickNearestFeature}
\title{Pick feature closest to reference peak}
\usage{
pickNearestFeature(eXpRT, analyte, oswFiles, runname, adaptiveRT,
  featureFDR)
}
\value{
Invisible NULL
}
\description{
Pick feature closest to reference peak
}
