% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotAlignedAnalytes}
\alias{plotAlignedAnalytes}
\title{Plot aligned XICs group for a specific peptide.
AlignObjOutput is the output from getAlignObjs fucntion.}
\usage{
plotAlignedAnalytes(AlignObjOutput, plotType = "All",
  DrawAlignR = FALSE, annotatePeak = FALSE, saveFigs = FALSE)
}
\description{
Plot aligned XICs group for a specific peptide.
AlignObjOutput is the output from getAlignObjs fucntion.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt", "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt")
AlignObjOutput <- getAlignObjs(analytes = "QFNNTDIVLLEDFQK_3", runs, dataPath = dataPath)
plotAlignedAnalytes(AlignObjOutput)
}
