% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotAlignemntPath}
\alias{plotAlignemntPath}
\title{Plot aligned path through the similarity matrix.
Reference run has indices on X-axis, eXp run has them on Y-axis.}
\usage{
plotAlignemntPath(AlignObjOutput)
}
\description{
Plot aligned path through the similarity matrix.
Reference run has indices on X-axis, eXp run has them on Y-axis.
}
