% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotAlignmentPath}
\alias{plotAlignmentPath}
\title{Visualize alignment path through similarity matrix}
\usage{
plotAlignmentPath(AlignObjOutput)
}
\arguments{
\item{AlignObjOutput}{(list) The list contains AlignObj, raw XICs for reference and experiment, and reference-peak label.}
}
\value{
A plot to the current device.
}
\description{
Plot aligned path through the similarity matrix. Reference run has indices on X-axis, eXp run has them on Y-axis.
In getAlignObjs function, objType must be set to medium.
}
\examples{
library(lattice)
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt",
 "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt")
AlignObjOutput <- getAlignObjs(analytes = "QFNNTDIVLLEDFQK_3", runs, dataPath = dataPath,
 objType = "medium")
plotAlignmentPath(AlignObjOutput)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
