% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotAnalyteXICs}
\alias{plotAnalyteXICs}
\title{Plot extracted-ion chromatogram.}
\usage{
plotAnalyteXICs(analyte, run, dataPath = ".", maxFdrQuery = 1,
  XICfilter = "sgolay", SgolayFiltOrd = 4, SgolayFiltLen = 9,
  runType = "DIA_proteomics", oswMerged = TRUE,
  nameCutPattern = "(.*)(/)(.*)", analyteInGroupLabel = FALSE,
  peakAnnot = NULL, Title = NULL)
}
\arguments{
\item{analyte}{(string) An analyte is as PRECURSOR.GROUP_LABEL or as PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}

\item{run}{(string) Name of a mzml file without extension.}

\item{dataPath}{(char) Path to mzml and osw directory.}

\item{maxFdrQuery}{(numeric) A numeric value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{XICfilter}{(string) This must be one of the strings "sgolay", "none".}

\item{SgolayFiltOrd}{(integer) It defines the polynomial order of filer.}

\item{SgolayFiltLen}{(integer) Must be an odd number. It defines the length of filter.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{nameCutPattern}{(string) regex expression to fetch mzML file name from RUN.FILENAME columns of osw files.}

\item{analyteInGroupLabel}{(logical) TRUE for getting analytes as PRECURSOR.GROUP_LABEL from osw file.}

\item{peakAnnot}{(numeric) Peak-apex time.}

\item{Title}{(logical) TRUE: name of the list will be displayed as title.}
}
\value{
A plot to the current device.
}
\description{
Plot extracted-ion chromatogram.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
run <- "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"
plotAnalyteXICs(analyte = "QFNNTDIVLLEDFQK_3", run, dataPath = dataPath, XICfilter = "none")
plotAnalyteXICs(analyte = "14299_QFNNTDIVLLEDFQK/3", run, dataPath = dataPath,
XICfilter = "sgolay", analyteInGroupLabel = TRUE)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
