% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotPeptideXICs}
\alias{plotPeptideXICs}
\title{Plot Extracted-ion chromatogram group for a specific peptide.}
\usage{
plotPeptideXICs(peptides, runs, dataPath = ".", maxFdrQuery = 1,
  SgolayFiltOrd = 4, SgolayFiltLen = 9, query = NULL,
  oswMerged = TRUE, nameCutPattern = "(.*)(/)(.*)", peakAnnot = NULL,
  Title = NULL)
}
\description{
Plot Extracted-ion chromatogram group for a specific peptide.
}
