% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotSingleAlignedChrom}
\alias{plotSingleAlignedChrom}
\title{Plot an aligned XIC-group.}
\usage{
plotSingleAlignedChrom(XIC_group, idx, peakAnnot = NULL)
}
\description{
Plot an aligned XIC-group.
}
