% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_alignment.R
\name{populateReferenceExperimentFeatureAlignmentMap}
\alias{populateReferenceExperimentFeatureAlignmentMap}
\title{Populate Alignment Feature Mapping Table}
\usage{
populateReferenceExperimentFeatureAlignmentMap(
  df,
  feature_alignment_mapping,
  tAligned,
  ref,
  eXp,
  analyte_chr
)
}
\arguments{
\item{df}{(dataframe) a collection of features related to the peptide}

\item{feature_alignment_mapping}{(data.table)  contains experiment feature ids
mapped to corresponding reference feature id per analyte. This is an output of \code{\link{getRefExpFeatureMap}}.}

\item{tAligned}{(matrix) the first column corresponds to the aligned reference time, the second column is the aligned experiment time. #' @seealso \code{\link{getAlignedTimesFast}}}

\item{ref}{(string) name of the reference run. Must be in the rownames of fileInfo.}

\item{eXp}{(string) name of the run to be aligned to reference run. Must be in the rownames of fileInfo.}

\item{analyte_chr}{(string) name of highest quality precursor/transition_group_id}
}
\description{
Populate an alignment feature mapping table for star align method,
to map aligned features in experiment runs to the reference run for a given analyte.
Is a data.table generated from of \code{\link{getRefExpFeatureMap}}.
}
\author{
Justin Sing, \email{justinc.sing@mail.utoronto.ca}

ORCID: 0000-0003-0386-0092

License: (c) Author (2020) + GPL-3
Date: 2022-11-07
}
