% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressive_alignment.R
\name{progAlignRuns}
\alias{progAlignRuns}
\title{Peptide quantification through progressive alignment}
\usage{
progAlignRuns(
  dataPath,
  params,
  outFile = "DIAlignR",
  ropenms = NULL,
  oswMerged = TRUE,
  runs = NULL,
  newickTree = NULL,
  applyFun = lapply
)
}
\arguments{
\item{dataPath}{(string) path to mzml and osw directory.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{outFile}{(string) name of the output file.}

\item{ropenms}{(pyopenms module) get this python module through \code{\link{get_ropenms}}. Required only for chrom.mzML files.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{runs}{(string) names of mzml file without extension.}

\item{newickTree}{(string) guidance tree in newick format. Look up \code{\link{getTree}}.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\value{
(None)
}
\description{
This function expects osw and mzml directories at dataPath. It first reads osw files and fetches
 chromatogram indices for each analyte. To perform alignment, first a crude guide-tree is built which
 can also be provided with newickTree parameter. As we traverse from the leaf-nodes to the root node,
 runs are aligned pairwise. The root node is named master1 that has average of all fragment ion chromatograms
 and identified peak-groups. These features are propagated back to leaf nodes and finally aligned
 features are written in the output file.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
params <- paramsDIAlignR()
params[["context"]] <- "experiment-wide"
\dontrun{
ropenms <- get_ropenms(condaEnv = "envName")
progAlignRuns(dataPath, params = params, outFile = "test3", ropenms = ropenms)
# Removing aligned vectors
file.remove(list.files(dataPath, pattern = "*_av.rds", full.names = TRUE))
# Removing temporarily created master chromatograms
file.remove(list.files(file.path(dataPath, "mzml"), pattern = "^master[0-9]+\\\\.chrom\\\\.mzML$", full.names = TRUE))
file.remove(file.path(dataPath, "test3.temp.RData"))
file.remove(file.path(dataPath, "master.merged.osw"))
}
}
\seealso{
\code{\link{alignTargetedRuns}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-10
}
