% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nature_paper.R
\name{progAlignRuns2}
\alias{progAlignRuns2}
\title{ids <- as.integer(scan(file = "data/ids.txt"))
params <- paramsDIAlignR()
params[["maxFdrQuery"]] <- 1.0
params[["maxPeptideFdr"]] <- 1.0
params[["kernelLen"]] <- 11
params$batchSize <- 10000L
params$globalAlignmentFdr <- 0.001
BiocParallel::register(BiocParallel::MulticoreParam(workers = 6, log = FALSE, threshold = "INFO", stop.on.error = TRUE))
ropenms <- get_ropenms(condaEnv = "TricEnvr")
DIAlignR:::progAlignRuns2(dataPath = ".", params = params, outFile = "prog.tsv", ropenms = ropenms, ids = ids, applyFun = BiocParallel::bplapply)}
\usage{
progAlignRuns2(
  dataPath,
  params,
  outFile = "DIAlignR.tsv",
  ropenms,
  ids = NULL,
  oswMerged = TRUE,
  runs = NULL,
  newickTree = NULL,
  applyFun = lapply
)
}
\description{
ids <- as.integer(scan(file = "data/ids.txt"))
params <- paramsDIAlignR()
params[["maxFdrQuery"]] <- 1.0
params[["maxPeptideFdr"]] <- 1.0
params[["kernelLen"]] <- 11
params$batchSize <- 10000L
params$globalAlignmentFdr <- 0.001
BiocParallel::register(BiocParallel::MulticoreParam(workers = 6, log = FALSE, threshold = "INFO", stop.on.error = TRUE))
ropenms <- get_ropenms(condaEnv = "TricEnvr")
DIAlignR:::progAlignRuns2(dataPath = ".", params = params, outFile = "prog.tsv", ropenms = ropenms, ids = ids, applyFun = BiocParallel::bplapply)
}
