% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressive_alignment.R
\name{progSplit4}
\alias{progSplit4}
\title{Step 4 for progressive alignment}
\usage{
progSplit4(
  dataPath,
  params,
  outFile = "DIAlignR",
  oswMerged = TRUE,
  applyFun = lapply
)
}
\arguments{
\item{dataPath}{(string) path to xics and osw directory.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{outFile}{(string) name of the output file.}

\item{oswMerged}{(logical) TRUE if merged file from pyprophet is used.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\description{
Step 4 for progressive alignment
}
\seealso{
\code{\link{progAlignRuns}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-03-03
}
