% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressive_alignment.R
\name{progTree1}
\alias{progTree1}
\title{Step 1 for progressive alignment}
\usage{
progTree1(
  dataPath,
  params,
  outFile = "DIAlignR",
  oswMerged = TRUE,
  runs = NULL,
  newickTree = NULL,
  applyFun = lapply
)
}
\arguments{
\item{dataPath}{(string) path to xics and osw directory.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{outFile}{(string) name of the output file.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{runs}{(string) names of xics file without extension.}

\item{newickTree}{(string) guidance tree in newick format. Look up \code{\link{getTree}}.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\description{
Step 1 for progressive alignment
}
\seealso{
\code{\link{progAlignRuns}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-03-03
}
