% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplit_alignment.R
\name{psuedo_iRTlibrary}
\alias{psuedo_iRTlibrary}
\title{Builds assay library and pseudo iRT library}
\usage{
psuedo_iRTlibrary(
  psuedoSpectraPattern,
  msplitAssaySuffix,
  masterLibPath,
  startTime = 0,
  stopTime = 7200
)
}
\arguments{
\item{msplitAssaySuffix}{(numeric) suffix of MSPLIT assay library.}

\item{masterLibPath}{(numeric) path of the master library to be converted to sample-specific assay library.}

\item{startTime}{(integer) start time of the run in second. Usually it is 0.}

\item{stopTime}{(integer) stop time of the run in seconds. Usually it is 7200 sec (2 hour run).}

\item{msplitFilteredPattern}{(string) pattern of pseudo iRT spectra files. Pattern must start with '*'.}
}
\value{
(None) writes _norm_iRT_assaylib.tsv and _norm_masterLib.tsv files.
}
\description{
Builds assay library and pseudo iRT library
}
\examples{
psuedoSpectraPattern <- "*_MSPLITfiltered_iRT.txt"
msplitAssaySuffix <- "_Openswath_assaylib.tsv"
masterLibPath <- "masterLib.tsv"
\dontrun{
psuedo_iRTlibrary(psuedoSpectraPattern, msplitAssaySuffix, masterLibPath)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-14
}
\keyword{internal}
