% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplit_alignment.R
\name{psuedo_iRTspectra}
\alias{psuedo_iRTspectra}
\title{Get psuedo iRT spectra}
\usage{
psuedo_iRTspectra(
  msplitFilteredPattern,
  maxQval = 1e-04,
  zoneWidth = 300,
  minPep = 5L,
  maxPep = 10L,
  minIntensity = 1e+06
)
}
\arguments{
\item{msplitFilteredPattern}{(string) pattern of MSPLITfiltered.txt files. Pattern must start with '*'.}

\item{maxQval}{(numeric) upper limit of qvalue for spectra to be considered for psuedo iRT.}

\item{zoneWidth}{(numeric)}

\item{minPep}{(integer) minimum number of precursors requires in each retention time zone.}

\item{maxPep}{(integer) maximum number of precursors requires in each retention time zone.}

\item{minIntensity}{(numeric) lower limit of ion count for spectra to be considered for psuedo iRT.}
}
\value{
(None) writes *_iRT.txt files.
}
\description{
This function cuts the retention time axis in multiple zones from which intense and confident spectra
are selected for the purspose of buillding psuedo iRT library.
}
\examples{
msplitFilteredPattern <- "*_MSPLITfiltered.txt"
\dontrun{
psuedo_iRTspectra(msplitFilteredPattern)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-14
}
\keyword{internal}
