% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzml.R
\name{readChromatogramHeader}
\alias{readChromatogramHeader}
\title{Get a table of chromatogram indices and respective transition IDs.}
\usage{
readChromatogramHeader(mzmlName)
}
\value{
A data-frame.
}
\description{
Get a table of chromatogram indices and respective transition IDs.
}
