% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzml.R
\name{readSqMassHeader}
\alias{readSqMassHeader}
\title{Get chromatogram header from a sqMass file}
\usage{
readSqMassHeader(con)
}
\arguments{
\item{mzmlName}{(char) path to xics file.}
}
\value{
(A data-frame) It has 10 columns. The two important columns are:
\item{chromatogramId}{(integer) Fragment-ion ID that matches with transition ID in osw file.}
\item{chromatogramIndex}{(integer) Index of chromatogram in mzML file.}
}
\description{
Get a table of chromatogram indices and respective transition IDs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
sqName <-paste0(dataPath,"/xics/hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt.chrom.sqMass")
\dontrun{
chromHead <- readChromatogramHeader(sqName)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-12-25
}
\keyword{internal}
