% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_area.R
\name{recalculateIntensity}
\alias{recalculateIntensity}
\title{Calculates area of peaks in peakTable}
\usage{
recalculateIntensity(
  peakTable,
  dataPath = ".",
  oswMerged = TRUE,
  XICfilter = "sgolay",
  polyOrd = 4,
  kernelLen = 9,
  baseSubtraction = TRUE,
  baselineType = "base_to_base",
  integrationType = "intensity_sum",
  fitEMG = FALSE,
  smoothPeakArea = FALSE
)
}
\arguments{
\item{peakTable}{(data-frame) usually an output of alignTargetedRuns. Must have these columns: run, precursor, leftWidth, rightWidth.}

\item{dataPath}{(string) path to mzml and osw directory.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{XICfilter}{(string) must be either sgolay, boxcar, gaussian, loess or none.}

\item{polyOrd}{(integer) order of the polynomial to be fit in the kernel.}

\item{kernelLen}{(integer) number of data-points to consider in the kernel.}

\item{baseSubtraction}{(logical) TRUE: remove background from peak signal using estimated noise levels.}

\item{baselineType}{(string) method to estimate the background of a peak contained in XICs. Must be
from "base_to_base", "vertical_division_min", "vertical_division_max".}

\item{integrationType}{(string) method to ompute the area of a peak contained in XICs. Must be
from "intensity_sum", "trapezoid", "simpson".}

\item{fitEMG}{(logical) enable/disable exponentially modified gaussian peak model fitting.}

\item{smoothPeakArea}{(logical) FALSE: raw chromatograms will be used for quantification. TRUE: smoothed chromatograms will be used for quantification.}
}
\value{
(data-frame)
}
\description{
For the give peak boundary in peakTable, the function extracts raw chromatograms and recalculate intensities.
}
\examples{
peakTable <- data.frame(precursor = c(1967L, 1967L, 2474L, 2474L),
                   run = rep(c("hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt",
                   "hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"), 2),
                   intensity = c(186.166, 579.832, 47.9525, 3.7413),
                   leftWidth = c(5001.76, 5025.66, 6441.51, 6516.6),
                   rightWidth = c(5076.86, 5121.25, 6475.65, 6554.2), stringsAsFactors = FALSE)
dataPath <- system.file("extdata", package = "DIAlignR")
newTable <- recalculateIntensity(peakTable, dataPath)
}
\seealso{
\code{\link{alignTargetedRuns}, \link{calculateIntensity}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-28
}
