% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_files.R
\name{reduceXICs}
\alias{reduceXICs}
\title{Subset an XIC file}
\usage{
reduceXICs(nativeIDs, xicFileIn, xicFileOut)
}
\arguments{
\item{nativeIDs}{(integer) transition IDs to be kept.}

\item{xicFileIn}{(character) name to the current file.}

\item{xicFileOut}{(character) name of the new file.}
}
\value{
(None)
}
\description{
Create a sqMass file that has chromatograms for given native IDs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
oswIn <- file.path(dataPath, "osw", "merged.osw")
params <- paramsDIAlignR()
params[["context"]] <- "experiment-wide"
ids <- getNativeIDs(oswIn, 1338L, params)
xicFileIn <- file.path(dataPath, "xics", "hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt.chrom.sqMass")
reduceXICs(ids, xicFileIn, xicFileOut = "temp.chrom.sqMass")
file.remove("temp.chrom.sqMass")
}
\seealso{
\code{\link{createSqMass}, \link{getNativeIDs}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2022) + GPL-3
Date: 2022-04-19
}
