% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{script2}
\alias{script2}
\title{Performs alignment using script1 output}
\usage{
script2(
  dataPath,
  outFile = "DIAlignR",
  params = paramsDIAlignR(),
  oswMerged = TRUE,
  runs = NULL,
  refRun = NULL,
  applyFun = lapply
)
}
\arguments{
\item{dataPath}{(string) path to mzml and osw directory.}

\item{outFile}{(string) name of the output file.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{runs}{(a vector of string) names of mzml file without extension.}

\item{refRun}{(string) reference for alignment. If no run is provided, m-score is used to select reference run.}

\item{applyFun}{(function) value must be either lapply or BiocParallel::bplapply.}
}
\description{
Performs alignment using script1 output
}
\examples{
params <- paramsDIAlignR()
params[["context"]] <- "experiment-wide"
dataPath <- system.file("extdata", package = "DIAlignR")
BiocParallel::register(BiocParallel::MulticoreParam(workers = 4, progressbar = TRUE))
script1(dataPath, outFile = "testDIAlignR", params = params, applyFun = BiocParallel::bplapply)
script2(dataPath, outFile = "testDIAlignR", params = params, applyFun = lapply)
file.remove(file.path(dataPath, "testDIAlignR_script1.RData"))
}
\seealso{
\code{\link{alignTargetedRuns}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-02-20
}
