% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectChromIndices}
\alias{selectChromIndices}
\title{Provides the index of reference run.}
\usage{
selectChromIndices(oswFiles, runname, analyte)
}
\value{
A vector of Integers
}
\description{
Provides the index of reference run.
}
