% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_alignment.R
\name{setAlignmentRank}
\alias{setAlignmentRank}
\title{Set Alignment rank to the aligned feature}
\usage{
setAlignmentRank(df, ref, eXp, tAligned, XICs.eXp, params, adaptiveRT)
}
\arguments{
\item{df}{(dataframe) a collection of features related to analyte_chr.}

\item{ref}{(string) name of the reference run. Must be in the rownames of fileInfo.}

\item{eXp}{(string) name of the run to be aligned to reference run. Must be in the rownames of fileInfo.}

\item{tAligned}{(list) the first element corresponds to the aligned reference time,
the second element is the aligned experiment time.}

\item{XICs.eXp}{(list) list of extracted ion chromatograms from experiment run.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{adaptiveRT}{(numeric) defines the window around the aligned retention time, within which
features with m-score below aligned FDR are considered for quantification.}
}
\value{
(dataframe) a collection of aligned features related to analyte_chr.
}
\description{
Picks the top feature in run by comparing m-score to unaligned FDR and aligned FDR.
If no satisfactory feature is found, peak-integration is carried out by mapping left and right peak
boundaries from the reference feature and integrating area under the curve.
}
\examples{
data(multipeptide_DIAlignR, package="DIAlignR")
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
\dontrun{
# Use getAlignedTimes() to get tAligned.
setAlignmentRank(multipeptide_DIAlignR[["4618"]], ref = "run1", eXp = "run2", tAligned, XICs.ref,
params, adaptiveRT)
multipeptide[["4618"]]
}
}
\seealso{
\code{\link{getMultipeptide}, \link{calculateIntensity}, \link{alignToRef}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-04-13
}
\keyword{internal}
