% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setAlignObj_S4}
\alias{setAlignObj_S4}
\title{Get a dummy S4 object of C++ class AlignObj}
\usage{
setAlignObj_S4(ROW_SIZE, COL_SIZE)
}
\arguments{
\item{ROW_SIZE}{(int) Number of rows of a matrix}

\item{COL_SIZE}{(int) Number of columns of a matrix}
}
\value{
AlignObj (S4class) A S4class dummy object from C++ AlignObj struct
}
\description{
Get a dummy S4 object of C++ class AlignObj
}
\examples{
x <- setAlignObj_S4(4, 5)
x@signalA_len # 3
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
