% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sgolayCpp}
\alias{sgolayCpp}
\title{Smooth chromatogram with savitzky-golay filter.}
\usage{
sgolayCpp(chrom, kernelLen, polyOrd)
}
\arguments{
\item{chrom}{(matrix) chromatogram containing time and intensity vectors.}

\item{kernelLen}{(integer) length of filter. Must be an odd number.}

\item{polyOrd}{(integer) TRUE: remove background from peak signal using estimated noise levels.}
}
\value{
(matrix).
}
\description{
Smooth chromatogram with savitzky-golay filter.
}
\examples{
data("XIC_QFNNTDIVLLEDFQK_3_DIAlignR", package = "DIAlignR")
XICs <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
xic <- sgolayCpp(as.matrix(XICs[[1]]), kernelLen = 11L, polyOrd = 4L)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2020) + MIT
Date: 2019-12-31
}
