% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_intensity.R
\name{sgolayFill}
\alias{sgolayFill}
\title{Fill missing values using spline}
\usage{
sgolayFill(chrom, polyOrd, kernelLen)
}
\arguments{
\item{chrom}{(data-frames) first column is time (must be equidistant), second column is intensity.}

\item{polyOrd}{(integer) must be less than kernelLen.}

\item{kernelLen}{(integer) must be an odd integer.}
}
\value{
(dataframe) has two columns:
\item{time}{(numeric)}
\item{intensity}{(numeric)}
}
\description{
Fill missing values using spline
}
\examples{
time <- seq(from = 3003.4, to = 3048, by = 3.4)
y <- c(0.2050595, 0.8850070, 2.2068768, 3.7212677, 5.1652605, 5.8288915, 5.5446804,
 4.5671360, 3.3213154, 1.9485889, 0.9520709, 0.3294218, 0.2009581, 0.1420923)
chrom <- data.frame(time, y)
chrom$y[c(1,8)] <- NA
\dontrun{
sgolayFill(chrom, polyOrd = 3, kernelLen = 9)
}
}
\seealso{
\code{\link[pracma]{polyfit}, \link[signal]{sgolayfilt}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-21
}
\keyword{internal}
