% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatogram_smooth.R
\name{smoothXICs}
\alias{smoothXICs}
\title{Smooth chromatogram signals from a list}
\usage{
smoothXICs(XICs, type = "none", samplingTime = NULL,
  kernelLen = NULL, polyOrd = NULL)
}
\arguments{
\item{XICs}{(A list) A list of dataframe that consists of two columns. First column must be
monotonically increasing.}

\item{type}{(char) must be either sgolay, boxcar, gaussian, loess or none.}

\item{samplingTime}{(numeric) Time difference between neighboring points.}

\item{kernelLen}{(integer) Number of data-points to consider in the kernel.}

\item{polyOrd}{(integer) Order of the polynomial to be fit in the kernel.}
}
\value{
A list.
}
\description{
Smoothing methods are Savitzky-Golay, Boxcar, Gaussian kernel and LOESS. Savitzky-Golay smoothing
is good at preserving peak-shape compared to gaussian and boxcar smoothing. However, it assumes
equidistant points that fortunately is the case for DIA data. This requires a quadratic memory
to store the fit and slower than other smoothing methods.
}
\examples{
data("XIC_QFNNTDIVLLEDFQK_3_DIAlignR")
XICs <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run0"]][["14299_QFNNTDIVLLEDFQK/3"]]
newXICs <- smoothXICs(XICs, type = "sgolay", samplingTime = 3.42, kernelLen = 9,
 polyOrd = 3)
}
\seealso{
\url{https://terpconnect.umd.edu/~toh/spectrum/Smoothing.html},
 \url{https://rafalab.github.io/dsbook/smoothing.html}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL3
Date: 2020-02-21
}
