% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstAlign.R
\name{traverseMST}
\alias{traverseMST}
\title{Reorder MST}
\usage{
traverseMST(net, ref0)
}
\arguments{
\item{net}{(matrix) array of tree-edges.}

\item{ref0}{(character) start node for traversal.}
}
\value{
(matrix) array of tree-edges reordered for traversal.
}
\description{
Reorder MST for traversal from a given node.
}
\examples{
net <- cbind("A" = c("run1","run2","run3"), "B" = c("run2","run3","run4"))
\dontrun{
traverseMST(net, "run2")
}
}
\seealso{
\code{\link{mstAlignRuns}, \link{getMST}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-05-14
}
\keyword{internal}
