% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_order.R
\name{traverseUp}
\alias{traverseUp}
\title{Get node IDs from tree}
\usage{
traverseUp(
  tree,
  dataPath,
  fileInfo,
  features,
  mzPntrs,
  prec2chromIndex,
  precursors,
  params,
  adaptiveRTs,
  refRuns,
  ropenms
)
}
\arguments{
\item{tree}{(phylo) a phylogenetic tree.}

\item{dataPath}{(string) path to mzml and osw directory.}

\item{fileInfo}{(data-frame) output of \code{\link{getRunNames}}.}

\item{features}{(list of data-frames) contains features and their properties identified in each run.}

\item{mzPntrs}{(list) a list of mzRpwiz.}

\item{prec2chromIndex}{(list) A list of dataframes having following columns: \cr
transition_group_id: it is PRECURSOR.ID from osw file. \cr
chromatogramIndex: index of chromatogram in mzML file.}

\item{precursors}{(data-frame) atleast two columns transition_group_id and transition_ids are required.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}
}
\value{
(integer) a vector with names being node IDs.
}
\description{
Get node IDs from tree
}
\examples{
m <- matrix(c(0,1,2, 1,0,1.5, 2,1.5,0), byrow = TRUE, ncol = 3,
 dimnames = list(c("run1", "run2", "run3"), c("run1", "run2", "run3")))
distMat <- as.dist(m)
\dontrun{
tree <- getTree(distMat)
traverseUp(tree, dataPath, fileInfo, features, mzPntrs, prec2chromIndex, precursors, params)
}
}
\seealso{
\code{\link{getTree}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-01
}
\keyword{internal}
