% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_runs.R
\name{trfrParentFeature}
\alias{trfrParentFeature}
\title{Transform features to child time-domain}
\usage{
trfrParentFeature(XICs, timeParent, df, i, params)
}
\arguments{
\item{XICs}{(list of data-frames) extracted ion chromatograms from the child run.}

\item{timeParent}{(data-frame) has two columns: tAligned and alignedChildTime. tAligned is time vector from one of the parent run.}

\item{df}{(data-frame) contains features to be transformed. It has a format of \code{\link{getFeatures}} output.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}
}
\value{
(data-frame) this has a format of \code{\link{getFeatures}} output.
}
\description{
This function transforms the peaks' times to child run's time-domain. The feature intensity is
calculated with appropriate method stated in params.
Internal missing values are not allowed in timeParent.
}
\examples{
data(masterXICs_DIAlignR, package="DIAlignR")
newXICs <- masterXICs_DIAlignR
timeParent <- newXICs[[2]][, c("tAligned.ref", "alignedChildTime")]
colnames(timeParent) <- c("tAligned", "alignedChildTime")
params <- paramsDIAlignR()
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- DIAlignR::getRunNames(dataPath = dataPath)
features <- getFeatures(fileInfo, maxFdrQuery = 1.00, runType = "DIA_proteomics")
df <- features$run1[features$run1$transition_group_id == 4618L, ]
\dontrun{
trfrParentFeature(newXICs[[1]], timeParent, df, params)
}
}
\seealso{
\code{\link{getChildFeature}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-17
}
\keyword{internal}
