% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatogram_manip.R
\name{trimXICs}
\alias{trimXICs}
\title{Selects a part of chromatograms}
\usage{
trimXICs(XICs, len = 1)
}
\arguments{
\item{XICs}{(A list) A list of dataframe that consists of two columns. First column must be
monotonically increasing.}

\item{length}{(numeric) must be between 0.1 and 1.}
}
\value{
A list.
}
\description{
This function trims chromatograms from the end-points.
}
\examples{
data("XIC_QFNNTDIVLLEDFQK_3_DIAlignR")
XICs <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run0"]][["14299_QFNNTDIVLLEDFQK/3"]]
\dontrun{
newXICs <- smoothXICs(XICs, len = 0.5)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL3
Date: 2020-04-01
}
