% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{uncompressVec}
\alias{uncompressVec}
\title{Uncompress a Blob object}
\usage{
uncompressVec(x, type)
}
\arguments{
\item{x}{(Blob object)}

\item{type}{(integer) must either be 5L or 6L to indicate linear and short logged float compression, respectively.}
}
\value{
A numeric vector. Uncompressed form of the Blob.
}
\description{
Uncompress a Blob object
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-12-13
}
\keyword{internal}
