% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeOutFeatureAlignmentMap}
\alias{writeOutFeatureAlignmentMap}
\title{Write out alignment map to disk}
\usage{
writeOutFeatureAlignmentMap(multiFeatureAlignmentMap, oswMerged, fileInfo)
}
\arguments{
\item{multiFeatureAlignmentMap}{(list) contains multiple data-frames that are collection of experiment feature ids
mapped to corresponding reference feature id per analyte. This is an output of \code{\link{getRefExpFeatureMap}}.}

\item{oswMerged}{(logical) TRUE if merged file from pyprophet is used.}

\item{fileInfo}{(data-frame) output of \code{\link{getRunNames}}.}
}
\description{
Save alignment mapping to disk, either append table to OSW file, or save TSV file(s)
}
\author{
Justin Sing, \email{justinc.sing@mail.utoronto.ca}

ORCID: 0000-0003-0386-0092

License: (c) Author (2020) + GPL-3
Date: 2022-11-07
}
