% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeTables}
\alias{writeTables}
\title{Writes the output table for DIAlignR}
\usage{
writeTables(refAnalytes, runs, filename, intesityTbl, rtTbl, lwTbl, rwTbl)
}
\arguments{
\item{refAnalytes}{(list of data-frames) it is the output from getOswFiles function.}

\item{filename}{(list of data-frames) it is the output from getOswFiles function.}

\item{intesityTbl}{(string) Must be a combination of "run" and an iteger e.g. "run2".}

\item{rtTbl}{(string) analyte is as PRECURSOR.GROUP_LABEL or as PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}

\item{lwTbl}{dd}

\item{rwTbl}{fff}
}
\value{
None
}
\description{
select chromatogram indices by matching analyte and runname in oswFiles.
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
\dontrun{
selectChromIndices(oswFiles = oswFiles_DIAlignR, runname = "run2", analyte = "AQPPVSTEY_2")
selectChromIndices(oswFiles = oswFiles_DIAlignR, runname = "run0",
 analyte = "14299_QFNNTDIVLLEDFQK/3")
}
}
\seealso{
\code{\link{getOswFiles}, \link{getOswAnalytes}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
\keyword{internal}
