% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{XIC_QFNNTDIVLLEDFQK_3_DIAlignR}
\alias{XIC_QFNNTDIVLLEDFQK_3_DIAlignR}
\title{Extracted-ion chromatograms (XICs) of a peptide}
\format{A list of three elements where each element consists of a list of six data frames. Each
 data frame has two columns:
\describe{
  \item{time}{Retention time of ananlyte in the run, in sec}
  \item{intensity}{Intensity of signal for the transition}
}}
\source{
data-raw/test_GenerateData.R has source code.
}
\usage{
XIC_QFNNTDIVLLEDFQK_3_DIAlignR
}
\description{
XICs of peptide QFNNTDIVLLEDFQK/3 from three SWATH runs:

run0 : hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt.chrom.mzML\cr
run1 : hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML\cr
run2 : hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML
}
\keyword{datasets}
