% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_aligned_obj.R
\docType{class}
\name{AffineAlignObj-class}
\alias{AffineAlignObj-class}
\alias{AffineAlignObj}
\title{An S4 object for class AffineAlignObj}
\description{
s is a point-wise similarity matrix between signalA and signalB.
Intermediate matrices M,A,B are calculated from s for affine-alignment. Each cell of the Traceback
matrix has coordinate of its parent cell. path matrix is a binary matrix with ones indicating path
of maximum cumulative score.
GapOpen and GapExten are gap-opening and gap-extension penalties used by affine alignment algorithm.
indexA_aligned and indexB_aligned are aligned indices of signalA and SignalB. The cumulative
alignment score is in score vector.
}
\seealso{
\code{\link{doAffineAlignmentCpp}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
