% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_aligned_obj.R
\docType{class}
\name{AffineAlignObjLight-class}
\alias{AffineAlignObjLight-class}
\alias{AffineAlignObjLight}
\title{An S4 object for class AffineAlignObj. It only contains aligned indices.}
\description{
indexA_aligned and indexB_aligned are aligned indices of signalA and SignalB. The cumulative
alignment score is in score vector.
}
\seealso{
\code{\link{doAffineAlignmentCpp}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
