% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine_aligned_obj.R
\docType{class}
\name{AffineAlignObjLight-class}
\alias{AffineAlignObjLight-class}
\alias{AffineAlignObjLight}
\title{An S4 object for class AffineAlignObj. It only contains aligned indices.}
\description{
An S4 object for class AffineAlignObj. It only contains aligned indices.
}
