% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dia_runs.R
\name{alignTargetedRuns}
\alias{alignTargetedRuns}
\title{Outputs intensities for each analyte from aligned Targeted-MS runs}
\usage{
alignTargetedRuns(dataPath, alignType = "hybrid",
  analyteInGroupLabel = FALSE, oswMerged = TRUE, runs = NULL,
  analytes = NULL, nameCutPattern = "(.*)(/)(.*)",
  maxFdrQuery = 0.05, maxFdrLoess = 0.01, analyteFDR = 0.01,
  spanvalue = 0.1, runType = "DIA_Proteomics",
  normalization = "mean", simMeasure = "dotProductMasked",
  XICfilter = "sgolay", SgolayFiltOrd = 4, SgolayFiltLen = 9,
  goFactor = 0.125, geFactor = 40, cosAngleThresh = 0.3,
  OverlapAlignment = TRUE, dotProdThresh = 0.96, gapQuantile = 0.5,
  hardConstrain = FALSE, samples4gradient = 100, samplingTime = 3.4,
  RSEdistFactor = 3.5, saveFiles = FALSE)
}
\arguments{
\item{dataPath}{(char) Path to mzml and osw directory.}

\item{alignType}{Available alignment methods are "global", "local" and "hybrid".}

\item{analyteInGroupLabel}{(logical) TRUE for getting analytes as PRECURSOR.GROUP_LABEL from osw file.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{runs}{(A vector of string) Names of mzml file without extension.}

\item{analytes}{(vector of strings) transition_group_ids for which features are to be extracted. analyteInGroupLabel must be set according the pattern used here.}

\item{nameCutPattern}{(string) regex expression to fetch mzML file name from RUN.FILENAME columns of osw files.}

\item{maxFdrQuery}{(numeric) A numeric value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{maxFdrLoess}{(numeric) A numeric value between 0 and 1. Features should have m-score lower than this value for participation in LOESS fit.}

\item{analyteFDR}{(numeric) only analytes that have m-score less than this, will be included in the output.}

\item{spanvalue}{(numeric) Spanvalue for LOESS fit. For targeted proteomics 0.1 could be used.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{normalization}{(character) Must be selected from "mean", "l2".}

\item{simMeasure}{(string) Must be selected from dotProduct, cosineAngle,
cosine2Angle, dotProductMasked, euclideanDist, covariance and correlation.}

\item{XICfilter}{(string) This must be one of the strings "sgolay", "none".}

\item{SgolayFiltOrd}{(integer) It defines the polynomial order of filer.}

\item{SgolayFiltLen}{(integer) Must be an odd number. It defines the length of filter.}

\item{goFactor}{(numeric) Penalty for introducing first gap in alignment. This value is multiplied by base gap-penalty.}

\item{geFactor}{(numeric) Penalty for introducing subsequent gaps in alignment. This value is multiplied by base gap-penalty.}

\item{cosAngleThresh}{(numeric) In simType = dotProductMasked mode, angular similarity should be higher than cosAngleThresh otherwise similarity is forced to zero.}

\item{OverlapAlignment}{(logical) An input for alignment with free end-gaps. False: Global alignment, True: overlap alignment.}

\item{dotProdThresh}{(numeric) In simType = dotProductMasked mode, values in similarity matrix higher than dotProdThresh quantile are checked for angular similarity.}

\item{gapQuantile}{(numeric) Must be between 0 and 1. This is used to calculate base gap-penalty from similarity distribution.}

\item{hardConstrain}{(logical) If FALSE; indices farther from noBeef distance are filled with distance from linear fit line.}

\item{samples4gradient}{(numeric) This parameter modulates penalization of masked indices.}

\item{samplingTime}{(numeric) Time difference between two data-points in each chromatogram. For hybrid and local alignment, samples are assumed to be equally time-spaced.}

\item{RSEdistFactor}{(numeric) This defines how much distance in the unit of rse remains a noBeef zone.}

\item{saveFiles}{(logical) Must be selected from light, medium and heavy.}
}
\value{
Two tables of intensity and rention times for every analyte in each run.
}
\description{
This function expects osw and mzml directories at dataPath. It first reads osw files and fetches chromatogram indices for each analyte.
It then align XICs of each analyte to its reference XICs. Best peak, which has lowest m-score, about the aligned retention time is picked for quantification.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt",
 "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt")
intensityTbl <- alignTargetedRuns(dataPath, runs = runs, analytes = c("QFNNTDIVLLEDFQK_3"),
 analyteInGroupLabel = FALSE)
intensityTbl <- alignTargetedRuns(dataPath, runs = runs, analytes = c("14299_QFNNTDIVLLEDFQK/3"),
 analyteInGroupLabel = TRUE)
}
\references{
Gupta S, Ahadi S, Zhou W, Röst H. "DIAlignR Provides Precise Retention Time Alignment Across Distant Runs in DIA and Targeted Proteomics." Mol Cell Proteomics. 2019 Apr;18(4):806-817. doi: https://doi.org/10.1074/mcp.TIR118.001132 Epub 2019 Jan 31.
}
\seealso{
\code{\link{getRunNames}, \link{getOswFiles}, \link{getAnalytesName}, \link{getMappedRT}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-14
}
