% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dia_runs.R
\name{alignToRef}
\alias{alignToRef}
\title{Aligns an analyte from an experiment to the reference run}
\usage{
alignToRef(
  eXp,
  ref,
  analyte_chr,
  rownum,
  fileInfo,
  XICs.ref.s,
  params,
  prec2chromIndex,
  mzPntrs,
  df,
  globalFits,
  RSE
)
}
\arguments{
\item{eXp}{(string) name of the run to be aligned to reference run. Must be in the rownames of fileInfo.}

\item{ref}{(string) name of the reference run. Must be in the rownames of fileInfo.}

\item{analyte_chr}{(string) Precursor ID of the requested analyte.}

\item{rownum}{(integer) represnts the index of the multipepetide to be aligned.}

\item{fileInfo}{(data-frame) output of \code{\link{getRunNames}}.}

\item{XICs.ref.s}{(list of dataframes) Smoothed fragment-ion chromatograms of the analyte_chr from the reference run.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{prec2chromIndex}{(list) a list of dataframes having following columns: \cr
transition_group_id: it is PRECURSOR.ID from osw file. \cr
chromatogramIndex: index of chromatogram in mzML file.}

\item{mzPntrs}{(list) a list of mzRpwiz.}

\item{df}{(dataframe) a collection of features related to analyte_chr.}

\item{globalFits}{(list) each element is either of class lm or loess. This is an output of \code{\link{getGlobalFits}}.}

\item{RSE}{(list) Each element represents Residual Standard Error of corresponding fit in globalFits.}
}
\value{
(dataframe) aligned features of analyte_chr in eXp run.
}
\description{
df contains unaligned features for an analyte across multiple runs. This function aligns eXp run to
ref run and updates corresponding features.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
}
\seealso{
\code{\link{alignTargetedRuns}, \link{alignIthAnalyte}, \link{setAlignmentRank}, \link{getMultipeptide}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-26
}
\keyword{internal}
