% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dia_runs.R
\name{alignToRef}
\alias{alignToRef}
\title{Aligns an analyte from an experiment to the reference run}
\usage{
alignToRef(
  eXp,
  ref,
  refIdx,
  fileInfo,
  XICs,
  XICs.ref,
  params,
  df,
  globalFits,
  RSE,
  feature_alignment_map = NULL
)
}
\arguments{
\item{eXp}{(string) name of the run to be aligned to reference run. Must be in the rownames of fileInfo.}

\item{ref}{(string) name of the reference run. Must be in the rownames of fileInfo.}

\item{refIdx}{(integer) index of the reference feature in df.}

\item{fileInfo}{(data-frame) output of \code{\link{getRunNames}}.}

\item{XICs}{(list of dataframes) fragment-ion chromatograms of the analytes for all runs.}

\item{XICs.ref}{(list of dataframes) fragment-ion chromatograms of the analyte_chr from the reference run.}

\item{params}{(list) parameters are entered as list. Output of the \code{\link{paramsDIAlignR}} function.}

\item{df}{(dataframe) a collection of features related to the peptide}

\item{globalFits}{(list) each element is either of class lm or loess. This is an output of \code{\link{getGlobalFits}}.}

\item{RSE}{(list) Each element represents Residual Standard Error of corresponding fit in globalFits.}

\item{feature_alignment_mapping}{(data.table)  contains experiment feature ids
mapped to corresponding reference feature id per analyte. This is an output of \code{\link{getRefExpFeatureMap}}.}
}
\value{
invisible NULL
}
\description{
df contains unaligned features for an analyte across multiple runs. This function aligns eXp run to
ref run and updates corresponding features.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
}
\seealso{
\code{\link{alignTargetedRuns}, \link{perBatch}, \link{setAlignmentRank}, \link{getMultipeptide}, \link{getRefExpFeatureMap}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-07-26
}
\keyword{internal}
