% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{areaIntegrator}
\alias{areaIntegrator}
\title{Calculates area between signal-boundaries.}
\usage{
areaIntegrator(
  l1,
  l2,
  left,
  right,
  integrationType,
  baselineType,
  fitEMG,
  baseSubtraction
)
}
\arguments{
\item{l1}{(list) A list of time vectors.}

\item{l2}{(list) A list of intensity vectors.}

\item{left}{(numeric) left boundary of the peak.}

\item{right}{(numeric) right boundary of the peak.}

\item{integrationType}{(string) method to ompute the area of a peak contained in XICs. Must be
from "intensity_sum", "trapezoid", "simpson".}

\item{baselineType}{(string) method to estimate the background of a peak contained in XICs. Must be
from "base_to_base", "vertical_division_min", "vertical_division_max".}

\item{fitEMG}{(logical) enable/disable exponentially modified gaussian peak model fitting.}

\item{baseSubtraction}{(logical) TRUE: remove background from peak signal using estimated noise levels.}
}
\value{
area (numeric).
}
\description{
This function sums all the intensities between left-index and right-index.
}
\examples{
data("XIC_QFNNTDIVLLEDFQK_3_DIAlignR", package = "DIAlignR")
XICs <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run1"]][["14299_QFNNTDIVLLEDFQK/3"]]
l1 <- lapply(XICs, `[[`, 1)
l2 <- lapply(XICs, `[[`, 2)
areaIntegrator(l1, l2, left = 5203.7, right = 5268.5, "intensity_sum", "base_to_base", FALSE, TRUE)
# 224.9373
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
