% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_chromatogram.R
\name{childXIC}
\alias{childXIC}
\title{Get a child chromatogram from parents}
\usage{
childXIC(
  XIC.ref,
  XIC.eXp,
  alignedIndices,
  method = "spline",
  polyOrd = 4,
  kernelLen = 9,
  splineMethod = "fmm",
  mergeStrategy = "avg",
  keepFlanks = TRUE
)
}
\arguments{
\item{XIC.ref}{(data-frame) extracted ion chromatogram from reference run. Must not contain missing values.}

\item{XIC.eXp}{(data-frame) extracted ion chromatogram from experiment run. Must not contain missing values.}

\item{alignedIndices}{(data-frame) must have two columns "indexAligned.ref" and "indexAligned.eXp".}

\item{method}{(string) must be either "spline", "sgolay" or "linear".}

\item{polyOrd}{(integer) must be less than kernelLen.}

\item{kernelLen}{(integer) must be an odd integer.}

\item{splineMethod}{(string) must be either "fmm" or "natural".}

\item{mergeStrategy}{(string) must be either ref, avg, refStart or refEnd.}

\item{keepFlanks}{(logical) TRUE: Flanking chromatogram is not removed.}
}
\value{
(list) the first element is chromatogram. The second element is aligned parent time-vectors.
}
\description{
Internal gaps in reference chromatogram and corresponding indices in experiment chromatogram
are discarded while merging both chromatogram to obtain a child chromatogram.
}
\section{time-merging}{

There are three startegies for calculating time to keep sampling rate equal to parents:\cr
 - Use reference time. \cr
 - Average time (Non-gap region will have parent's sampling rate). \cr
 - Start/end with the average then fixed sampling rate.
}

\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
data(alignObj_DIAlignR, package="DIAlignR")
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run1"]][["14299_QFNNTDIVLLEDFQK/3"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["run2"]][["14299_QFNNTDIVLLEDFQK/3"]]
alignedIndices <- cbind(alignObj_DIAlignR@indexA_aligned, alignObj_DIAlignR@indexB_aligned)
colnames(alignedIndices) <- c("indexAligned.ref", "indexAligned.eXp")
alignedIndices[, 1:2][alignedIndices[, 1:2] == 0] <- NA_integer_
\dontrun{
plot(childXIC(XICs.ref[[1]], XICs.eXp[[1]], alignedIndices)[[1]], type = 'l')
}
}
\seealso{
\code{\link{mergeXIC}, \link{alignedXIC}, \link{childXICs}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-23
}
\keyword{internal}
