% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osw_mzml.R
\name{chromatogramIdAsInteger}
\alias{chromatogramIdAsInteger}
\title{Coerce chromatogram ids as integer}
\usage{
chromatogramIdAsInteger(chromatogramHeader)
}
\arguments{
\item{chromatogramHeader}{(dataframe)}
}
\value{
Invisible NULL
}
\description{
chromatogramHeader has 10 columns. The two important columns are:
"chromatogramId" which has fragment-ion ID that matches with transition ID in osw file.
"chromatogramIndex" that lists indices of chromatograms in mzML file.
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
