% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyopenms.R
\name{createMZML}
\alias{createMZML}
\title{Create an mzML file}
\usage{
createMZML(ropenms, filename, XICs, transitionIDs)
}
\arguments{
\item{ropenms}{(pyopenms module) get this python module through get_ropenms().}

\item{filename}{(string) name of the mzML file to be written. Extension should be .chrom.mzML.}

\item{XICs}{(list of list of data-frames) list of extracted ion chromatograms of all precursors.}

\item{transitionIDs}{(list of integer) length must be the same as of XICs.}
}
\value{
(None)
}
\description{
Writes an mzML file having chromatograms and their native IDs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
filename <- paste0(dataPath, "/xics/hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML")
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR)
XICs <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt"]]
nativeIds <- list(27706:27711)
\dontrun{
ropenms <- get_ropenms(condaEnv = "envName")
createMZML(ropenms, "testfile.chrom.mzML", XICs, nativeIds)
mzR::chromatogramHeader(mzR::openMSfile("testfile.chrom.mzML", backend = "pwiz"))
file.remove("testfile.chrom.mzML")
}
}
\seealso{
\code{\link{get_ropenms}, \link{addXIC}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-06-06
}
