% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doAffineAlignment_S4}
\alias{doAffineAlignment_S4}
\title{Initialize a S4 object AffineAlignObj}
\usage{
doAffineAlignment_S4(s, signalA_len, signalB_len, go, ge, OverlapAlignment)
}
\arguments{
\item{seq1Len}{(int) Length of sequence1}

\item{seq2Len}{(int) Length of sequence2}
}
\value{
affineAlignObj (S4class) An object from C++ class of AffineAlignObj
}
\description{
Initialize a S4 object AffineAlignObj
}
\examples{
# Get sequence similarity of two DNA strings
Match=10; MisMatch=-2
seq1 = "GCAT"; seq2 = "CAGTG"
s <- getSeqSimMat(seq1, seq2, Match, MisMatch)
objAffine_Global <- doAffineAlignment_S4(s, 4, 5, 22, 7, FALSE)
objAffine_Global@score # -2  -4  -6  4 -18
objAffine_Olap <- doAffineAlignment_S4(s, 4, 5, 22, 7, TRUE)
objAffine_Olap@score # 0 10 20 18 18 18
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
