% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doAlignmentCpp}
\alias{doAlignmentCpp}
\title{Perform non-affine global and overlap alignment on a similarity matrix}
\usage{
doAlignmentCpp(sim, gap, OverlapAlignment)
}
\arguments{
\item{sim}{(NumericMatrix) A numeric matrix with similarity values of two sequences or signals.}

\item{gap}{(double) Penalty for introducing gaps in alignment.}

\item{OverlapAlignment}{(logical) An input for alignment with free end-gaps. False: Global alignment, True: overlap alignment.}
}
\value{
AlignObj (S4class) An object from C++ class of AlignObj.
}
\description{
Perform non-affine global and overlap alignment on a similarity matrix
}
\examples{
# Get sequence similarity of two DNA strings
Match=10; MisMatch=-2
seq1 = "GCAT"; seq2 = "CAGTG"
s <- getSeqSimMat(seq1, seq2, Match, MisMatch)
obj_Global <- doAlignmentCpp(s, 22, FALSE)
obj_Global@score # -2 -4 -6 4 -18
obj_Olap <- doAlignmentCpp(s, 22, TRUE)
obj_Olap@score # 0 10 20 18 18 18
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
