% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{extractXIC_group}
\alias{extractXIC_group}
\title{Extract XICs of all transitions requested in chromIndices.}
\usage{
extractXIC_group(mz, chromIndices, SgolayFiltOrd = 4,
  SgolayFiltLen = 9)
}
\value{
A list of data-frames. Each data frame has elution time and intensity of fragment-ion XIC.
}
\description{
Extract XICs of all transitions requested in chromIndices.
}
