% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osw.R
\name{fetchAnalytesInfo}
\alias{fetchAnalytesInfo}
\title{Fetch features of analytes}
\usage{
fetchAnalytesInfo(oswName, maxFdrQuery, oswMerged, analytes, filename,
  runType, analyteInGroupLabel = FALSE)
}
\arguments{
\item{oswName}{(char) path to the osw file.}

\item{maxFdrQuery}{(numeric) A numeric value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{analytes}{(vector of strings) transition_group_ids for which features are to be extracted. analyteInGroupLabel must be set according the pattern used here.}

\item{filename}{(data-frame) Should be from the RUN.FILENAME column from osw files.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{analyteInGroupLabel}{(logical) TRUE for getting analytes as PRECURSOR.GROUP_LABEL from osw file.
FALSE for fetching analytes as PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}
}
\value{
(data-frames) Data-frame has following columns:
\item{transition_group_id}{(string) it is either fetched from PRECURSOR.GROUP_LABEL or a combination of PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}
\item{filename}{(string) as mentioned in RUN table of osw files.}
\item{RT}{(numeric) retention time as in FEATURE.EXP_RT of osw files.}
\item{delta_rt}{(numeric) as in FEATURE.DELTA_RT of osw files.}
\item{assay_RT}{(numeric) library retention time as in PRECURSOR.LIBRARY_RT of osw files.}
\item{Intensity}{(numeric) peak intensity as in FEATURE_MS2.AREA_INTENSITY of osw files.}
\item{leftWidth}{(numeric) as in FEATURE.LEFT_WIDTH of osw files.}
\item{rightWidth}{(numeric) as in FEATURE.RIGHT_WIDTH of osw files.}
\item{peak_group_rank}{(integer) rank of each feature associated with transition_group_id.}
\item{m_score}{(numeric) q-value of each feature associated with transition_group_id.}
\item{transition_id}{(integer) fragment-ion ID associated with transition_group_id. This is matched with chromatogram ID in mzML file.}
}
\description{
Get a data-frame of analytes' transition_group_ids, their OpenSwath features, chromatogram indices and associated FDR-scores.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
filenames <- DIAlignR::getRunNames(dataPath = dataPath)
oswName <- paste0(dataPath,"/osw/merged.osw")
\dontrun{
analytesInfo <- fetchAnalytesInfo(oswName, maxFdrQuery = 0.05, oswMerged = TRUE,
 analytes = c("19051_KLIVTSEGC[160]FK/2"), filename = filenames$filename[2],
  runType = "DIA_proteomics", analyteInGroupLabel = TRUE)
analytesInfo <- fetchAnalytesInfo(oswName, maxFdrQuery = 0.05, oswMerged = TRUE,
 analytes = c("IHFLSPVRPFTLTPGDEEESFIQLITPVR_3"), filename = filenames$filename[3],
  runType = "DIA_proteomics", analyteInGroupLabel = FALSE)
}
}
\seealso{
\code{\link{getRunNames}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
\keyword{internal}
