% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osw.R
\name{fetchPrecursorsInfo}
\alias{fetchPrecursorsInfo}
\title{Get precursors from a feature file}
\usage{
fetchPrecursorsInfo(filename, runType, selectIDs = NULL)
}
\arguments{
\item{filename}{(string) Should be from the RUN.FILENAME column from osw files.}

\item{runType}{(string) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}
}
\value{
(data-frames) Data-frame has following columns:
\item{transition_group_id}{(integer) a unique id for each precursor.}
\item{transition_id}{(list) fragment-ion ID associated with transition_group_id. This is matched with chromatogram ID in mzML file.}
\item{peptide_id}{(integer) a unique id for each peptide. A peptide can have multiple precursors.}
\item{sequence}{(string) amino-acid sequence of the precursor with possible modifications.}
\item{charge}{(integer) charge on the precursor.}
\item{group_label}{(string) TODO Figure it out.}
}
\description{
Get a data-frame of analytes' transition_group_id, transition_ids, peptide_id and amino-acid sequences.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
filename <- paste0(dataPath,"/osw/merged.osw")
\dontrun{
precursorsInfo <- fetchPrecursorsInfo(filename, runType = "DIA_proteomics")
dim(precursorsInfo) # 303  6
}
}
\seealso{
\code{\link{getRunNames}, \link{getPrecursors}, \link{getPrecursorsQuery}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-04-04
}
\keyword{internal}
