% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filenames.R
\name{filenamesFromMZML}
\alias{filenamesFromMZML}
\title{Get mzML filenames from the directory.}
\usage{
filenamesFromMZML(dataPath, chromFile)
}
\arguments{
\item{dataPath}{(char) Path to mzml and osw directory.}
}
\value{
A dataframe with two columns:
\item{runName}{(string) contain respective mzML names without extension.}
\item{chromatogramFile}{(string) Path to the chromatogram file.}
}
\description{
Reads all mzML names avaialble in the directory.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
\dontrun{
filenamesFromMZML(dataPath)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
\keyword{internal}
