% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_filenames.R
\name{filenamesFromOSW}
\alias{filenamesFromOSW}
\title{Get mzML filenames from osw RUN table.}
\usage{
filenamesFromOSW(dataPath, pattern)
}
\arguments{
\item{dataPath}{(char) path to xics and osw directory.}

\item{pattern}{(char) must be either *.osw or *merged.osw .}
}
\value{
A dataframe with three columns:
\item{spectraFile}{(string) as mentioned in RUN table of osw files.}
\item{spectraFileID}{(string) ID in RUN table of osw files.}
\item{featureFile}{(string) Path to the feature file.}
}
\description{
Get mzML filenames from osw RUN table.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
\dontrun{
filenamesFromOSW(dataPath, "*.osw")
filenamesFromOSW(dataPath, "*merged.osw")
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
\keyword{internal}
