% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pep_peak_corp.R
\name{getAlignObj}
\alias{getAlignObj}
\title{Outputs AlignObj from an alignment of two XIC-groups}
\usage{
getAlignObj(
  XICs.ref,
  XICs.eXp,
  globalFit,
  alignType,
  adaptiveRT,
  normalization,
  simType,
  goFactor,
  geFactor,
  cosAngleThresh,
  OverlapAlignment,
  dotProdThresh,
  gapQuantile,
  kerLen,
  hardConstrain,
  samples4gradient,
  objType = "light"
)
}
\arguments{
\item{XICs.ref}{List of extracted ion chromatograms from reference run.}

\item{XICs.eXp}{List of extracted ion chromatograms from experiment run.}

\item{globalFit}{Linear or loess fit object between reference and experiment run.}

\item{alignType}{Available alignment methods are "global", "local" and "hybrid".}

\item{adaptiveRT}{(numeric) Similarity matrix is not penalized within adaptive RT.}

\item{normalization}{(character) Must be selected from "mean", "l2".}

\item{simType}{(string) Must be selected from dotProduct, cosineAngle, crossCorrelation,
cosine2Angle, dotProductMasked, euclideanDist, covariance and correlation.}

\item{goFactor}{(numeric) Penalty for introducing first gap in alignment. This value is multiplied by base gap-penalty.}

\item{geFactor}{(numeric) Penalty for introducing subsequent gaps in alignment. This value is multiplied by base gap-penalty.}

\item{cosAngleThresh}{(numeric) In simType = dotProductMasked mode, angular similarity should be higher than cosAngleThresh otherwise similarity is forced to zero.}

\item{OverlapAlignment}{(logical) An input for alignment with free end-gaps. False: Global alignment, True: overlap alignment.}

\item{dotProdThresh}{(numeric) In simType = dotProductMasked mode, values in similarity matrix higher than dotProdThresh quantile are checked for angular similarity.}

\item{gapQuantile}{(numeric) Must be between 0 and 1. This is used to calculate base gap-penalty from similarity distribution.}

\item{kerLen}{(integer) In simType = crossCorrelation, length of the kernel used to sum similarity score. Must be an odd number.}

\item{hardConstrain}{(logical) If FALSE; indices farther from noBeef distance are filled with distance from linear fit line.}

\item{samples4gradient}{(numeric) This parameter modulates penalization of masked indices.}

\item{objType}{(char) Must be selected from light, medium and heavy.}
}
\value{
A S4 object. Three most-important slots are:
\item{indexA_aligned}{(integer) aligned indices of reference run.}
\item{indexB_aligned}{(integer) aligned indices of experiment run.}
\item{score}{(numeric) cumulative score of alignment.}
}
\description{
Outputs AlignObj from an alignment of two XIC-groups
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
data(oswFiles_DIAlignR, package="DIAlignR")
run1 <- "hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"
run2 <- "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[[run1]][["4618"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[[run2]][["4618"]]
globalFit <- getGlobalAlignment(oswFiles_DIAlignR, ref = "run1", eXp = "run2",
 maxFdrGlobal = 0.05, spanvalue = 0.1)
AlignObj <- getAlignObj(XICs.ref, XICs.eXp, globalFit, alignType = "hybrid", adaptiveRT = 77.82315,
 normalization = "mean", simType = "dotProductMasked", goFactor = 0.125,
  geFactor = 40, cosAngleThresh = 0.3, OverlapAlignment = TRUE, dotProdThresh = 0.96,
  gapQuantile = 0.5, kerLen = 9L, hardConstrain = FALSE, samples4gradient = 100, objType = "light")
}
\seealso{
\code{\link{alignChromatogramsCpp}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
