% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_pep_peak_corp.R
\name{getAlignObj}
\alias{getAlignObj}
\title{Outputs AlignObj from an alignment of two XIC-groups}
\usage{
getAlignObj(XICs.ref, XICs.eXp, Loess.fit, adaptiveRT, samplingTime,
  normalization, simType, goFactor, geFactor, cosAngleThresh,
  OverlapAlignment, dotProdThresh, gapQuantile, hardConstrain,
  samples4gradient, objType = "light")
}
\arguments{
\item{XICs.ref}{List of extracted ion chromatograms from reference run.}

\item{XICs.eXp}{List of extracted ion chromatograms from experiment run.}

\item{Loess.fit}{LOESS fit object between reference and experiment run.}

\item{adaptiveRT}{(numeric) Similarity matrix is not penalized within adaptive RT.}

\item{samplingTime}{(numeric) Time difference between two data-points in each chromatogram. For hybrid and local alignment, samples are assumed to be equally time-spaced.}

\item{normalization}{(character) Must be selected from "mean", "l2".}

\item{simType}{(string) Must be selected from dotProduct, cosineAngle,
cosine2Angle, dotProductMasked, euclideanDist, covariance and correlation.}

\item{goFactor}{(numeric) Penalty for introducing first gap in alignment. This value is multiplied by base gap-penalty.}

\item{geFactor}{(numeric) Penalty for introducing subsequent gaps in alignment. This value is multiplied by base gap-penalty.}

\item{cosAngleThresh}{(numeric) In simType = dotProductMasked mode, angular similarity should be higher than cosAngleThresh otherwise similarity is forced to zero.}

\item{OverlapAlignment}{(logical) An input for alignment with free end-gaps. False: Global alignment, True: overlap alignment.}

\item{dotProdThresh}{(numeric) In simType = dotProductMasked mode, values in similarity matrix higher than dotProdThresh quantile are checked for angular similarity.}

\item{gapQuantile}{(numeric) Must be between 0 and 1. This is used to calculate base gap-penalty from similarity distribution.}

\item{hardConstrain}{(logical) If FALSE; indices farther from noBeef distance are filled with distance from linear fit line.}

\item{samples4gradient}{(numeric) This parameter modulates penalization of masked indices.}

\item{objType}{(char) Must be selected from light, medium and heavy.}
}
\value{
A S4 object. Three most-important slots are:
\item{indexA_aligned}{(integer) aligned indices of reference run.}
\item{indexB_aligned}{(integer) aligned indices of experiment run.}
\item{score}{(numeric) cumulative score of alignment.}
}
\description{
Outputs AlignObj from an alignment of two XIC-groups
}
\seealso{
\code{\link{alignChromatogramsCpp}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
