% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dia_runs.R
\name{getAlignObjs}
\alias{getAlignObjs}
\title{AlignObj for analytes between a pair}
\usage{
getAlignObjs(analytes, runs, dataPath = ".", alignType = "hybrid",
  runType = "DIA_Proteomics", refRun = NULL, oswMerged = TRUE,
  nameCutPattern = "(.*)(/)(.*)", maxFdrQuery = 0.05,
  maxFdrLoess = 0.01, analyteFDR = 1, spanvalue = 0.1,
  normalization = "mean", simMeasure = "dotProductMasked",
  SgolayFiltOrd = 4, SgolayFiltLen = 9, goFactor = 0.125,
  geFactor = 40, cosAngleThresh = 0.3, OverlapAlignment = TRUE,
  dotProdThresh = 0.96, gapQuantile = 0.5, hardConstrain = FALSE,
  samples4gradient = 100, expRSE = 8, samplingTime = 3.4,
  RSEdistFactor = 3.5, objType = "light")
}
\value{
A list of AlignObj. Each AlignObj contains alignment path, similarity matrix and related parameters.
}
\description{
AlignObj for analytes between a pair
}
