% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytes.R
\name{getAnalytes}
\alias{getAnalytes}
\title{Get names of analytes found in all runs.}
\usage{
getAnalytes(dataPath, runs = NULL, oswMerged = TRUE,
  runType = "DIA_Proteomics", commonAnalytes = FALSE,
  maxFdrQuery = 0.05, nameCutPattern = "(.*)(/)(.*)",
  analyteInGroupLabel = FALSE)
}
\value{
A vector of strings.
}
\description{
Get names of analytes found in all runs.
}
