% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytes.R
\name{getAnalytes}
\alias{getAnalytes}
\title{Get names of analytes found in all runs.}
\usage{
getAnalytes(dataPath, runs = NULL, oswMerged = TRUE,
  runType = "DIA_Proteomics", commonAnalytes = FALSE,
  maxFdrQuery = 0.05, nameCutPattern = "(.*)(/)(.*)",
  analyteInGroupLabel = FALSE)
}
\arguments{
\item{dataPath}{(char) Path to mzml and osw directory.}

\item{runs}{(A vector of string) Names of mzml file without extension.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{commonAnalytes}{(logical) TRUE: intersect across all runs, FASLE: union across all runs.}

\item{maxFdrQuery}{(numeric) A numeric value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{nameCutPattern}{(string) regex expression to fetch mzML file name from RUN.FILENAME columns of osw files.}

\item{analyteInGroupLabel}{(logical) TRUE for getting analytes as PRECURSOR.GROUP_LABEL from osw file.}
}
\value{
A vector of strings.
}
\description{
Get names of analytes found in all runs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt",
 "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt")
analytes <- getAnalytes(dataPath, runs, oswMerged = TRUE, maxFdrQuery = 0.01,
 commonAnalytes = TRUE)
}
\seealso{
\code{\link{getRunNames}, \link{getOswAnalytes}, \link{getAnalytesName}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
