% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytes.R
\name{getAnalytesName}
\alias{getAnalytesName}
\title{Get names of analytes found in all runs.}
\usage{
getAnalytesName(oswFiles, analyteFDR = 1, commonAnalytes = TRUE)
}
\value{
A vector of strings.
}
\description{
Get names of analytes found in all runs.
}
