% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_analytes.R
\name{getAnalytesName}
\alias{getAnalytesName}
\title{Get names of analytes found in all runs}
\usage{
getAnalytesName(oswFiles, analyteFDR = 1, commonAnalytes = TRUE)
}
\arguments{
\item{oswFiles}{(list of data-frames) it is output from getOswFiles function.}

\item{analyteFDR}{(numeric) only analytes that have m-score less than this, will be included in the output.}

\item{commonAnalytes}{(logical) TRUE: intersect across all runs, FASLE: union across all runs.}
}
\value{
A vector of strings.
}
\description{
This function provides all found analytes or only common analytes that have m-score less than analyteFDR.
}
\examples{
data("oswFiles_DIAlignR", package = "DIAlignR")
\dontrun{
getAnalytesName(oswFiles = oswFiles_DIAlignR, analyteFDR = 0.01, commonAnalytes = TRUE)
}
}
\seealso{
\code{\link{getOswFiles}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-14
}
