% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osw_query.R
\name{getAnalytesQuery}
\alias{getAnalytesQuery}
\title{This is a query that will be used to fetch information from osw files.}
\usage{
getAnalytesQuery(maxFdrQuery, oswMerged = TRUE, filename = NULL,
  runType = "DIA_Proteomics")
}
\value{
SQL query to be searched.
}
\description{
This is a query that will be used to fetch information from osw files.
}
