% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getChromSimMat}
\alias{getChromSimMat}
\title{Calculate similarity matrix of two fragment-ion chromatogram sets.}
\usage{
getChromSimMat(l1, l2)
}
\arguments{
\item{seq1}{(char) A single string}

\item{seq2}{(char) A single string}

\item{Match}{(float) Score for character match}

\item{MisMatch}{(float) score for character mismatch}
}
\value{
s (matrix) Numeric similarity matrix. Rows and columns expresses seq1 and seq2, respectively
}
\description{
Calculate similarity matrix of two fragment-ion chromatogram sets.
}
\examples{
# Get sequence similarity of two DNA strings
Match=10; MisMatch=-2
seq1 = "GCAT"; seq2 = "CAGTG"
getChromSimMat(seq1, seq2, Match, MisMatch)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-05
}
