% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGlobalAlignMask}
\alias{getGlobalAlignMask}
\title{Get a mask for constraining similarity matrix.}
\usage{
getGlobalAlignMask(tA, tB, B1p, B2p, noBeef = 50L,
  hardConstrain = FALSE)
}
\arguments{
\item{ROW_SIZE}{(int) Number of rows of a matrix}

\item{COL_SIZE}{(int) Number of columns of a matrix}
}
\value{
affineAlignObj (S4class) A S4class dummy object from C++ AffineAlignObj struct
}
\description{
This function takes in timeVectors from both runs, a global-fit object and
sample-length of window of no constraining. Outside of window, all elements
of matrix are either equally weighted or weighted proportional to distance
from window-boundry.
}
\examples{
tA <- c(3353.2, 3356.6, 3360.0, 3363.5)
tB <- c(3325.9, 3329.3, 3332.7, 3336.1)
B1p <- 3325.751; B2p <- 3336.119
noBeef <- 1
m <- getGlobalAlignMask(tA, tB, B1p, B2p, noBeef, FALSE)
matrix(c(0.0000, 0.0000, 0.7071, 1.4142, 0.0000, 0.0000, 0.0000, 0.7071, 0.7071, 0.0000,
0.0000, 0.0000, 1.4142, 0.7071, 0.0000, 0.0000), 4, 4, byrow = F)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
