% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getGlobalAlignMask}
\alias{getGlobalAlignMask}
\title{Get a mask for constraining similarity matrix}
\usage{
getGlobalAlignMask(tA, tB, B1p, B2p, noBeef = 50L,
  hardConstrain = FALSE)
}
\arguments{
\item{ROW_SIZE}{(int) Number of rows of a matrix}

\item{COL_SIZE}{(int) Number of columns of a matrix}
}
\value{
affineAlignObj (S4class) A S4class dummy object from C++ AffineAlignObj struct
}
\description{
Get a mask for constraining similarity matrix
}
\examples{
x <- setAffineAlignObj_S4(4, 5)
x@signalA_len # 3
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
