% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_fit.R
\name{getLinearfit}
\alias{getLinearfit}
\title{Calculates linear fit between RT of two runs}
\usage{
getLinearfit(RUNS_RT)
}
\arguments{
\item{RUNS_RT}{(data-frame) must have three calumns: transition_group_id, RT.eXp, and RT.ref.}
}
\value{
An object of class "lm".
}
\description{
This function uses output of getRTdf that selects features from oswFiles which has m-score < maxFdrLoess. It fits Linear model on these feature.
Loess mapping is established from reference to experiment run.
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
\dontrun{
RUNS_RT <- getRTdf(oswFiles = oswFiles_DIAlignR, ref = "run1", eXp = "run2", maxFdrGlobal = 0.05)
lm.fit <- getLinearfit(RUNS_RT)
}
}
\seealso{
\code{\link{getLOESSfit}, \link{getFeatures}, \link{getRTdf}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
\keyword{internal}
