% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_fit.R
\name{getLOESSfit}
\alias{getLOESSfit}
\title{calculates LOESS fit between RT of two runs}
\usage{
getLOESSfit(oswRef, oswExp, peptides, spanvalue = 0.1)
}
\arguments{
\item{peptides}{Test peptides' names.}

\item{spanvalue}{A numeric Spanvalue for LOESS fit. For targeted proteomics
0.1 could be used.}

\item{pairName}{Names of runs joined with underscore. e.g. runA_runB, This will allow
alignment of runB to runA.}

\item{oswOutput}{list of list (OpenSWATH output).}
}
\value{
An object of class "loess".
}
\description{
This function takes in run-pairs, names of test peptides, span paprameter for
LOESS function, output of OpenSWATH which include estimated retention time of
peptides.
}
