% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_fit.R
\name{getLOESSfit}
\alias{getLOESSfit}
\title{Calculates LOESS fit between RT of two runs}
\usage{
getLOESSfit(RUNS_RT, spanvalue = 0.1)
}
\arguments{
\item{RUNS_RT}{(data-frame) must have three calumns: transition_group_id, RT.eXp, and RT.ref.}

\item{spanvalue}{(numeric) Spanvalue for LOESS fit. For targeted proteomics 0.1 could be used.}
}
\value{
An object of class "loess".
}
\description{
This function uses output of getRTdf that selects features from oswFiles which has m-score < maxFdrLoess. It fits LOESS on these feature.
Loess mapping is established from reference to experiment run.
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
\dontrun{
RUNS_RT <- getRTdf(oswFiles = oswFiles_DIAlignR, ref = "run1", eXp = "run2", maxFdrGlobal = 0.05)
Loess.fit <- getLOESSfit(RUNS_RT, spanvalue = 0.1)
}
}
\seealso{
\code{\link{getLinearfit}, \link{getFeatures}, \link{getRTdf}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-14
}
\keyword{internal}
