% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getMappedRT}
\alias{getMappedRT}
\title{Extract XICs of all transitions requested in chromIndices.}
\usage{
getMappedRT(XICs.ref, XICs.eXp, Loess.fit, alignType, adaptiveRT,
  samplingTime, normalization, simMeasure, goFactor, geFactor,
  cosAngleThresh, OverlapAlignment, dotProdThresh, gapQuantile,
  hardConstrain, samples4gradient)
}
\value{
A list of list of data-frames. Each data frame has elution time and intensity of fragment-ion XIC.
}
\description{
Extract XICs of all transitions requested in chromIndices.
}
