% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_mrm_runs.R
\name{getMRMoswFiles}
\alias{getMRMoswFiles}
\title{Get list of peptides and their chromatogram indices.}
\usage{
getMRMoswFiles(filenames, dataPath = ".", peptides = NULL,
  query = NULL, oswMerged = FALSE, maxFdrQuery = 1)
}
\value{
A list of data-frames.
}
\description{
Get list of peptides and their chromatogram indices.
}
