% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstAlign.R
\name{getMST}
\alias{getMST}
\title{Create a MST from distance matrix}
\usage{
getMST(distMat)
}
\arguments{
\item{distMat}{(dist) a pairwise distance matrix.}
}
\value{
(matrix) array of tree-edges.
}
\description{
Builds a minimum spanning tree from the distance.
}
\examples{
m <- matrix(c(0,1,2,3, 1,0,1.5,1.5, 2,1.5,0,1, 3,1.5,1,0), byrow = TRUE,
            ncol = 4, dimnames = list(c("run1", "run2", "run3", "run4"),
                                      c("run1", "run2", "run3", "run4")))
distMat <- as.dist(m, diag = FALSE, upper = FALSE)
\dontrun{
x <- getMST(distMat)
g1 <- igraph::graph_from_edgelist(edges1, directed = FALSE)
plot(x)
}
}
\seealso{
\code{\link[ape]{mst}, \link{traverseMST}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2021) + GPL-3
Date: 2021-05-14
}
\keyword{internal}
