% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzml.R
\name{getMZMLpointers}
\alias{getMZMLpointers}
\title{Get pointers to each mzML file.}
\usage{
getMZMLpointers(dataPath, runs)
}
\arguments{
\item{dataPath}{(char) path to mzml and osw directory.}

\item{runs}{(A vector of string) names of mzml file without extension. Vector must have names as shown in the example.}
}
\value{
(A list of mzRpwiz)
}
\description{
Returns instantiated mzRpwiz object associated to mzML file.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("run0" = "hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt",
 "run1" = "hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt")
mzPntrs <- getMZMLpointers(dataPath = dataPath, runs = runs)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
