% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_runs.R
\name{getOswFiles}
\alias{getOswFiles}
\title{Get list of peptides and their chromatogram indices.}
\usage{
getOswFiles(filenames, dataPath = ".", peptides = NULL, query = NULL,
  oswMerged = TRUE, maxFdrQuery = 1, nameCutPattern = "(.*)(/)(.*)")
}
\value{
A list of data-frames.
}
\description{
Get list of peptides and their chromatogram indices.
}
