% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_osw_mzml.R
\name{getOswFiles}
\alias{getOswFiles}
\title{Get list of peptides and their chromatogram indices.}
\usage{
getOswFiles(
  fileInfo,
  mzPntrs,
  maxFdrQuery = 0.05,
  analyteFDR = 0.01,
  oswMerged = TRUE,
  analytes = NULL,
  runType = "DIA_Proteomics",
  analyteInGroupLabel = FALSE
)
}
\arguments{
\item{mzPntrs}{A list of mzRpwiz.
FALSE for fetching analytes as PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}

\item{maxFdrQuery}{(numeric) A numeric value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{analyteFDR}{(numeric) Not used.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{analytes}{(string) analyte is as PRECURSOR.GROUP_LABEL or as PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}

\item{runType}{(char) This must be one of the strings "DIA_Proteomics", "DIA_Metabolomics".}

\item{analyteInGroupLabel}{(logical) TRUE for getting analytes as PRECURSOR.GROUP_LABEL from osw file.}

\item{dataPath}{(char) path to xics and osw directory.}

\item{filenames}{(data-frame) column "filename" contains RUN table from osw files. column "runs" contain respective mzML names without extension.
To get filenames use DIAlignR::getRunNames function.}
}
\value{
(data-frames) Data-frame has following columns:
\item{transition_group_id}{(string) it is either fetched from PRECURSOR.GROUP_LABEL or a combination of PEPTIDE.MODIFIED_SEQUENCE and PRECURSOR.CHARGE from osw file.}
\item{RT}{(numeric) retention time as in FEATURE.EXP_RT of osw files.}
\item{delta_rt}{(numeric) as in FEATURE.DELTA_RT of osw files.}
\item{assay_RT}{(numeric) library retention time as in PRECURSOR.LIBRARY_RT of osw files.}
\item{Intensity}{(numeric) peak intensity as in FEATURE_MS2.AREA_INTENSITY of osw files.}
\item{leftWidth}{(numeric) as in FEATURE.LEFT_WIDTH of osw files.}
\item{rightWidth}{(numeric) as in FEATURE.RIGHT_WIDTH of osw files.}
\item{peak_group_rank}{(integer) rank of each feature associated with transition_group_id.}
\item{m_score}{(numeric) q-value of each feature associated with transition_group_id.}
\item{chromatogramIndex}{(integer) Index of chromatogram in mzML file.}
\item{transition_ids}{(integer) fragment-ion ID associated with transition_group_id. This is matched with chromatogram ID in mzML file.}
}
\description{
This function reads all osw and xics files in the directories at dataPath. It selects analytes which has associated features with m-score < maxFdrQuery.
For these analytes it fetches chromatogram indices by matching transition_id(osw) with chromatogramID(xics).
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
filenames <- getRunNames(dataPath = dataPath)
\dontrun{
mzPntrs <- getMZMLpointers(filenames)
oswFiles <- getOswFiles(filenames, mzPntrs, analyteInGroupLabel = TRUE)
rm(mzPntrs)
}
}
\seealso{
\code{\link{getRunNames}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
\keyword{internal}
