% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_osw.R
\name{getPrecursors}
\alias{getPrecursors}
\title{Get precursors from all feature files}
\usage{
getPrecursors(fileInfo, oswMerged = TRUE, runType = "DIA_Proteomics")
}
\arguments{
\item{fileInfo}{(data-frame) Output of DIAlignR::getRunNames function.}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{runType}{(char) This must be one of the strings "DIA_Proteomics", "DIA_Metabolomics".}
}
\value{
(data-frames) A data-frame having following columns:
\item{transition_group_id}{(integer) a unique id for each precursor.}
\item{transition_id}{(list) fragment-ion ID associated with transition_group_id. This is matched with chromatogram ID in mzML file.}
\item{peptide_id}{(integer) a unique id for each peptide. A peptide can have multiple precursors.}
\item{sequence}{(string) amino-acid sequence of the precursor with possible modifications.}
\item{charge}{(integer) charge on the precursor.}
\item{group_label}{(string) TODO Figure it out.}
}
\description{
Get a data-frame of analytes' transition_group_id, transition_ids, peptide_id and amino-acid sequences.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
fileInfo <- DIAlignR::getRunNames(dataPath = dataPath)
\dontrun{
precursorsInfo <- getPrecursors(fileInfo, oswMerged = TRUE, runType = "DIA_Proteomics")
dim(precursorsInfo) # 322  6
}
}
\seealso{
\code{\link{getRunNames}, \link{fetchPrecursorsInfo}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-04-06
}
