% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osw_query.R
\name{getPrecursorsQuery}
\alias{getPrecursorsQuery}
\title{Get precursor Info}
\usage{
getPrecursorsQuery(runType = "DIA_Proteomics", level = "Peptide")
}
\arguments{
\item{runType}{(string) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{level}{(string) Apply maxPeptideFDR on Protein as well if specified as "Protein". Default: "Peptide".}
}
\value{
SQL query to be searched.
}
\description{
For each precursor in the table respective transition ids are fetched.
Order of transition is kept same as the order of their intensities in \code{\link{getTransitionsQuery}}.
}
\seealso{
\code{\link{fetchPrecursorsInfo}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2020-04-04
}
\keyword{internal}
