% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_runs.R
\name{getQuery}
\alias{getQuery}
\title{This is a query that will be used to fetch information from osw files.}
\usage{
getQuery(maxFdrQuery, peptides = NULL)
}
\value{
SQL query to be searched.
}
\description{
This is a query that will be used to fetch information from osw files.
}
