% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_order.R
\name{getTree}
\alias{getTree}
\title{Create a phylogenetic tree}
\usage{
getTree(distMat, method = "average", prefix = "master")
}
\arguments{
\item{distMat}{(dist) a pairwise distance matrix.}
}
\value{
(phylo) a tree.
}
\description{
Builds a phylogenetic tree from the distance matrix using UPGMA algorithm.
}
\examples{
m <- matrix(c(0,1,2,3, 1,0,1.5,1.5, 2,1.5,0,1, 3,1.5,1,0), byrow = TRUE,
            ncol = 4, dimnames = list(c("run1", "run2", "run3", "run4"),
                                      c("run1", "run2", "run3", "run4")))
distMat <- as.dist(m, diag = FALSE, upper = FALSE)
\dontrun{
tree <- getTree(distMat)
}
tree <- ape::read.tree(text = "(run1:9,(run2:7,run0:2)master2:5)master1;")
plot(tree, type = "phylogram", show.node.label = TRUE)
ape::axisPhylo(1)
plot(tree, type = "unrooted", show.node.label = TRUE)
ape::edgelabels(tree$edge.length)
tree <- ape::nj(distMat) # Neighbor-Joining tree
plot(tree, type = "unrooted", show.node.label = TRUE)
ape::edgelabels(tree$edge.length)
}
\seealso{
\code{\link[phangorn]{upgma}, \link{getNodeIDs}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-31
}
\keyword{internal}
