% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getXICs}
\alias{getXICs}
\title{Get XICs of all analytes}
\usage{
getXICs(
  analytes,
  runs,
  dataPath = ".",
  maxFdrQuery = 1,
  runType = "DIA_proteomics",
  oswMerged = TRUE
)
}
\arguments{
\item{analytes}{(integer) a vector of precursor IDs.}

\item{runs}{(vector of string) names of mzML files without extension.}

\item{dataPath}{(string) Path to mzml and osw directory.}

\item{maxFdrQuery}{(numeric) A numeric value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}
}
\value{
A list of list. Each list contains XIC-group for that run. XIC-group is a list of dataframe that has elution time and intensity of fragment-ion XIC.
}
\description{
For all the analytes requested in runs, it first creates oswFiles, then, fetches chromatogram indices from oswFiles and
extract chromatograms from mzML files.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt",
 "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt")
analytes <- c(32L, 898L, 2474L)
XICs <- getXICs(analytes, runs = runs, dataPath = dataPath)
}
\seealso{
\code{\link{getXICs4AlignObj}, \link{getRunNames}, \link{analytesFromFeatures}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
