% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_runs.R
\name{getXICs}
\alias{getXICs}
\title{Get XICs for a list of peptides}
\usage{
getXICs(peptides, runs, dataPath = ".", maxFdrQuery = 1,
  SgolayFiltOrd = 4, SgolayFiltLen = 9, query = NULL)
}
\value{
A list of list. Each list contains XICs for that run.
}
\description{
Get XICs for a list of peptides
}
