% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getXICs4AlignObj}
\alias{getXICs4AlignObj}
\title{Extract XICs of all transitions requested in chromIndices.}
\usage{
getXICs4AlignObj(dataPath, runs, oswFiles, analytes,
  XICfilter = "sgolay", SgolayFiltOrd = 4, SgolayFiltLen = 9)
}
\value{
A list of list of data-frames. Each data frame has elution time and intensity of fragment-ion XIC.
}
\description{
Extract XICs of all transitions requested in chromIndices.
}
