% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{getXICs4AlignObj}
\alias{getXICs4AlignObj}
\title{Extract XICs of analytes}
\usage{
getXICs4AlignObj(mzPntrs, fileInfo, runs, prec2chromIndex, analytes)
}
\arguments{
\item{mzPntrs}{a list of mzRpwiz.}

\item{fileInfo}{(data-frame) output of getRunNames().}

\item{runs}{(vector of string) names of mzML files without extension.}

\item{prec2chromIndex}{(list of data-frames) output of getChromatogramIndices(). Each dataframe has two columns: transition_group_id and chromatogramIndex.}

\item{analytes}{(integer) a vector of precursor IDs.}
}
\value{
A list of list of data-frames. Each data frame has elution time and intensity of fragment-ion XIC.
}
\description{
For all the analytes requested, it fetches chromatogram indices from prec2chromIndex and
extracts chromatograms using mzPntrs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt",
          "hroest_K120808_Strep10\%PlasmaBiolRepl1_R03_SW_filt")
analytes <- c(32L, 898L, 2474L)
params <- paramsDIAlignR()
params[["chromFile"]] <- "mzML"
fileInfo <- getRunNames(dataPath = dataPath)
fileInfo <- updateFileInfo(fileInfo, runs)
precursors <- getPrecursorByID(analytes,fileInfo)
mzPntrs <- getMZMLpointers(fileInfo)
prec2chromIndex <- getChromatogramIndices(fileInfo, precursors, mzPntrs)
XICs <- getXICs4AlignObj(mzPntrs, fileInfo, runs, prec2chromIndex, analytes)
rm(mzPntrs)
}
\seealso{
\code{\link{getChromatogramIndices}, \link{getRunNames}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
