% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initializeMatrix}
\alias{initializeMatrix}
\title{Calculate similarity matrix for two sequences}
\usage{
initializeMatrix(initVal, ROW_SIZE, COL_SIZE)
}
\arguments{
\item{initVal}{(char) Matrix intialization value}

\item{ROW_SIZE}{(int) Number of rows}

\item{COL_SIZE}{(int) Number of columns}
}
\value{
s (matrix) a similarity matrix
}
\description{
Calculate similarity matrix for two sequences
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-05
}
