% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_alignment.R
\name{mapIdxToTime}
\alias{mapIdxToTime}
\title{Establishes mapping from index to time}
\usage{
mapIdxToTime(timeVec, idx)
}
\arguments{
\item{timeVec}{A numeric vector}

\item{idx}{An integer vector}
}
\value{
A mutated time vector
}
\description{
Takes a time vector and index vector of same length. This function create a
new time vector given indices specified in \code{idx}. For \code{NA} indices
it uses last index to fill time value. For \code{NA} appearing at the start
of \code{idx}, it uses next index to get time value.
}
\examples{
timeVec <- c(1.3,5.6,7.8)
idx <- c(NA, NA, 1L, 2L, NA, NA, 3L, NA)
mapIdxToTime(timeVec, idx) # c(1.3, 1.3, 1.3, 5.6, 5.6, 5.6, 7.8, 7.8)

}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
