% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_alignment.R
\name{mappedRTfromAlignObj}
\alias{mappedRTfromAlignObj}
\title{Map reference run time on the experiment run.}
\usage{
mappedRTfromAlignObj(refRT, tVec.ref, tVec.eXp, AlignObj)
}
\arguments{
\item{refRT}{(numeric) retention time in reference run.}

\item{tVec.ref}{(numeric) time vector of refernce run.}

\item{tVec.eXp}{(numeric) time vector of experiment run.}

\item{AlignObj}{(S4 object) must have indexA_aligned, indexB_aligned and score as slots.}
}
\value{
(numeric)
}
\description{
Retention time from reference run is mapped to experiment run using AlignObj.
}
\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
tVec.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]][[1]][, "time"]
tVec.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]][[1]][, "time"]
\dontrun{
AlignObj <- testAlignObj()
mappedRTfromAlignObj(refRT= 5238.35, tVec.ref, tVec.eXp, AlignObj)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
\keyword{internal}
