% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_chromatogram.R
\name{mergeXIC}
\alias{mergeXIC}
\title{Merge two XICs into one}
\usage{
mergeXIC(XIC.ref, XIC.eXp, mergeStrategy)
}
\arguments{
\item{XIC.ref}{(data-frame) extracted ion chromatogram from reference run. Must not contain missing values.}

\item{XIC.eXp}{(data-frame) extracted ion chromatogram from experiment run. Must not contain missing values.}

\item{mergeStrategy}{(string) must be either ref, avg, refStart or refEnd.}
}
\value{
(dataframe) has two columns:
\item{time}{(numeric)}
\item{intensity}{(numeric)}
}
\description{
Both extracted-ion chromatograms must have same number of rows. Missing values are not allowed.
Intensities are weighted-averaged.
}
\section{time-merging}{

There are three startegies for calculating time to keep sampling rate equal to parents:\cr
 - Use reference time. \cr
 - Average time (Non-gap region will have parent's sampling rate). \cr
 - Start/end with the average then fixed sampling rate.
}

\examples{
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"]][["14299_QFNNTDIVLLEDFQK/3"]]
\dontrun{
plot(mergeXIC(XICs.ref[[1]], XICs.eXp[[1]], mergeStrategy="ref"), type = "l")
}
}
\seealso{
\code{\link{childXIC}, \link{alignedXIC}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-05-23
}
\keyword{internal}
