% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotAlignedAnalytes}
\alias{plotAlignedAnalytes}
\title{Plot aligned XICs group for a specific peptide.
AlignObjOutput is the output from getAlignObjs fucntion.}
\usage{
plotAlignedAnalytes(AlignObjOutput, plotType = "All",
  annotatePeak = FALSE, DrawAlignR = FALSE)
}
\description{
Plot aligned XICs group for a specific peptide.
AlignObjOutput is the output from getAlignObjs fucntion.
}
