% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotAlignedAnalytes}
\alias{plotAlignedAnalytes}
\title{Plot aligned XICs group for a specific peptide.
AlignObjOutput is the output from getAlignObjs fucntion.}
\usage{
plotAlignedAnalytes(AlignObjOutput, plotType = "All",
  outFile = "AlignedAnalytes.pdf", annotatePeak = FALSE,
  saveFigs = FALSE)
}
\arguments{
\item{AlignObjOutput}{(list) list contains fileInfo, AlignObj, raw XICs for reference and experiment, and reference-peak label.}

\item{plotType}{(string) must be one of the strings "All", "onlyUnaligned" and "onlyAligned".}

\item{outFile}{(string) name of the output pdf file.}

\item{annotatePeak}{(logical) TRUE: Peak boundaries and apex will be highlighted.}

\item{saveFigs}{(logical) TRUE: Figures will be saved in AlignedAnalytes.pdf .}
}
\value{
A plot to the current device.
}
\description{
Plot aligned XICs group for a specific peptide.
AlignObjOutput is the output from getAlignObjs fucntion.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt",
 "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt")
AlignObjOutput <- getAlignObjs(analytes = 4618L, runs, dataPath = dataPath)
plotAlignedAnalytes(AlignObjOutput)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
