% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotAnalyteXICs}
\alias{plotAnalyteXICs}
\title{Plot extracted-ion chromatogram.}
\usage{
plotAnalyteXICs(
  analyte,
  run,
  dataPath = ".",
  maxFdrQuery = 1,
  XICfilter = "sgolay",
  polyOrd = 4,
  kernelLen = 9,
  runType = "DIA_proteomics",
  oswMerged = TRUE,
  peakAnnot = NULL,
  Title = NULL
)
}
\arguments{
\item{analyte}{(integer) an analyte is a PRECURSOR.ID from the osw file.}

\item{run}{(string) Name of a xics file without extension.}

\item{dataPath}{(string) path to xics and osw directory.}

\item{maxFdrQuery}{(numeric) A numeric value between 0 and 1. It is used to filter features from osw file which have SCORE_MS2.QVALUE less than itself.}

\item{XICfilter}{(string) must be either sgolay, boxcar, gaussian, loess or none.}

\item{polyOrd}{(integer) order of the polynomial to be fit in the kernel.}

\item{kernelLen}{(integer) number of data-points to consider in the kernel.}

\item{runType}{(char) This must be one of the strings "DIA_proteomics", "DIA_Metabolomics".}

\item{oswMerged}{(logical) TRUE for experiment-wide FDR and FALSE for run-specific FDR by pyprophet.}

\item{peakAnnot}{(numeric) Peak-apex time.}

\item{Title}{(logical) TRUE: name of the list will be displayed as title.}
}
\value{
A plot to the current device.
}
\description{
Plot extracted-ion chromatogram.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
run <- "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"
plotAnalyteXICs(analyte = 2474L, run, dataPath = dataPath, oswMerged = TRUE, XICfilter = "none")
plotAnalyteXICs(analyte = 2474L, run, dataPath = dataPath, oswMerged = TRUE, XICfilter = "sgolay")
}
\seealso{
\code{\link{plotXICgroup}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
