% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotAnalyteXICs}
\alias{plotAnalyteXICs}
\title{Plot extracted-ion chromatogram.}
\usage{
plotAnalyteXICs(analyte, run, dataPath = ".", maxFdrQuery = 1,
  XICfilter = "sgolay", SgolayFiltOrd = 4, SgolayFiltLen = 9,
  runType = "DIA_proteomics", oswMerged = TRUE,
  nameCutPattern = "(.*)(/)(.*)", analyteInGroupLabel = FALSE,
  peakAnnot = NULL, Title = NULL)
}
\description{
Plot extracted-ion chromatogram.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
run <- "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"
plotAnalyteXICs(analyte = "QFNNTDIVLLEDFQK_3", run, dataPath = dataPath, XICfilter = "none")
plotAnalyteXICs(analyte = "14299_QFNNTDIVLLEDFQK/3", run, dataPath = dataPath,
XICfilter = "sgolay", analyteInGroupLabel = TRUE)
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
