% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotSingleAlignedChrom}
\alias{plotSingleAlignedChrom}
\title{Plot an aligned XIC-group.}
\usage{
plotSingleAlignedChrom(XIC_group, idx, peakAnnot = NULL)
}
\arguments{
\item{XIC_group}{(list) It is a list of dataframe which has two columns. First column is for time
and second column indicates intensity.}

\item{idx}{(integer) Indices of aligned chromatograms.}

\item{peakAnnot}{(numeric) Peak-apex time.}
}
\value{
A plot to the current device.
}
\description{
Plot an aligned XIC-group.
}
\details{
x-axis cannot have the same time-values, therefore, x-axis is indecized.
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
