% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotSingleAlignedChrom}
\alias{plotSingleAlignedChrom}
\title{Plot an aligned XIC-group.
x-axis cannot have the same time-values, therefore, x-axis is indecized.}
\usage{
plotSingleAlignedChrom(XIC_group, idx, peakAnnot = NULL)
}
\description{
Plot an aligned XIC-group.
x-axis cannot have the same time-values, therefore, x-axis is indecized.
}
