% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotXICgroup}
\alias{plotXICgroup}
\title{Plot Extracted-ion chromatogram group.}
\usage{
plotXICgroup(XIC_group, peakAnnot = NULL, Title = NULL)
}
\description{
Plot Extracted-ion chromatogram group.
}
