% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_chromatograms.R
\name{plotXICgroup}
\alias{plotXICgroup}
\title{Plot Extracted-ion chromatogram group.}
\usage{
plotXICgroup(XIC_group, peakAnnot = NULL, Title = NULL)
}
\arguments{
\item{XIC_group}{(list) It is a list of dataframe which has two columns. First column is for time
and second column indicates intensity.}

\item{peakAnnot}{(numeric) Peak-apex time.}

\item{Title}{(logical) TRUE: name of the list will be displayed as title.}
}
\value{
A plot to the current device.
}
\description{
Plot Extracted-ion chromatogram group.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
runs <- c("hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt",
 "hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt")
XICs <- getXICs(analytes = 4618L, runs = runs, dataPath = dataPath, oswMerged = TRUE)
plotXICgroup(XICs[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]])
XICs <- smoothXICs(XICs[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]],
  type = "sgolay", kernelLen = 13, polyOrd = 4)
plotXICgroup(XICs, Title = "Precursor 4618 \n
 run hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt")
}
\seealso{
\code{\link{plotAnalyteXICs}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + GPL-3
Date: 2019-12-13
}
