% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzml.R
\name{readChromatogramHeader}
\alias{readChromatogramHeader}
\title{Get chromatogram header from a mzML file}
\usage{
readChromatogramHeader(mzmlName)
}
\arguments{
\item{mzmlName}{(char) path to mzml file.}
}
\value{
(A data-frame) It has 10 columns. The two important columns are:
\item{chromatogramId}{(integer) Fragment-ion ID that matches with transition ID in osw file.}
\item{chromatogramIndex}{(integer) Index of chromatogram in mzML file.}
}
\description{
Get a table of chromatogram indices and respective transition IDs.
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
mzmlName <-paste0(dataPath,"/mzml/hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt.chrom.mzML")
\dontrun{
chromHead <- readChromatogramHeader(mzmlName = mzmlName)
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2019) + MIT
Date: 2019-12-13
}
