% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setAffineAlignObj_S4}
\alias{setAffineAlignObj_S4}
\title{Initialize a S4 object AffineAlignObj1}
\usage{
setAffineAlignObj_S4(ROW_SIZE, COL_SIZE)
}
\arguments{
\item{ROW_SIZE}{(int) Number of rows of a matrix}

\item{COL_SIZE}{(int) Number of columns of a matrix}
}
\value{
affineAlignObj (S4class) A S4class dummy object from C++ class of AffineAlignObj
}
\description{
Initialize a S4 object AffineAlignObj1
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}
ORCID: 0000-0003-3500-8152
License: (c) Author (2019) + MIT
Date: 2019-03-08
}
