% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_alignment.R
\name{setAlignmentRank}
\alias{setAlignmentRank}
\title{Set Alignment rank to the aligned feature}
\usage{
setAlignmentRank(df, refIdx, eXp, tAligned, XICs, params, adaptiveRT)
}
\arguments{
\item{tAligned}{(list) the first element corresponds to the aligned reference time,
the second element is the aligned experiment time.}

\item{adaptiveRT}{(numeric) defines the window around the aligned retention time, within which
features with m-score below aligned FDR are considered for quantification.}

\item{XICs.eXp}{(list) list of extracted ion chromatograms from experiment run.}
}
\description{
Picks the top feature in run by comparing m-score to unaligned FDR and aligned FDR.
If no satisfactory feature is found, peak-integration is carried out by mapping left and right peak
boundaries from the reference feature and integrating area under the curve.
}
\examples{
data(multipeptide_DIAlignR, package="DIAlignR")
data(XIC_QFNNTDIVLLEDFQK_3_DIAlignR, package="DIAlignR")
params <- paramsDIAlignR()
df <- multipeptide_DIAlignR[["14383"]]
df$alignment_rank[2] <- 1L
XICs.ref <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep0\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
XICs.eXp <- XIC_QFNNTDIVLLEDFQK_3_DIAlignR[["hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt"]][["4618"]]
\dontrun{
# Use getAlignedTimes() to get tAligned.
alignObj <- testAlignObj()
tAligned <- alignedTimes2(alignObj, XICs.ref, XICs.eXp)
setAlignmentRank(df, refIdx = 3L, eXp = "run2", tAligned, XICs.eXp,
params, adaptiveRT = 38.66)
}
}
\seealso{
\code{\link{getMultipeptide}, \link{calculateIntensity}, \link{alignToRef}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-04-13
}
\keyword{internal}
