% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peaks_chromatograms.R
\name{uncompressVec}
\alias{uncompressVec}
\title{Uncompress a Blob object}
\usage{
uncompressVec(x, type)
}
\arguments{
\item{x}{(Blob object)}

\item{type}{(integer) must either be 5L or 6L to indicate linear and short logged float compression, respectively.}
}
\value{
A numeric vector. Uncompressed form of the Blob.
}
\description{
compression is one of 0 = no, 1 = zlib, 2 = np-linear,
 3 = np-slof, 4 = np-pic, 5 = np-linear + zlib,
  6 = np-slof + zlib, 7 = np-pic + zlib
}
\examples{
dataPath <- system.file("extdata", package = "DIAlignR")
sqName <- paste0(dataPath,"/mzml/hroest_K120809_Strep10\%PlasmaBiolRepl2_R04_SW_filt.chrom.sqMass")
con <- DBI::dbConnect(RSQLite::SQLite(), dbname = sqName)
df1 <- DBI::dbGetQuery(con, "SELECT CHROMATOGRAM_ID, COMPRESSION, DATA_TYPE, DATA FROM DATA WHERE CHROMATOGRAM_ID = 36;")
DBI::dbDisconnect(con)
\dontrun{
time = uncompressVec(df1[["DATA"]][[1]], df1$COMPRESSION[[1]])
intensity = uncompressVec(df1[["DATA"]][[2]], df1$COMPRESSION[[2]])
}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-12-13
}
\keyword{internal}
