% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeTables}
\alias{writeTables}
\title{Writes the output table post-alignment}
\usage{
writeTables(fileInfo, multipeptide, precursors)
}
\arguments{
\item{fileInfo}{(data-frame) Output of getRunNames function.}

\item{multipeptide}{(list of data-frames) Each element of the list is collection of features associated with a precursor.}

\item{precursors}{(data-frame) for each transition_group_id, contains peptide sequence and charge.}

\item{filename}{(string) Name of the output file.}
}
\value{
An output table with following columns: precursor, run, intensity, RT, leftWidth, rightWidth,
 peak_group_rank, m_score, alignment_rank, peptide_id, sequence, charge, group_label.
}
\description{
Selects all features from multipeptide with alignment rank = 1, and write them onto the disk.
}
\examples{
data(oswFiles_DIAlignR, package="DIAlignR")
\dontrun{
writeTables(fileInfo, multipeptide, precursors)
}
}
\seealso{
\code{\link{getRunNames}, \link{getMultipeptide}, \link{getPrecursors}}
}
\author{
Shubham Gupta, \email{shubh.gupta@mail.utoronto.ca}

ORCID: 0000-0003-3500-8152

License: (c) Author (2020) + GPL-3
Date: 2020-04-14
}
\keyword{internal}
